### 2025-09-19 #9025 @editorialbot
[PRE REVIEW]: GeoAI: A Python package for integrating artificial intelligence with geospatial data analysis and visualization
**Submitting author:** @giswqs (<a href="http://orcid.org/0000-0001-5437-4073">Qiusheng Wu</a>)
**Repository:** https://github.com/opengeos/geoai
**Branch with paper.md** (empty if default branch):
**Version:** v0.11.1
