### 2025-09-22 95698fb
Clean up code
