+++ geoai/geoai.py
@@ -10,38 +10,23 @@ import leafmap.maplibregl as maplibregl
 
 from .change_detection import ChangeDetection
 from .classify import classify_image, classify_images, train_classifier
-from .download import (
-    download_naip,
-    download_overture_buildings,
-    download_pc_stac_item,
-    extract_building_stats,
-    get_overture_data,
-    pc_collection_list,
-    pc_item_asset_list,
-    pc_stac_download,
-    pc_stac_search,
-    read_pc_item_asset,
-    view_pc_item,
-    view_pc_items,
-)
+from .download import (download_naip, download_overture_buildings,
+                       download_pc_stac_item, extract_building_stats,
+                       get_overture_data, pc_collection_list,
+                       pc_item_asset_list, pc_stac_download, pc_stac_search,
+                       read_pc_item_asset, view_pc_item, view_pc_items)
 from .extract import *
 from .hf import *
-from .segment import *
-from .train import (
-    get_instance_segmentation_model,
-    instance_segmentation,
-    instance_segmentation_batch,
-    instance_segmentation_inference_on_geotiff,
-    object_detection,
-    object_detection_batch,
-    semantic_segmentation,
-    semantic_segmentation_batch,
-    train_instance_segmentation_model,
-    train_MaskRCNN_model,
-    train_segmentation_model,
-)
-from .utils import *
 from .map_widgets import DINOv3GUI
+from .segment import *
+from .train import (get_instance_segmentation_model, instance_segmentation,
+                    instance_segmentation_batch,
+                    instance_segmentation_inference_on_geotiff,
+                    object_detection, object_detection_batch,
+                    semantic_segmentation, semantic_segmentation_batch,
+                    train_instance_segmentation_model, train_MaskRCNN_model,
+                    train_segmentation_model)
+from .utils import *
 
 
 class LeafMap(leafmap.Map):
