+++ geoai/utils.py
@@ -11,8 +11,17 @@ import xml.etree.ElementTree as ET
 from collections import OrderedDict
 from collections.abc import Iterable
 from pathlib import Path
-from typing import (Any, Callable, Dict, Generator, Iterator, List, Optional,
-                    Tuple, Union)
+from typing import (
+    Any,
+    Callable,
+    Dict,
+    Generator,
+    Iterator,
+    List,
+    Optional,
+    Tuple,
+    Union,
+)
 
 # Third-Party Libraries
 import cv2
