+++ geoai/utils.py
@@ -182,7 +182,6 @@ def view_image(
         image (Union[np.ndarray, torch.Tensor]): The image to visualize.
         transpose (bool, optional): Whether to transpose the image. Defaults to False.
         bdx (Optional[int], optional): The band index to visualize. Defaults to None.
-        scale_factor (float, optional): The scale factor to apply to the image. Defaults to 1.0.
         figsize (Tuple[int, int], optional): The size of the figure. Defaults to (10, 5).
         axis_off (bool, optional): Whether to turn off the axis. Defaults to True.
         title (Optional[str], optional): The title of the plot. Defaults to None.
@@ -711,7 +710,7 @@ def view_vector_interactive(
         tiles_args (dict, optional): Additional arguments for the localtileserver client.
             get_folium_tile_layer function. Defaults to None.
         **kwargs: Additional keyword arguments to pass to GeoDataFrame.explore() function.
-        See https://geopandas.org/en/stable/docs/reference/api/geopandas.GeoDataFrame.explore.html
+            See https://geopandas.org/en/stable/docs/reference/api/geopandas.GeoDataFrame.explore.html
 
     Returns:
         folium.Map: The map object with the vector data added.
@@ -2528,7 +2527,7 @@ def batch_vector_to_raster(
         nodata (int): No data value for the output raster.
 
     Returns:
-        list: List of paths to the created raster files.
+        List[str]: List of paths to the created raster files.
     """
     # Create output directory if it doesn't exist
     os.makedirs(output_dir, exist_ok=True)
@@ -3157,7 +3156,7 @@ def export_geotiff_tiles_batch(
         mask_extensions (list): List of mask file extensions to process (default: common raster/vector formats)
 
     Returns:
-        dict: Dictionary containing batch processing statistics
+        Dict[str, Any]: Dictionary containing batch processing statistics
 
     Raises:
         ValueError: If no images or masks found, or if counts don't match
@@ -4631,13 +4630,13 @@ def export_training_data(
 
 
 def masks_to_vector(
-    mask_path,
-    output_path=None,
-    simplify_tolerance=1.0,
-    mask_threshold=0.5,
-    min_object_area=100,
-    max_object_area=None,
-    nms_iou_threshold=0.5,
+    mask_path: str,
+    output_path: Optional[str] = None,
+    simplify_tolerance: float = 1.0,
+    mask_threshold: float = 0.5,
+    min_object_area: int = 100,
+    max_object_area: Optional[int] = None,
+    nms_iou_threshold: float = 0.5,
 ):
     """
     Convert a building mask GeoTIFF to vector polygons and save as a vector dataset.
@@ -4652,7 +4651,7 @@ def masks_to_vector(
         nms_iou_threshold: IoU threshold for non-maximum suppression (default: self.nms_iou_threshold)
 
     Returns:
-        GeoDataFrame with building footprints
+        Any: GeoDataFrame with building footprints
     """
     # Set default output path if not provided
     # if output_path is None:
@@ -5679,7 +5678,7 @@ def orthogonalize(
         detect_triangles (bool, optional): If True, performs additional check to avoid creating triangular shapes.
 
     Returns:
-        geopandas.GeoDataFrame: A GeoDataFrame containing the orthogonalized features.
+        Any: A GeoDataFrame containing the orthogonalized features.
     """
 
     from functools import partial
@@ -7086,8 +7085,8 @@ def regularize(
     num_cores: int = 1,
     include_metadata: bool = False,
     output_path: Optional[str] = None,
-    **kwargs,
-) -> gpd.GeoDataFrame:
+    **kwargs: Any,
+) -> Any:
     """Regularizes polygon geometries in a GeoDataFrame by aligning edges.
 
     Aligns edges to be parallel or perpendicular (optionally also 45 degrees)
