### 2025-09-22 3587501
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
