### 2025-09-22 181b05c
Fix docs build warnings
