### 2025-09-22 d657696
Add titiler endpoint
