+++ geoai/__init__.py
@@ -99,5 +99,5 @@ def set_proj_lib_path(verbose=False):
 # if ("google.colab" not in sys.modules) and (sys.platform != "windows"):
 #     set_proj_lib_path()
 
+from .dinov3 import DINOv3GeoProcessor, analyze_image_patches, create_similarity_map
 from .geoai import *
-from .dinov3 import DINOv3GeoProcessor, create_similarity_map, analyze_image_patches
