+++ tests/test_classify.py
@@ -5,6 +5,7 @@
 import unittest
 
 from geoai import classify
+
 from .test_fixtures import get_test_data_paths
 
 
