+++ tests/test_segment.py
@@ -5,6 +5,7 @@
 import unittest
 
 from geoai import segment
+
 from .test_fixtures import get_test_data_paths
 
 
