+++ geoai/agents/geo_agents.py
@@ -1,17 +1,23 @@
 from __future__ import annotations
 
 import asyncio
+import os
 import uuid
 from concurrent.futures import ThreadPoolExecutor
 from types import SimpleNamespace
 from typing import Any, Callable, Optional
 
+import boto3
 import ipywidgets as widgets
 import leafmap.maplibregl as leafmap
+from botocore.config import Config as BotocoreConfig
 from ipyevents import Event
 from IPython.display import display
 from strands import Agent
+from strands.models import BedrockModel
+from strands.models.anthropic import AnthropicModel
 from strands.models.ollama import OllamaModel
+from strands.models.openai import OpenAIModel
 
 from .map_tools import MapSession, MapTools
 
@@ -21,6 +27,112 @@ try:
     nest_asyncio.apply()
 except Exception:
     pass
+
+
+def create_ollama_model(
+    host: str = "http://localhost:11434",
+    model_id: str = "llama3.1",
+    **kwargs: Any,
+) -> OllamaModel:
+    """Create an Ollama model.
+
+    Args:
+        host: Ollama host URL.
+        model_id: Ollama model ID.
+        **kwargs: Additional keyword arguments for the Ollama model.
+
+    Returns:
+        OllamaModel: An Ollama model.
+    """
+    return OllamaModel(host=host, model_id=model_id, **kwargs)
+
+
+def create_openai_model(
+    model_id: str = "gpt-4o-mini",
+    api_key: str = None,
+    **kwargs: Any,
+) -> OpenAIModel:
+    """Create an OpenAI model.
+
+    Args:
+        model_id: OpenAI model ID.
+        api_key: OpenAI API key.
+        **kwargs: Additional keyword arguments for the OpenAI model.
+
+    Returns:
+        OpenAIModel: An OpenAI model.
+    """
+
+    if api_key is not None:
+        try:
+            api_key = os.getenv("OPENAI_API_KEY", None)
+            if api_key is None:
+                raise ValueError("OPENAI_API_KEY is not set")
+        except Exception as e:
+            raise ValueError("OPENAI_API_KEY is not set")
+
+    client_args = kwargs.get("client_args", {})
+    if "api_key" not in client_args:
+        client_args["api_key"] = api_key
+
+    return OpenAIModel(client_args=client_args, model_id=model_id, **kwargs)
+
+
+def create_anthropic_model(
+    model_id: str = "claude-sonnet-4-20250514",
+    api_key: str = None,
+    **kwargs: Any,
+) -> AnthropicModel:
+    """Create an Anthropic model.
+
+    Args:
+        model_id: Anthropic model ID. Defaults to "claude-sonnet-4-20250514".
+            For a complete list of supported models,
+            see https://docs.claude.com/en/docs/about-claude/models/overview.
+        api_key: Anthropic API key.
+        **kwargs: Additional keyword arguments for the Anthropic model.
+    """
+
+    if api_key is not None:
+        try:
+            api_key = os.getenv("ANTHROPIC_API_KEY", None)
+            if api_key is None:
+                raise ValueError("ANTHROPIC_API_KEY is not set")
+        except Exception as e:
+            raise ValueError("ANTHROPIC_API_KEY is not set")
+
+    client_args = kwargs.get("client_args", {})
+    if "api_key" not in client_args:
+        client_args["api_key"] = api_key
+
+    return AnthropicModel(client_args=client_args, model_id=model_id, **kwargs)
+
+
+def create_bedrock_model(
+    model_id: str = "anthropic.claude-sonnet-4-20250514-v1:0",
+    region_name: str = None,
+    boto_session: Optional[boto3.Session] = None,
[…]
