### 2025-09-30 ad94d40
Make torchange optional (#288)
* Make torchange optional

* Update requirements.txt

* Update notebook
