### 2025-09-30 edaa1f3
Add more model providers for AI agents (#282)
* Add more model providers

* Update notebook
