+++ geoai/agents/geo_agents.py
@@ -1,17 +1,23 @@
 from __future__ import annotations
 
 import asyncio
+import os
 import uuid
 from concurrent.futures import ThreadPoolExecutor
 from types import SimpleNamespace
 from typing import Any, Callable, Optional
 
+import boto3
 import ipywidgets as widgets
 import leafmap.maplibregl as leafmap
+from botocore.config import Config as BotocoreConfig
 from ipyevents import Event
 from IPython.display import display
 from strands import Agent
-from strands.models.ollama import OllamaModel
+from strands.models import BedrockModel
+from strands.models.anthropic import AnthropicModel
+from strands.models.ollama import OllamaModel as _OllamaModel
+from strands.models.openai import OpenAIModel
 
 from .map_tools import MapSession, MapTools
 
@@ -21,6 +27,147 @@ try:
     nest_asyncio.apply()
 except Exception:
     pass
+
+
+class OllamaModel(_OllamaModel):
+    """Fixed OllamaModel that ensures proper model_id handling."""
+
+    async def stream(self, *args, **kwargs):
+        """Override stream to ensure model_id is passed as string."""
+        # Patch the ollama client to handle model object correctly
+        import ollama
+
+        # Save original method if not already saved
+        if not hasattr(ollama.AsyncClient, "_original_chat"):
+            ollama.AsyncClient._original_chat = ollama.AsyncClient.chat
+
+            async def fixed_chat(self, **chat_kwargs):
+                # If model is an OllamaModel object, extract the model_id
+                if "model" in chat_kwargs and hasattr(chat_kwargs["model"], "config"):
+                    chat_kwargs["model"] = chat_kwargs["model"].config["model_id"]
+                return await ollama.AsyncClient._original_chat(self, **chat_kwargs)
+
+            ollama.AsyncClient.chat = fixed_chat
+
+        # Call the original stream method
+        async for chunk in super().stream(*args, **kwargs):
+            yield chunk
+
+
+def create_ollama_model(
+    host: str = "http://localhost:11434",
+    model_id: str = "llama3.1",
+    client_args: dict = None,
+    **kwargs: Any,
+) -> OllamaModel:
+    """Create an Ollama model.
+
+    Args:
+        host: Ollama host URL.
+        model_id: Ollama model ID.
+        client_args: Client arguments for the Ollama model.
+        **kwargs: Additional keyword arguments for the Ollama model.
+
+    Returns:
+        OllamaModel: An Ollama model.
+    """
+    if client_args is None:
+        client_args = {}
+    return OllamaModel(host=host, model_id=model_id, client_args=client_args, **kwargs)
+
+
+def create_openai_model(
+    model_id: str = "gpt-4o-mini",
+    api_key: str = None,
+    client_args: dict = None,
+    **kwargs: Any,
+) -> OpenAIModel:
+    """Create an OpenAI model.
+
+    Args:
+        model_id: OpenAI model ID.
+        api_key: OpenAI API key.
+        client_args: Client arguments for the OpenAI model.
+        **kwargs: Additional keyword arguments for the OpenAI model.
+
+    Returns:
+        OpenAIModel: An OpenAI model.
+    """
+
+    if api_key is None:
+        try:
+            api_key = os.getenv("OPENAI_API_KEY", None)
+            if api_key is None:
+                raise ValueError("OPENAI_API_KEY is not set")
+        except Exception:
+            raise ValueError("OPENAI_API_KEY is not set")
+
+    if client_args is None:
+        client_args = kwargs.get("client_args", {})
+    if "api_key" not in client_args and api_key is not None:
+        client_args["api_key"] = api_key
+
[…]
