### 2025-09-30 c937a8c
Add more model providers for AI agents (#282)
* Add more model providers

* Update notebook
