+++ geoai/agents/__init__.py
@@ -1,2 +1,8 @@
-from .geo_agents import GeoAgent
+from .geo_agents import (
+    GeoAgent,
+    create_ollama_model,
+    create_anthropic_model,
+    create_openai_model,
+    create_bedrock_model,
+)
 from .map_tools import MapTools
