+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "geoai-py"
-version = "0.12.0"
+version = "0.13.0"
 dynamic = [
     "dependencies",
 ]
@@ -44,7 +44,7 @@ universal = true
 
 
 [tool.bumpversion]
-current_version = "0.12.0"
+current_version = "0.13.0"
 commit = true
 tag = true
 
