### 2025-10-01 14affb2
Add system_prompt param (#289)
