### 2025-10-01 6db205e
Add lightly-train support
