### 2025-10-01 a15cfa9
Update geoai/train.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
