### 2025-10-01 1e9dbdc
Add device parameter to load_lightly_pretrained_model
Updated load_lightly_pretrained_model to accept device parameter for loading model weights.
