### 2025-10-02 8a75ff4
Add support for saving performance metrics plot
