### 2025-10-02 4b9a51b
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
