### 2025-10-02 009e2a1
Add support for saving performance metrics plot (#292)
* Add support for saving performance metrics plot

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
