+++ geoai/__init__.py
@@ -2,7 +2,7 @@
 
 __author__ = """Qiusheng Wu"""
 __email__ = "giswqs@gmail.com"
-__version__ = "0.13.0"
+__version__ = "0.13.1"
 
 
 import os
