+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "geoai-py"
-version = "0.13.0"
+version = "0.13.1"
 dynamic = [
     "dependencies",
 ]
@@ -44,7 +44,7 @@ universal = true
 
 
 [tool.bumpversion]
-current_version = "0.13.0"
+current_version = "0.13.1"
 commit = true
 tag = true
 
