### 2025-10-02 6b3cc05
Add support for probability output (#294)
