+++ geoai/train.py
@@ -2625,6 +2625,8 @@ def semantic_inference_on_geotiff(
     num_channels: int = 3,
     num_classes: int = 2,
     device: Optional[torch.device] = None,
+    probability_path: Optional[str] = None,
+    probability_threshold: Optional[float] = None,
     quiet: bool = False,
     **kwargs: Any,
 ) -> Tuple[str, float]:
@@ -2641,6 +2643,11 @@ def semantic_inference_on_geotiff(
         num_channels (int): Number of channels to use from the input image.
         num_classes (int): Number of classes in the model output.
         device (torch.device, optional): Device to run inference on.
+        probability_path (str, optional): Path to save probability map. If provided,
+            the normalized class probabilities will be saved as a multi-band raster.
+        probability_threshold (float, optional): Probability threshold for binary classification.
+            Only used when num_classes=2. If provided, pixels with class 1 probability >= threshold
+            are classified as class 1, otherwise class 0. If None (default), uses argmax.
         quiet (bool): If True, suppress progress bar. Defaults to False.
         **kwargs: Additional arguments.
 
@@ -2811,10 +2818,19 @@ def semantic_inference_on_geotiff(
                     / count_accumulator[valid_pixels]
                 )
 
-            # Take argmax to get final class predictions
-            mask[valid_pixels] = np.argmax(
-                normalized_probs[:, valid_pixels], axis=0
-            ).astype(np.uint8)
+            # Apply threshold for binary classification or use argmax
+            if probability_threshold is not None and num_classes == 2:
+                # Use threshold: classify as class 1 if probability >= threshold
+                mask[valid_pixels] = (
+                    normalized_probs[1, valid_pixels] >= probability_threshold
+                ).astype(np.uint8)
+                if not quiet:
+                    print(f"Using probability threshold: {probability_threshold}")
+            else:
+                # Take argmax to get final class predictions
+                mask[valid_pixels] = np.argmax(
+                    normalized_probs[:, valid_pixels], axis=0
+                ).astype(np.uint8)
 
             # Check class distribution in predictions (summary only)
             unique_classes, class_counts = np.unique(
@@ -2839,6 +2855,29 @@ def semantic_inference_on_geotiff(
         if not quiet:
             print(f"Saved prediction to {output_path}")
 
+        # Save probability map if requested
+        if probability_path is not None:
+            prob_dir = os.path.abspath(os.path.dirname(probability_path))
+            os.makedirs(prob_dir, exist_ok=True)
+
+            # Prepare probability output metadata
+            prob_meta = meta.copy()
+            prob_meta.update({"count": num_classes, "dtype": "float32"})
+
+            # Save normalized probabilities
+            with rasterio.open(probability_path, "w", **prob_meta) as dst:
+                for class_idx in range(num_classes):
+                    # Normalize probabilities
+                    prob_band = np.zeros((height, width), dtype=np.float32)
+                    prob_band[valid_pixels] = (
+                        prob_accumulator[class_idx, valid_pixels]
+                        / count_accumulator[valid_pixels]
+                    )
+                    dst.write(prob_band, class_idx + 1)
+
+            if not quiet:
+                print(f"Saved probability map to {probability_path}")
+
         return output_path, inference_time
 
 
@@ -2853,6 +2892,8 @@ def semantic_inference_on_image(
     num_classes: int = 2,
     device: Optional[torch.device] = None,
     binary_output: bool = True,
+    probability_path: Optional[str] = None,
+    probability_threshold: Optional[float] = None,
     quiet: bool = False,
     **kwargs: Any,
 ) -> Tuple[str, float]:
@@ -2870,6 +2911,11 @@ def semantic_inference_on_image(
         num_classes (int): Number of classes in the model output.
         device (torch.device, optional): Device to run inference on.
[…]
