+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "geoai-py"
-version = "0.13.1"
+version = "0.13.2"
 dynamic = [
     "dependencies",
 ]
@@ -44,7 +44,7 @@ universal = true
 
 
 [tool.bumpversion]
-current_version = "0.13.1"
+current_version = "0.13.2"
 commit = true
 tag = true
 
