### 2025-10-04 f94a924
Add instance segmentation example
