+++ mkdocs.yml
@@ -121,6 +121,7 @@ nav:
           - examples/data_visualization.ipynb
           - examples/train_object_detection_model.ipynb
           - examples/train_segmentation_model.ipynb
+          - examples/train_instance_segmentation_model.ipynb
           - examples/train_landcover_classification.ipynb
           - examples/train_building_footprints_usa.ipynb
           - examples/train_solar_panel_detection.ipynb
