### 2025-10-04 5cd10d7
Add instance segmentation example (#298)
