+++ geoai/train.py
@@ -433,7 +433,7 @@ def train_one_epoch(
             elapsed_time = time.time() - start_time
             if verbose:
                 print(
-                    f"Epoch: {epoch}, Batch: {i}/{len(data_loader)}, Loss: {losses.item():.4f}, Time: {elapsed_time:.2f}s"
+                    f"Epoch: {epoch + 1}, Batch: {i + 1}/{len(data_loader)}, Loss: {losses.item():.4f}, Time: {elapsed_time:.2f}s"
                 )
             start_time = time.time()
 
