### 2025-10-04 3aad180
Add supporting for multiple vector files
