### 2025-10-04 d670c6c
Add supporting for creating training data from multiple vector files (#299)
* Add instance segmentation example

* Add supporting for multiple vector files
