+++ mkdocs.yml
@@ -108,6 +108,7 @@ nav:
           - examples/edit_vector.ipynb
           - examples/image_chips.ipynb
           - examples/image_tiling.ipynb
+          - examples/create_training_data.ipynb
           - examples/building_footprints_usa.ipynb
           - examples/building_footprints_africa.ipynb
           - examples/building_footprints_china.ipynb
