+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "geoai-py"
-version = "0.13.2"
+version = "0.14.0"
 dynamic = [
     "dependencies",
 ]
@@ -44,7 +44,7 @@ universal = true
 
 
 [tool.bumpversion]
-current_version = "0.13.2"
+current_version = "0.14.0"
 commit = true
 tag = true
 
