+++ geoai/timm_train.py
@@ -1,0 +1,658 @@
+"""Module for training and fine-tuning models using timm (PyTorch Image Models) with remote sensing imagery."""
+
+import os
+from typing import Any, Callable, Dict, List, Optional, Tuple, Union
+
+import numpy as np
+import torch
+import torch.nn as nn
+from torch.utils.data import Dataset, DataLoader
+from tqdm import tqdm
+
+try:
+    import timm
+
+    TIMM_AVAILABLE = True
+except ImportError:
+    TIMM_AVAILABLE = False
+
+try:
+    import lightning.pytorch as pl
+    from lightning.pytorch.callbacks import ModelCheckpoint, EarlyStopping
+    from lightning.pytorch.loggers import CSVLogger
+
+    LIGHTNING_AVAILABLE = True
+except ImportError:
+    LIGHTNING_AVAILABLE = False
+
+
+def get_timm_model(
+    model_name: str = "resnet50",
+    num_classes: int = 10,
+    in_channels: int = 3,
+    pretrained: bool = True,
+    features_only: bool = False,
+    **kwargs: Any,
+) -> nn.Module:
+    """
+    Create a timm model with custom input channels for remote sensing imagery.
+
+    Args:
+        model_name (str): Name of the timm model (e.g., 'resnet50', 'efficientnet_b0',
+            'vit_base_patch16_224', 'convnext_base').
+        num_classes (int): Number of output classes for classification.
+        in_channels (int): Number of input channels (3 for RGB, 4 for RGBN, etc.).
+        pretrained (bool): Whether to use pretrained weights.
+        features_only (bool): If True, return feature extraction model without classifier.
+        **kwargs: Additional arguments to pass to timm.create_model.
+
+    Returns:
+        nn.Module: Configured timm model.
+
+    Raises:
+        ImportError: If timm is not installed.
+        ValueError: If model_name is not available in timm.
+    """
+    if not TIMM_AVAILABLE:
+        raise ImportError("timm is required. Install it with: pip install timm")
+
+    # Check if model exists
+    if model_name not in timm.list_models():
+        available_models = timm.list_models(pretrained=True)[:10]
+        raise ValueError(
+            f"Model '{model_name}' not found in timm. "
+            f"First 10 available models: {available_models}. "
+            f"See all models at: https://github.com/huggingface/pytorch-image-models"
+        )
+
+    # Create base model
+    model = timm.create_model(
+        model_name,
+        pretrained=pretrained,
+        num_classes=num_classes if not features_only else 0,
+        in_chans=in_channels,
+        features_only=features_only,
+        **kwargs,
+    )
+
+    return model
+
+
+def modify_first_conv_for_channels(
+    model: nn.Module,
+    in_channels: int,
+    pretrained_channels: int = 3,
+) -> nn.Module:
+    """
+    Modify the first convolutional layer of a model to accept different number of input channels.
+
+    This is useful when you have a pretrained model with 3 input channels but want to use
+    imagery with more channels (e.g., 4 for RGBN, or more for multispectral).
+
+    Args:
+        model (nn.Module): PyTorch model to modify.
+        in_channels (int): Desired number of input channels.
+        pretrained_channels (int): Number of channels in pretrained weights (usually 3).
+
+    Returns:
+        nn.Module: Modified model with updated first conv layer.
+    """
+    if in_channels == pretrained_channels:
+        return model
+
+    # Find the first conv layer (different models have different architectures)
+    first_conv_name = None
+    first_conv = None
+
+    # Common patterns for first conv layers
+    possible_names = ["conv1", "conv_stem", "patch_embed.proj", "stem.conv1"]
+
+    for name in possible_names:
+        try:
+            parts = name.split(".")
+            module = model
[…]
