### 2025-10-04 5a36e86
Add support for timm (#301)
