+++ geoai/timm_segment.py
@@ -1,0 +1,1069 @@
+"""Module for training semantic segmentation models using timm encoders with PyTorch Lightning."""
+
+import os
+from typing import Any, Callable, Dict, List, Optional, Tuple, Union
+
+import numpy as np
+import torch
+import torch.nn as nn
+from torch.utils.data import Dataset, DataLoader
+from tqdm import tqdm
+
+try:
+    import timm
+
+    TIMM_AVAILABLE = True
+except ImportError:
+    TIMM_AVAILABLE = False
+
+try:
+    import segmentation_models_pytorch as smp
+
+    SMP_AVAILABLE = True
+except ImportError:
+    SMP_AVAILABLE = False
+
+try:
+    import lightning.pytorch as pl
+    from lightning.pytorch.callbacks import ModelCheckpoint, EarlyStopping
+    from lightning.pytorch.loggers import CSVLogger
+
+    LIGHTNING_AVAILABLE = True
+except ImportError:
+    LIGHTNING_AVAILABLE = False
+
+
+class TimmSegmentationModel(pl.LightningModule):
+    """
+    PyTorch Lightning module for semantic segmentation using timm encoders with SMP decoders,
+    or pure timm models from Hugging Face Hub.
+    """
+
+    def __init__(
+        self,
+        encoder_name: str = "resnet50",
+        architecture: str = "unet",
+        num_classes: int = 2,
+        in_channels: int = 3,
+        encoder_weights: str = "imagenet",
+        learning_rate: float = 1e-3,
+        weight_decay: float = 1e-4,
+        freeze_encoder: bool = False,
+        loss_fn: Optional[nn.Module] = None,
+        class_weights: Optional[torch.Tensor] = None,
+        use_timm_model: bool = False,
+        timm_model_name: Optional[str] = None,
+        **decoder_kwargs: Any,
+    ):
+        """
+        Initialize TimmSegmentationModel.
+
+        Args:
+            encoder_name (str): Name of encoder (e.g., 'resnet50', 'efficientnet_b0').
+            architecture (str): Segmentation architecture ('unet', 'unetplusplus', 'deeplabv3',
+                'deeplabv3plus', 'fpn', 'pspnet', 'linknet', 'manet', 'pan').
+                Ignored if use_timm_model=True.
+            num_classes (int): Number of output classes.
+            in_channels (int): Number of input channels.
+            encoder_weights (str): Pretrained weights for encoder ('imagenet', 'ssl', 'swsl', None).
+            learning_rate (float): Learning rate for optimizer.
+            weight_decay (float): Weight decay for optimizer.
+            freeze_encoder (bool): Freeze encoder weights during training.
+            loss_fn (nn.Module, optional): Custom loss function. Defaults to CrossEntropyLoss.
+            class_weights (torch.Tensor, optional): Class weights for loss function.
+            use_timm_model (bool): If True, load a complete segmentation model from timm/HF Hub
+                instead of using SMP architecture. Defaults to False.
+            timm_model_name (str, optional): Name or path of timm model from HF Hub
+                (e.g., 'hf-hub:timm/segformer_b0.ade_512x512' or 'nvidia/mit-b0').
+                Only used if use_timm_model=True.
+            **decoder_kwargs: Additional arguments for decoder (only used with SMP).
+        """
+        super().__init__()
+
+        if not TIMM_AVAILABLE:
+            raise ImportError("timm is required. Install it with: pip install timm")
+
+        self.save_hyperparameters()
+
+        # Check if using a pure timm model from HF Hub
+        if use_timm_model:
+            if timm_model_name is None:
+                timm_model_name = encoder_name
+
+            # Load model from timm (supports HF Hub with 'hf-hub:' prefix)
+            try:
[…]
