### 2025-10-04 335e15a
Add support for timm (#301)
