+++ docs/timm_segment.md
@@ -1,0 +1,3 @@
+# timm_segment module
+
+::: geoai.timm_segment
