### 2025-10-05 2650906
Add more image display functions (#302)
* Add more image display functions

* Update notebook

* Add citations

* Update notebook example

* Update LICENSE file
