### 2025-10-05 f43a357
Fix local raster visualization bug
