### 2025-10-05 f0227f7
Add support for exporting COCO and YOLO formats (#305)
* Add support for exporting COCO and YOLO formats

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
