+++ mkdocs.yml
@@ -109,6 +109,7 @@ nav:
           - examples/image_chips.ipynb
           - examples/image_tiling.ipynb
           - examples/create_training_data.ipynb
+          - examples/export_training_data_formats.ipynb
           - examples/building_footprints_usa.ipynb
           - examples/building_footprints_africa.ipynb
           - examples/building_footprints_china.ipynb
