### 2025-10-05 9f778c1
Add training support for COCO and YOLO
