+++ geoai/train.py
@@ -35,6 +35,89 @@ except ImportError:
     SMP_AVAILABLE = False
 
 
+def parse_coco_annotations(
+    coco_json_path: str, images_dir: str, labels_dir: str
+) -> Tuple[List[str], List[str]]:
+    """
+    Parse COCO format annotations and return lists of image and label paths.
+
+    Args:
+        coco_json_path (str): Path to COCO annotations JSON file (instances.json).
+        images_dir (str): Directory containing image files.
+        labels_dir (str): Directory containing label mask files.
+
+    Returns:
+        Tuple[List[str], List[str]]: Lists of image paths and corresponding label paths.
+    """
+    import json
+
+    with open(coco_json_path, "r") as f:
+        coco_data = json.load(f)
+
+    # Create mapping from image_id to filename
+    image_files = []
+    label_files = []
+
+    for img_info in coco_data["images"]:
+        img_filename = img_info["file_name"]
+        img_path = os.path.join(images_dir, img_filename)
+
+        # Derive label filename (same as image filename)
+        label_path = os.path.join(labels_dir, img_filename)
+
+        if os.path.exists(img_path) and os.path.exists(label_path):
+            image_files.append(img_path)
+            label_files.append(label_path)
+
+    return image_files, label_files
+
+
+def parse_yolo_annotations(
+    data_dir: str, images_subdir: str = "images", labels_subdir: str = "labels"
+) -> Tuple[List[str], List[str]]:
+    """
+    Parse YOLO format annotations and return lists of image and label paths.
+
+    YOLO format structure:
+    - data_dir/images/: Contains image files (.tif, .png, .jpg)
+    - data_dir/labels/: Contains label masks (.tif, .png) and YOLO .txt files
+    - data_dir/classes.txt: Class names (one per line)
+
+    Args:
+        data_dir (str): Root directory containing YOLO-format data.
+        images_subdir (str): Subdirectory name for images. Defaults to 'images'.
+        labels_subdir (str): Subdirectory name for labels. Defaults to 'labels'.
+
+    Returns:
+        Tuple[List[str], List[str]]: Lists of image paths and corresponding label paths.
+    """
+    images_dir = os.path.join(data_dir, images_subdir)
+    labels_dir = os.path.join(data_dir, labels_subdir)
+
+    if not os.path.exists(images_dir):
+        raise FileNotFoundError(f"Images directory not found: {images_dir}")
+    if not os.path.exists(labels_dir):
+        raise FileNotFoundError(f"Labels directory not found: {labels_dir}")
+
+    # Get all image files
+    image_extensions = (".tif", ".tiff", ".png", ".jpg", ".jpeg")
+    image_files = []
+    label_files = []
+
+    for img_file in os.listdir(images_dir):
+        if img_file.lower().endswith(image_extensions):
+            img_path = os.path.join(images_dir, img_file)
+
+            # Find corresponding label mask (same filename)
+            label_path = os.path.join(labels_dir, img_file)
+
+            if os.path.exists(label_path):
+                image_files.append(img_path)
+                label_files.append(label_path)
+
+    return sorted(image_files), sorted(label_files)
+
+
 def get_instance_segmentation_model(
     num_classes: int = 2, num_channels: int = 3, pretrained: bool = True
 ) -> torch.nn.Module:
@@ -617,6 +700,7 @@ def train_MaskRCNN_model(
     images_dir: str,
     labels_dir: str,
     output_dir: str,
+    input_format: str = "directory",
     num_channels: int = 3,
     model: Optional[torch.nn.Module] = None,
     pretrained: bool = True,
@@ -640,9 +724,17 @@ def train_MaskRCNN_model(
[…]
