### 2025-10-05 e109cd1
Add timm support for COCO and YOLO
