+++ geoai/utils.py
@@ -3324,6 +3324,7 @@ def export_geotiff_tiles_batch(
     image_extensions=None,
     mask_extensions=None,
     match_by_name=True,
+    metadata_format="PASCAL_VOC",
 ) -> Dict[str, Any]:
     """
     Export georeferenced GeoTIFF tiles from images and masks.
@@ -3364,6 +3365,8 @@ def export_geotiff_tiles_batch(
         mask_extensions (list): List of mask file extensions to process (default: common raster/vector formats)
         match_by_name (bool): If True, match image and mask files by base filename.
             If False, match by sorted order (alphabetically). Only applies when masks_folder is used.
+        metadata_format (str): Annotation format - "PASCAL_VOC" (XML), "COCO" (JSON), or "YOLO" (TXT).
+            Default is "PASCAL_VOC".
 
     Returns:
         Dict[str, Any]: Dictionary containing batch processing statistics
@@ -3446,6 +3449,19 @@ def export_geotiff_tiles_batch(
     output_masks_dir = os.path.join(output_folder, "masks")
     os.makedirs(output_images_dir, exist_ok=True)
     os.makedirs(output_masks_dir, exist_ok=True)
+
+    # Create annotation directory based on metadata format
+    if metadata_format in ["PASCAL_VOC", "COCO"]:
+        ann_dir = os.path.join(output_folder, "annotations")
+        os.makedirs(ann_dir, exist_ok=True)
+
+    # Initialize COCO annotations dictionary
+    coco_annotations = None
+    if metadata_format == "COCO":
+        coco_annotations = {"images": [], "annotations": [], "categories": []}
+
+    # Initialize YOLO class set
+    yolo_classes = set() if metadata_format == "YOLO" else None
 
     # Get list of image files
     image_files = []
@@ -3600,6 +3616,8 @@ def export_geotiff_tiles_batch(
                 quiet=quiet,
                 mask_gdf=mask_gdf,  # Pass pre-loaded GeoDataFrame if using single mask
                 use_single_mask_file=use_single_mask_file,
+                metadata_format=metadata_format,
+                ann_dir=ann_dir if metadata_format in ["PASCAL_VOC", "COCO"] else None,
             )
 
             # Update counters
@@ -3621,6 +3639,23 @@ def export_geotiff_tiles_batch(
                 }
             )
 
+            # Aggregate COCO annotations
+            if metadata_format == "COCO" and "coco_data" in tiles_generated:
+                coco_data = tiles_generated["coco_data"]
+                # Add images and annotations
+                coco_annotations["images"].extend(coco_data.get("images", []))
+                coco_annotations["annotations"].extend(coco_data.get("annotations", []))
+                # Merge categories (avoid duplicates)
+                for cat in coco_data.get("categories", []):
+                    if not any(
+                        c["id"] == cat["id"] for c in coco_annotations["categories"]
+                    ):
+                        coco_annotations["categories"].append(cat)
+
+            # Aggregate YOLO classes
+            if metadata_format == "YOLO" and "yolo_classes" in tiles_generated:
+                yolo_classes.update(tiles_generated["yolo_classes"])
+
         except Exception as e:
             if not quiet:
                 print(f"ERROR processing {base_name}: {e}")
@@ -3628,6 +3663,33 @@ def export_geotiff_tiles_batch(
                 {"image": image_file, "mask": mask_file, "error": str(e)}
             )
             batch_stats["errors"] += 1
+
+    # Save aggregated COCO annotations
+    if metadata_format == "COCO" and coco_annotations:
+        import json
+
+        coco_path = os.path.join(ann_dir, "instances.json")
+        with open(coco_path, "w") as f:
[…]
