### 2025-10-05 4643bec
Update geoai/utils.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
