### 2025-10-05 b8a01bb
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
