### 2025-10-08 1c71381
Add comprehensive notebook example for Lightly Train functionality
