+++ geoai/geoai.py
@@ -32,6 +32,9 @@ from .train import (
     instance_segmentation,
     instance_segmentation_batch,
     instance_segmentation_inference_on_geotiff,
+    lightly_embed_images,
+    lightly_train_model,
+    load_lightly_pretrained_model,
     object_detection,
     object_detection_batch,
     semantic_segmentation,
