### 2025-10-09 f7d5ec7
Remove lightly-train from requirements
Remove 'lightly-train' from the requirements.
