### 2025-10-09 160f4af
Add 'lightly-train' to optional dependencies
