### 2025-10-09 7e1c572
Add lightly-train support (#291)
* Add lightly-train support

* Update geoai/train.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update geoai/train.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Add device parameter to load_lightly_pretrained_model

Updated load_lightly_pretrained_model to accept device parameter for loading model weights.

* Add comprehensive notebook example for Lightly Train functionality

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Remove Python 3.13 fr
[…]
