### 2025-10-09 caba451
Remove hardcoded date in plots (#312)
