+++ examples/jupytext.toml
@@ -1,1 +1,0 @@
-formats="ipynb,myst"
