### 2025-10-09 2abdeb4
Make masks optional for exporting tiles
