### 2025-10-09 10edc3d
Update geoai/utils.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
