### 2025-10-09 e22c73f
Make masks optional for exporting tiles (#313)
* Make masks optional for exporting tiles

* Update geoai/utils.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

---------

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
