### 2025-10-09 5ee7550
Fix lightly-train notebook errors (#314)
