+++ geoai/train.py
@@ -3932,7 +3932,10 @@ def lightly_train_model(
         model (str): Model architecture to train. Supports models from torchvision,
             timm, ultralytics, etc. Default is "torchvision/resnet50".
         method (str): Self-supervised learning method. Options include:
-            "dinov2_distillation" (recommended), "dinov2", "dino", "simclr".
+            - "simclr": Works with CNN models (ResNet, EfficientNet, etc.)
+            - "dino": Works with both CNNs and ViTs
+            - "dinov2": Requires ViT models only
+            - "dinov2_distillation": Requires ViT models only (recommended for ViTs)
             Default is "dinov2_distillation".
         epochs (int): Number of training epochs. Default is 100.
         batch_size (int): Batch size for training. Default is 64.
@@ -3944,16 +3947,30 @@ def lightly_train_model(
 
     Raises:
         ImportError: If lightly-train is not installed.
-        ValueError: If data_dir does not exist or is empty.
+        ValueError: If data_dir does not exist, is empty, or incompatible model/method.
+
+    Note:
+        Model/Method compatibility:
+        - CNN models (ResNet, EfficientNet): Use "simclr" or "dino"
+        - ViT models: Use "dinov2", "dinov2_distillation", or "dino"
 
     Example:
+        >>> # For CNN models (ResNet, EfficientNet)
         >>> model_path = lightly_train_model(
         ...     data_dir="path/to/unlabeled/images",
         ...     output_dir="path/to/output",
         ...     model="torchvision/resnet50",
+        ...     method="simclr",  # Use simclr for CNNs
         ...     epochs=50
         ... )
-        >>> print(f"Pretrained model saved to: {model_path}")
+        >>> # For ViT models
+        >>> model_path = lightly_train_model(
+        ...     data_dir="path/to/unlabeled/images",
+        ...     output_dir="path/to/output",
+        ...     model="timm/vit_base_patch16_224",
+        ...     method="dinov2",  # dinov2 requires ViT
+        ...     epochs=50
+        ... )
     """
     if not LIGHTLY_TRAIN_AVAILABLE:
         raise ImportError(
@@ -3973,11 +3990,46 @@ def lightly_train_model(
     if not image_files:
         raise ValueError(f"No image files found in {data_dir}")
 
+    # Validate model/method compatibility
+    is_vit_model = "vit" in model.lower() or "vision_transformer" in model.lower()
+
+    if method in ["dinov2", "dinov2_distillation"] and not is_vit_model:
+        raise ValueError(
+            f"Method '{method}' requires a Vision Transformer (ViT) model, but got '{model}'.\n"
+            f"Solutions:\n"
+            f"  1. Use a ViT model: model='timm/vit_base_patch16_224'\n"
+            f"  2. Use a CNN-compatible method: method='simclr' or method='dino'\n"
+            f"\nFor CNN models (ResNet, EfficientNet), use 'simclr' or 'dino'.\n"
+            f"For ViT models, use 'dinov2', 'dinov2_distillation', or 'dino'."
+        )
+
     print(f"Found {len(image_files)} images in {data_dir}")
     print(f"Starting self-supervised pretraining with {method} method...")
+    print(f"Model: {model}")
 
     # Create output directory
     os.makedirs(output_dir, exist_ok=True)
+
+    # Detect if running in notebook environment and set appropriate configuration
+    def is_notebook():
+        try:
+            from IPython import get_ipython
+
+            if get_ipython() is not None:
+                return True
+        except (ImportError, NameError):
+            pass
+        return False
+
+    # Force single-device training in notebooks to avoid DDP strategy issues
+    if is_notebook():
+        # Only override if not explicitly set by user
+        if "accelerator" not in kwargs:
[…]
