### 2025-10-10 7aa0b93
feat(super-resolution): add super-resolution module for geospatial imagery
Add a new super-resolution module implementing ESRGAN and SRCNN architectures for enhancing low-resolution satellite and aerial images. Includes support for multiple upscaling factors (2x, 4x, 8x), geospatial metadata preservation, tiled processing for large images, training and inference pipelines, and evaluation metrics. Also adds comprehensive documentation and unit tests.
