+++ tests/test_super_resolution.py
@@ -1,0 +1,257 @@
+#!/usr/bin/env python
+
+"""Tests for `geoai.super_resolution` module."""
+
+import os
+import tempfile
+import unittest
+from unittest.mock import patch
+
+import numpy as np
+import rasterio
+import torch
+
+from geoai import SuperResolutionModel, create_super_resolution_model
+
+from .test_fixtures import get_test_data_paths
+
+
+class TestSuperResolution(unittest.TestCase):
+    """Tests for super-resolution module."""
+
+    def setUp(self):
+        """Set up test fixtures."""
+        self.test_paths = get_test_data_paths()
+        self.test_raster_rgb = self.test_paths["test_raster_rgb"]
+
+        # Create a simple model for testing
+        self.sr_model = SuperResolutionModel(upscale_factor=2, model_type="srcnn")
+
+    def test_module_imports(self):
+        """Test that super_resolution module imports correctly."""
+        try:
+            from geoai import super_resolution
+            self.assertTrue(hasattr(super_resolution, "SuperResolutionModel"))
+        except ImportError as e:
+            self.fail(f"Failed to import super_resolution module: {e}")
+
+    def test_model_initialization(self):
+        """Test model initialization with different parameters."""
+        # Test ESRGAN model
+        model_esrgan = SuperResolutionModel(model_type="esrgan", upscale_factor=4)
+        self.assertEqual(model_esrgan.model_type, "esrgan")
+        self.assertEqual(model_esrgan.upscale_factor, 4)
+
+        # Test SRCNN model
+        model_srcnn = SuperResolutionModel(model_type="srcnn", upscale_factor=2)
+        self.assertEqual(model_srcnn.model_type, "srcnn")
+
+        # Test invalid model type
+        with self.assertRaises(ValueError):
+            SuperResolutionModel(model_type="invalid")
+
+    def test_device_setup(self):
+        """Test device setup."""
+        # Test CPU device
+        model_cpu = SuperResolutionModel(device="cpu")
+        self.assertEqual(str(model_cpu.device), "cpu")
+
+        # Test CUDA if available
+        if torch.cuda.is_available():
+            model_cuda = SuperResolutionModel(device="cuda")
+            self.assertEqual(str(model_cuda.device), "cuda")
+
+    def test_model_save_load(self):
+        """Test saving and loading model weights."""
+        with tempfile.TemporaryDirectory() as temp_dir:
+            model_path = os.path.join(temp_dir, "test_model.pth")
+
+            # Save model
+            self.sr_model.save_model(model_path)
+            self.assertTrue(os.path.exists(model_path))
+
+            # Load model
+            new_model = SuperResolutionModel(upscale_factor=2, model_type="srcnn")
+            new_model.load_model(model_path)
+
+            # Check that weights are loaded (basic check)
+            self.assertIsNotNone(new_model.model)
+
+    def test_enhance_image_basic(self):
+        """Test basic image enhancement functionality."""
+        if not os.path.exists(self.test_raster_rgb):
+            self.skipTest("Test raster file not found")
+
+        with tempfile.TemporaryDirectory() as temp_dir:
+            output_path = os.path.join(temp_dir, "enhanced.tif")
+
+            # Enhance image
+            result = self.sr_model.enhance_image(self.test_raster_rgb, output_path)
+
+            # Check output
+            self.assertEqual(result, output_path)
+            self.assertTrue(os.path.exists(output_path))
+
+            # Verify dimensions
+            with rasterio.open(self.test_raster_rgb) as src:
+                original_width, original_height = src.width, src.height
+
+            with rasterio.open(output_path) as dst:
+                enhanced_width, enhanced_height = dst.width, dst.height
+
+            # Should be 2x larger
+            self.assertEqual(enhanced_width, original_width * 2)
[…]
