+++ geoai/super_resolution.py
@@ -1,0 +1,456 @@
+"""Super-resolution module for enhancing geospatial imagery resolution using deep learning."""
+
+import logging
+import os
+from pathlib import Path
+from typing import Any, Dict, List, Optional, Tuple, Union
+
+import numpy as np
+import rasterio
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+from rasterio.transform import from_bounds
+from torch.utils.data import DataLoader, Dataset
+from tqdm import tqdm
+
+# Configure logging
+logging.basicConfig(level=logging.INFO)
+logger = logging.getLogger(__name__)
+
+
+class ResidualBlock(nn.Module):
+    """Residual block for ESRGAN generator."""
+
+    def __init__(self, channels: int):
+        super(ResidualBlock, self).__init__()
+        self.conv1 = nn.Conv2d(channels, channels, kernel_size=3, padding=1)
+        self.bn1 = nn.BatchNorm2d(channels)
+        self.conv2 = nn.Conv2d(channels, channels, kernel_size=3, padding=1)
+        self.bn2 = nn.BatchNorm2d(channels)
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        residual = x
+        out = F.relu(self.bn1(self.conv1(x)))
+        out = self.bn2(self.conv2(out))
+        out += residual
+        return out
+
+
+class ESRGANGenerator(nn.Module):
+    """ESRGAN-inspired generator network for super-resolution."""
+
+    def __init__(self, upscale_factor: int = 4, num_channels: int = 3):
+        super(ESRGANGenerator, self).__init__()
+        self.upscale_factor = upscale_factor
+
+        # Initial convolution
+        self.conv1 = nn.Conv2d(num_channels, 64, kernel_size=9, padding=4)
+
+        # Residual blocks
+        self.residual_blocks = nn.Sequential(*[ResidualBlock(64) for _ in range(16)])
+
+        # Second convolution
+        self.conv2 = nn.Conv2d(64, 64, kernel_size=3, padding=1)
+        self.bn2 = nn.BatchNorm2d(64)
+
+        # Upsampling layers
+        self.upsample = nn.Sequential()
+        for _ in range(int(np.log2(upscale_factor))):
+            self.upsample.append(nn.Conv2d(64, 256, kernel_size=3, padding=1))
+            self.upsample.append(nn.PixelShuffle(2))
+            self.upsample.append(nn.ReLU(inplace=True))
+
+        # Final convolution
+        self.conv3 = nn.Conv2d(64, num_channels, kernel_size=9, padding=4)
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        out1 = F.relu(self.conv1(x))
+        out = self.residual_blocks(out1)
+        out2 = self.bn2(self.conv2(out))
+        out = out1 + out2  # Skip connection
+        out = self.upsample(out)
+        out = self.conv3(out)
+        return out
+
+
+class SRCNN(nn.Module):
+    """SRCNN (Super-Resolution CNN) implementation."""
+
+    def __init__(self, upscale_factor: int = 2, num_channels: int = 3):
+        super(SRCNN, self).__init__()
+        self.upscale_factor = upscale_factor
+
+        self.layers = nn.Sequential(
+            nn.Conv2d(num_channels, 64, kernel_size=9, padding=4),
+            nn.ReLU(inplace=True),
+            nn.Conv2d(64, 32, kernel_size=1),
+            nn.ReLU(inplace=True),
+            nn.Conv2d(32, num_channels, kernel_size=5, padding=2)
+        )
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        # Bicubic upsampling first
[…]

[…]
