### 2025-10-11 21e732e
Add support for calculating segmentation metrics
