+++ geoai/utils.py
@@ -383,6 +383,392 @@ def calc_stats(dataset, divide_by: float = 1.0) -> Tuple[np.ndarray, np.ndarray]
     return accum_mean / len(files), accum_std / len(files)
 
 
+def calc_iou(
+    ground_truth: Union[str, np.ndarray, torch.Tensor],
+    prediction: Union[str, np.ndarray, torch.Tensor],
+    num_classes: Optional[int] = None,
+    ignore_index: Optional[int] = None,
+    smooth: float = 1e-6,
+    band: int = 1,
+) -> Union[float, np.ndarray]:
+    """
+    Calculate Intersection over Union (IoU) between ground truth and prediction masks.
+
+    This function computes the IoU metric for segmentation tasks. It supports both
+    binary and multi-class segmentation, and can handle numpy arrays, PyTorch tensors,
+    or file paths to raster files.
+
+    Args:
+        ground_truth (Union[str, np.ndarray, torch.Tensor]): Ground truth segmentation mask.
+            Can be a file path (str) to a raster file, numpy array, or PyTorch tensor.
+            For binary segmentation: shape (H, W) with values {0, 1}.
+            For multi-class segmentation: shape (H, W) with class indices.
+        prediction (Union[str, np.ndarray, torch.Tensor]): Predicted segmentation mask.
+            Can be a file path (str) to a raster file, numpy array, or PyTorch tensor.
+            Should have the same shape and format as ground_truth.
+        num_classes (Optional[int], optional): Number of classes for multi-class segmentation.
+            If None, assumes binary segmentation. Defaults to None.
+        ignore_index (Optional[int], optional): Class index to ignore in computation.
+            Useful for ignoring background or unlabeled pixels. Defaults to None.
+        smooth (float, optional): Smoothing factor to avoid division by zero.
+            Defaults to 1e-6.
+        band (int, optional): Band index to read from raster file (1-based indexing).
+            Only used when input is a file path. Defaults to 1.
+
+    Returns:
+        Union[float, np.ndarray]: For binary segmentation, returns a single float IoU score.
+            For multi-class segmentation, returns an array of IoU scores for each class.
+
+    Examples:
+        >>> # Binary segmentation with arrays
+        >>> gt = np.array([[0, 0, 1, 1], [0, 1, 1, 1]])
+        >>> pred = np.array([[0, 0, 1, 1], [0, 0, 1, 1]])
+        >>> iou = calc_iou(gt, pred)
+        >>> print(f"IoU: {iou:.4f}")
+        IoU: 0.8333
+
+        >>> # Multi-class segmentation
+        >>> gt = np.array([[0, 0, 1, 1], [0, 2, 2, 1]])
+        >>> pred = np.array([[0, 0, 1, 1], [0, 0, 2, 2]])
+        >>> iou = calc_iou(gt, pred, num_classes=3)
+        >>> print(f"IoU per class: {iou}")
+        IoU per class: [0.8333 0.5000 0.5000]
+
+        >>> # Using PyTorch tensors
+        >>> gt_tensor = torch.tensor([[0, 0, 1, 1], [0, 1, 1, 1]])
+        >>> pred_tensor = torch.tensor([[0, 0, 1, 1], [0, 0, 1, 1]])
+        >>> iou = calc_iou(gt_tensor, pred_tensor)
+        >>> print(f"IoU: {iou:.4f}")
+        IoU: 0.8333
+
+        >>> # Using raster file paths
+        >>> iou = calc_iou("ground_truth.tif", "prediction.tif", num_classes=3)
+        >>> print(f"Mean IoU: {np.nanmean(iou):.4f}")
+        Mean IoU: 0.7500
+    """
+    # Load from file if string path is provided
+    if isinstance(ground_truth, str):
+        with rasterio.open(ground_truth) as src:
[…]
