### 2025-10-11 72821ce
Update geoai/utils.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
