### 2025-10-11 956f000
Add support for calculating segmentation metrics (#317)
* Add support for calculating segmentation metrics

* Update geoai/utils.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update geoai/utils.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

---------

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
