+++ geoai/train.py
@@ -2033,14 +2033,14 @@ def get_smp_model(
         )
 
 
-def dice_coefficient(
+def f1_score(
     pred: torch.Tensor,
     target: torch.Tensor,
     smooth: float = 1e-6,
     num_classes: Optional[int] = None,
 ) -> float:
     """
-    Calculate Dice coefficient for segmentation (binary or multi-class).
+    Calculate F1 score (also known as Dice coefficient) for segmentation (binary or multi-class).
 
     Args:
         pred (torch.Tensor): Predicted mask (probabilities or logits) with shape [C, H, W] or [H, W].
@@ -2049,7 +2049,7 @@ def dice_coefficient(
         num_classes (int, optional): Number of classes. If None, auto-detected.
 
     Returns:
-        float: Mean Dice coefficient across all classes.
+        float: Mean F1 score across all classes.
     """
     # Convert predictions to class predictions
     if pred.dim() == 3:  # [C, H, W] format
@@ -2064,8 +2064,8 @@ def dice_coefficient(
     if num_classes is None:
         num_classes = max(pred_classes.max().item(), target.max().item()) + 1
 
-    # Calculate Dice for each class and average
-    dice_scores = []
+    # Calculate F1 score for each class and average
+    f1_scores = []
     for class_id in range(num_classes):
         pred_class = (pred_classes == class_id).float()
         target_class = (target == class_id).float()
@@ -2074,10 +2074,10 @@ def dice_coefficient(
         union = pred_class.sum() + target_class.sum()
 
         if union > 0:
-            dice = (2.0 * intersection + smooth) / (union + smooth)
-            dice_scores.append(dice.item())
+            f1 = (2.0 * intersection + smooth) / (union + smooth)
+            f1_scores.append(f1.item())
 
-    return sum(dice_scores) / len(dice_scores) if dice_scores else 0.0
+    return sum(f1_scores) / len(f1_scores) if f1_scores else 0.0
 
 
 def iou_coefficient(
@@ -2208,12 +2208,12 @@ def evaluate_semantic(
         num_classes (int): Number of classes for evaluation metrics.
 
     Returns:
-        dict: Evaluation metrics including loss, IoU, and Dice.
+        dict: Evaluation metrics including loss, IoU, and F1.
     """
     model.eval()
 
     total_loss = 0
-    dice_scores = []
+    f1_scores = []
     iou_scores = []
     num_batches = len(data_loader)
 
@@ -2230,17 +2230,17 @@ def evaluate_semantic(
 
             # Calculate metrics for each sample in the batch
             for pred, target in zip(outputs, targets):
-                dice = dice_coefficient(pred, target, num_classes=num_classes)
+                f1 = f1_score(pred, target, num_classes=num_classes)
                 iou = iou_coefficient(pred, target, num_classes=num_classes)
-                dice_scores.append(dice)
+                f1_scores.append(f1)
                 iou_scores.append(iou)
 
     # Calculate metrics
     avg_loss = total_loss / num_batches
-    avg_dice = sum(dice_scores) / len(dice_scores) if dice_scores else 0
+    avg_f1 = sum(f1_scores) / len(f1_scores) if f1_scores else 0
     avg_iou = sum(iou_scores) / len(iou_scores) if iou_scores else 0
 
-    return {"loss": avg_loss, "Dice": avg_dice, "IoU": avg_iou}
+    return {"loss": avg_loss, "F1": avg_f1, "IoU": avg_iou}
 
 
 def train_segmentation_model(
@@ -2550,7 +2550,7 @@ def train_segmentation_model(
         print(f"Using {torch.cuda.device_count()} GPUs for training")
[…]
