### 2025-10-11 716266a
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
