### 2025-10-11 45e8c12
Rename Dice score as F1 score (#318)
* Rename Dice score as F1 score

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
