### 2025-10-12 ce99a2a
fix(super-resolution): correct patch extraction, transform calculations, and SSIM metric
- Fix LR image patch extraction by scaling indices and patch size with upscale factor
- Update geospatial transform calculations to use affine transform attributes (c, f, e, a)
- Remove unsupported 'edsr' model type from docstring
- Fix SSIM calculation by replacing multichannel with channel_axis parameter
- Delete standalone test file and update import in proper test module
- Notebook added in examples folder
