+++ .github/workflows/ubuntu.yml
@@ -23,7 +23,7 @@ jobs:
             #   run: sudo apt-get update && sudo apt-get install -y gdal-bin libgdal-dev
 
             - name: Install uv
-              uses: astral-sh/setup-uv@v6
+              uses: astral-sh/setup-uv@v7
               with:
                   version: "0.4.16"
                   enable-cache: false
