### 2025-10-14 f7f2f0e
Add early stop patience for training (#322)
* Add early stop patience for training

* Update mkdocs jupyter version
