### 2025-10-14 e1557c9
Add early stop patience for training (#322)
* Add early stop patience for training

* Update mkdocs jupyter version
