+++ requirements_docs.txt
@@ -7,7 +7,7 @@ livereload
 mkdocs
 mkdocs-git-revision-date-localized-plugin
 mkdocs-git-revision-date-plugin
-mkdocs-jupyter>=0.24.0
+mkdocs-jupyter>=0.25.1
 mkdocs-material>=9.1.3
 mkdocs-pdf-export-plugin
 mkdocstrings
