### 2025-10-15 83b8637
Update utils.py
