### 2025-10-15 0640e12
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
