### 2025-10-16 705ed84
Add precision and recall metrics
