### 2025-10-16 b902382
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
