### 2025-10-16 0157094
Update geoai/utils.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
