### 2025-10-16 6283c0e
Add support for exporting probability image for each class
