### 2025-10-16 b9906ba
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
