### 2025-10-16 bd33994
Add support for exporting probability image for each class (#326)
* Add support for exporting probability image for each class

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
