### 2025-10-17 a3b2446
Add STAC agent for interactive search
