+++ geoai/agents/__init__.py
@@ -1,8 +1,10 @@
 from .geo_agents import (
     GeoAgent,
+    STACAgent,
     create_ollama_model,
     create_anthropic_model,
     create_openai_model,
     create_bedrock_model,
 )
 from .map_tools import MapTools
+from .stac_tools import STACTools
