+++ geoai/agents/geo_agents.py
@@ -5,7 +5,7 @@ import os
 import uuid
 from concurrent.futures import ThreadPoolExecutor
 from types import SimpleNamespace
-from typing import Any, Callable, Optional
+from typing import Any, Callable, Dict, Optional
 
 import boto3
 import ipywidgets as widgets
@@ -20,6 +20,7 @@ from strands.models.ollama import OllamaModel as _OllamaModel
 from strands.models.openai import OpenAIModel
 
 from .map_tools import MapSession, MapTools
+from .stac_tools import STACTools
 
 try:
     import nest_asyncio
@@ -593,3 +594,232 @@ class GeoAgent(Agent):
             [map_panel, right], layout=widgets.Layout(width="100%", gap="8px")
         )
         display(root)
+
+
+class STACAgent(Agent):
+    """AI agent for searching and interacting with STAC catalogs."""
+
+    def __init__(
+        self,
+        *,
+        model: str = "llama3.1",
+        system_prompt: str = "default",
+        endpoint: str = "https://planetarycomputer.microsoft.com/api/stac/v1",
+        model_args: dict = None,
+        **kwargs: Any,
+    ) -> None:
+        """Initialize the STAC Agent.
+
+        Args:
+            model: Model identifier (default: "llama3.1").
+            system_prompt: System prompt for the agent (default: "default").
+            endpoint: STAC API endpoint URL. Defaults to Microsoft Planetary Computer.
+            model_args: Additional keyword arguments for the model.
+            **kwargs: Additional keyword arguments for the Agent.
+        """
+        self.tools: STACTools = STACTools(endpoint=endpoint)
+
+        if model_args is None:
+            model_args = {}
+
+        # --- save a model factory we can call each turn ---
+        if model == "llama3.1":
+            self._model_factory: Callable[[], OllamaModel] = (
+                lambda: create_ollama_model(
+                    host="http://localhost:11434", model_id=model, **model_args
+                )
+            )
+        elif isinstance(model, str):
+            self._model_factory: Callable[[], BedrockModel] = (
+                lambda: create_bedrock_model(model_id=model, **model_args)
+            )
+        elif isinstance(model, OllamaModel):
+            # Extract configuration from existing OllamaModel and create new instances
+            model_id = model.config["model_id"]
+            host = model.host
+            client_args = model.client_args
+            self._model_factory: Callable[[], OllamaModel] = (
+                lambda: create_ollama_model(
+                    host=host, model_id=model_id, client_args=client_args, **model_args
+                )
+            )
+        elif isinstance(model, OpenAIModel):
+            # Extract configuration from existing OpenAIModel and create new instances
+            model_id = model.config["model_id"]
+            client_args = model.client_args.copy()
+            self._model_factory: Callable[[], OpenAIModel] = (
+                lambda mid=model_id, client_args=client_args: create_openai_model(
+                    model_id=mid, client_args=client_args, **model_args
+                )
+            )
+        elif isinstance(model, AnthropicModel):
+            # Extract configuration from existing AnthropicModel and create new instances
+            model_id = model.config["model_id"]
+            client_args = model.client_args.copy()
+            self._model_factory: Callable[[], AnthropicModel] = (
+                lambda mid=model_id, client_args=client_args: create_anthropic_model(
+                    model_id=mid, client_args=client_args, **model_args
+                )
+            )
+        else:
+            raise ValueError(f"Invalid model: {model}")
+
+        # build initial model (first turn)
+        model = self._model_factory()
+
+        if system_prompt == "default":
+            system_prompt = """
+            You are a STAC catalog search agent. When users ask to find or search for imagery:
+
[…]
