+++ geoai/agents/stac_tools.py
@@ -1,0 +1,321 @@
+"""Tools for STAC catalog search and interaction."""
+
+import json
+from typing import Any, Dict, List, Optional
+
+from strands import tool
+
+from ..download import pc_collection_list, pc_stac_search
+from .stac_models import (
+    LocationInfo,
+    STACAssetInfo,
+    STACCollectionInfo,
+    STACItemInfo,
+    STACSearchResult,
+)
+
+
+class STACTools:
+    """Collection of tools for searching and interacting with STAC catalogs."""
+
+    def __init__(
+        self,
+        endpoint: str = "https://planetarycomputer.microsoft.com/api/stac/v1",
+    ) -> None:
+        """Initialize STAC tools.
+
+        Args:
+            endpoint: STAC API endpoint URL. Defaults to Microsoft Planetary Computer.
+        """
+        self.endpoint = endpoint
+
+    @tool(
+        description="List and search available STAC collections from Planetary Computer"
+    )
+    def list_collections(
+        self,
+        filter_keyword: Optional[str] = None,
+        detailed: bool = False,
+    ) -> str:
+        """List available STAC collections from Planetary Computer.
+
+        Args:
+            filter_keyword: Optional keyword to filter collections (searches in id, title, description).
+            detailed: If True, return detailed information including temporal extent, license, etc.
+
+        Returns:
+            JSON string containing list of collections with their metadata.
+        """
+        try:
+            # Get collections using existing function
+            df = pc_collection_list(
+                endpoint=self.endpoint,
+                detailed=detailed,
+                filter_by=None,
+                sort_by="id",
+            )
+
+            # Apply keyword filtering if specified
+            if filter_keyword:
+                mask = df["id"].str.contains(filter_keyword, case=False, na=False) | df[
+                    "title"
+                ].str.contains(filter_keyword, case=False, na=False)
+                if "description" in df.columns:
+                    mask |= df["description"].str.contains(
+                        filter_keyword, case=False, na=False
+                    )
+                df = df[mask]
+
+            # Convert to list of dictionaries
+            collections = df.to_dict("records")
+
+            # Convert to structured models
+            collection_models = []
+            for col in collections:
+                collection_models.append(
+                    STACCollectionInfo(
+                        id=col.get("id", ""),
+                        title=col.get("title", ""),
+                        description=col.get("description"),
+                        license=col.get("license"),
+                        temporal_extent=col.get("temporal_extent"),
+                        spatial_extent=col.get("bbox"),
+                        providers=col.get("providers"),
+                        keywords=col.get("keywords"),
+                    )
+                )
+
+            result = {
+                "count": len(collection_models),
+                "filter_keyword": filter_keyword,
+                "collections": [c.model_dump() for c in collection_models],
+            }
+
+            return json.dumps(result, indent=2)
+
+        except Exception as e:
+            return json.dumps({"error": str(e)})
+
+    @tool(
+        description="Search for STAC items in a specific collection with optional filters"
+    )
+    def search_items(
+        self,
+        collection: str,
+        bbox: Optional[List[float]] = None,
+        time_range: Optional[str] = None,
+        query: Optional[Dict[str, Any]] = None,
+        limit: Optional[int] = 10,
+        max_items: Optional[int] = None,
+    ) -> str:
+        """Search for STAC items in the Planetary Computer catalog.
+
+        Args:
+            collection: Collection ID to search (e.g., "sentinel-2-l2a", "naip", "landsat-c2-l2").
+            bbox: Bounding box as [west, south, east, north] in WGS84 coordinates.
+                Example: [-122.5, 37.7, -122.3, 37.8] for San Francisco area.
+            time_range: Time range as "start/end" string in ISO format.
+                Example: "2024-09-01/2024-09-30" or "2024-09-01/2024-09-01" for single day.
[…]
