+++ geoai/download.py
@@ -1,5 +1,6 @@
 """This module provides functions to download data, including NAIP imagery and building data from Overture Maps."""
 
+import datetime
 import logging
 import os
 import subprocess
