### 2025-10-17 ab18111
Fix bbox and time_range errors
