+++ geoai/agents/geo_agents.py
@@ -607,6 +607,7 @@ class STACAgent(Agent):
         system_prompt: str = "default",
         endpoint: str = "https://planetarycomputer.microsoft.com/api/stac/v1",
         model_args: dict = None,
+        map_instance: Optional[leafmap.Map] = None,
         **kwargs: Any,
     ) -> None:
         """Initialize the STAC Agent.
@@ -616,9 +617,11 @@ class STACAgent(Agent):
             system_prompt: System prompt for the agent (default: "default").
             endpoint: STAC API endpoint URL. Defaults to Microsoft Planetary Computer.
             model_args: Additional keyword arguments for the model.
+            map_instance: Optional leafmap.Map instance for visualization. If None, creates a new one.
             **kwargs: Additional keyword arguments for the Agent.
         """
         self.stac_tools: STACTools = STACTools(endpoint=endpoint)
+        self.map_instance = map_instance if map_instance is not None else leafmap.Map()
 
         if model_args is None:
             model_args = {}
@@ -670,7 +673,7 @@ class STACAgent(Agent):
 
         if system_prompt == "default":
             system_prompt = """
-            You are a STAC catalog search agent. When users ask to find or search for imagery:
+            You are a STAC catalog search agent. Follow these steps EXACTLY for every search:
 
             STEP 1: Determine the collection ID
             Common collections:
@@ -686,27 +689,39 @@ class STACAgent(Agent):
             - Extract the most relevant collection ID from the response
             - Use that collection ID in the next step
 
-            STEP 2: If they mention a location name (e.g., "San Francisco", "New York"), call geocode_location(location_name).
-            Extract the bbox array from the response.
+            STEP 2: Check if user mentioned a location
+            If the query contains ANY location name (e.g., "Paris", "San Francisco", "New York", "California", "Tokyo"):
+            - REQUIRED: Call geocode_location("<location_name>") FIRST
+            - Extract the bbox array from the response
+            - This bbox MUST be included in search_items()
 
-            STEP 3: Call search_items() with ONLY these parameters:
+            Location words include: city names, state names, country names, region names, place names
+            Examples: "Paris", "New York", "California", "Seattle", "San Francisco", "Tokyo", "London"
+
+            STEP 3: Build search_items() parameters
             - collection: REQUIRED - the collection ID from Step 1
-            - bbox: OPTIONAL - ONLY if user explicitly mentioned a location name or coordinates
-            - time_range: OPTIONAL - ONLY if user explicitly mentioned dates, months, or years (e.g., "in 2024", "August 2024", "from June to September")
+            - bbox: REQUIRED if Step 2 found a location, otherwise omit entirely
+            - time_range: REQUIRED if query has dates/months/years, otherwise omit entirely
             - max_items: default to 1
 
-            CRITICAL RULES FOR PARAMETERS:
-            - Do NOT add time_range unless the user's query contains explicit temporal information (dates, months, years, or time periods)
-            - Do NOT add bbox unless the user's query contains a location name or coordinates
-            - Do NOT infer or assume dates/times that were not mentioned by the user
-            - Omit optional parameters entirely - do not pass null or empty values
+            CRITICAL RULES:
+            - If location mentioned → MUST call geocode_location() → MUST include bbox in search_items()
+            - If dates mentioned → MUST include time_range in search_items()
+            - If NO location → omit bbox parameter
+            - If NO dates → omit time_range parameter
 
             Examples:
-            - "Show me NAIP imagery for New York" → geocode + bbox, NO time_range
-            - "Find Sentinel-2 imagery in August 2024" → time_range, NO bbox
-            - "Find Landsat over Paris from June to July 2023" → geocode + bbox + time_range
-            - "Show me MODIS data for California" → list_collections(filter_keyword="modis") + geocode + bbox
[…]
