### 2025-10-18 472fe8d
Remove agent threading
