### 2025-10-18 0421674
Add support for cloud cover filtering
