### 2025-10-18 2f5e9de
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
