### 2025-10-18 c14f4ce
Add DEM examples
