+++ geoai/agents/geo_agents.py
@@ -639,11 +639,14 @@ class STACAgent(Agent):
         if system_prompt == "default":
             system_prompt = """You are a STAC search agent. Follow these steps EXACTLY:
 
-1. Determine collection ID:
-   - "sentinel-2-l2a" for Sentinel-2
+1. Determine collection ID based on data type:
+   - "sentinel-2-l2a" for Sentinel-2 or optical satellite imagery
    - "landsat-c2-l2" for Landsat
-   - "naip" for NAIP
-   - "sentinel-1-grd" for Sentinel-1/SAR
+   - "naip" for NAIP or aerial imagery (USA only)
+   - "sentinel-1-grd" for Sentinel-1 or SAR/radar
+   - "cop-dem-glo-30" for DEM, elevation, or terrain data
+   - "aster-l1t" for ASTER
+   For other data (e.g., MODIS, land cover): call list_collections(filter_keyword="<keyword>")
 
 2. If location mentioned:
    - Call geocode_location("<name>") FIRST
@@ -675,8 +678,12 @@ Examples:
 4. Return first item as JSON:
 {"id": "...", "collection": "...", "datetime": "...", "bbox": [...], "assets": [...]}
 
-If no items: {"error": "No items found"}
-NO explanatory text - ONLY JSON."""
+ERROR HANDLING:
+- If no items found: {"error": "No items found"}
+- If tool result too large: {"error": "Result too large, try narrower search"}
+- If tool error: {"error": "Search failed: <error message>"}
+
+CRITICAL: Return ONLY JSON. NO explanatory text, NO made-up data."""
 
         super().__init__(
             name="STAC Search Agent",
@@ -727,13 +734,21 @@ NO explanatory text - ONLY JSON."""
                     for item in content:
                         if isinstance(item, dict) and "toolResult" in item:
                             tool_result = item["toolResult"]
+                            # Check if this is a search_items result
+                            # We need to look at the preceding assistant message to identify the tool
                             if tool_result.get("status") == "success":
                                 result_content = tool_result.get("content", [])
                                 if isinstance(result_content, list) and result_content:
                                     text_content = result_content[0].get("text", "")
                                     try:
                                         payload = json.loads(text_content)
-                                        if "items" in payload and payload.get("items"):
+                                        # Return ANY search_items payload, even if items is empty
+                                        # This is identified by having "query" and "collection" fields
+                                        if (
+                                            "query" in payload
+                                            and "collection" in payload
+                                            and "items" in payload
+                                        ):
                                             return payload
                                     except json.JSONDecodeError:
                                         continue
@@ -847,6 +862,8 @@ NO explanatory text - ONLY JSON."""
         collection = item.get("collection")
         item_id = item.get("id")
 
+        kwargs = {}
+
         # Determine which assets to display based on collection
         assets = None
         if collection == "sentinel-2-l2a":
@@ -857,6 +874,14 @@ NO explanatory text - ONLY JSON."""
             assets = ["image"]  # NAIP 4-band imagery
         elif "sentinel-1" in collection:
             assets = ["vv"]  # Sentinel-1 VV polarization
+        elif collection == "cop-dem-glo-30":
+            assets = ["data"]
+            kwargs["colormap_name"] = "terrain"
+        elif collection == "aster-l1t":
+            assets = ["VNIR"]  # ASTER L1T imagery
+        elif collection == "3dep-lidar-hag":
+            assets = ["data"]
+            kwargs["colormap_name"] = "terrain"
         else:
             # Try to find common asset names
             if "assets" in item:
@@ -864,7 +889,7 @@ NO explanatory text - ONLY JSON."""
                     a.get("key") for a in item["assets"] if isinstance(a, dict)
                 ]
[…]
