### 2025-10-18 c890a4f
Add notebook example
