+++ mkdocs.yml
@@ -150,6 +150,7 @@ nav:
           - examples/DINOv3_wetlands.ipynb
           - examples/AlphaEarth.ipynb
           - examples/AI_agents.ipynb
+          - examples/STAC_agents.ipynb
           - examples/train_timm_classifier.ipynb
           - examples/train_timm_segmentation.ipynb
           - examples/lightly_train_self_supervised.ipynb
