+++ geoai/agents/stac_tools.py
@@ -1,0 +1,435 @@
+"""Tools for STAC catalog search and interaction."""
+
+import ast
+import json
+from typing import Any, Dict, List, Optional, Union
+
+from strands import tool
+
+from ..download import pc_collection_list, pc_stac_search
+from .stac_models import (
+    LocationInfo,
+    STACAssetInfo,
+    STACCollectionInfo,
+    STACItemInfo,
+    STACSearchResult,
+)
+
+
+class STACTools:
+    """Collection of tools for searching and interacting with STAC catalogs."""
+
+    # Common location cache to avoid repeated geocoding
+    _LOCATION_CACHE = {
+        "san francisco": {
+            "name": "San Francisco",
+            "bbox": [-122.5155, 37.7034, -122.3549, 37.8324],
+            "center": [-122.4194, 37.7749],
+        },
+        "new york": {
+            "name": "New York",
+            "bbox": [-74.0479, 40.6829, -73.9067, 40.8820],
+            "center": [-73.9352, 40.7306],
+        },
+        "new york city": {
+            "name": "New York City",
+            "bbox": [-74.0479, 40.6829, -73.9067, 40.8820],
+            "center": [-73.9352, 40.7306],
+        },
+        "paris": {
+            "name": "Paris",
+            "bbox": [2.2241, 48.8156, 2.4698, 48.9022],
+            "center": [2.3522, 48.8566],
+        },
+        "london": {
+            "name": "London",
+            "bbox": [-0.5103, 51.2868, 0.3340, 51.6919],
+            "center": [-0.1276, 51.5074],
+        },
+        "tokyo": {
+            "name": "Tokyo",
+            "bbox": [139.5694, 35.5232, 139.9182, 35.8173],
+            "center": [139.6917, 35.6895],
+        },
+        "los angeles": {
+            "name": "Los Angeles",
+            "bbox": [-118.6682, 33.7037, -118.1553, 34.3373],
+            "center": [-118.2437, 34.0522],
+        },
+        "chicago": {
+            "name": "Chicago",
+            "bbox": [-87.9401, 41.6445, -87.5241, 42.0230],
+            "center": [-87.6298, 41.8781],
+        },
+        "seattle": {
+            "name": "Seattle",
+            "bbox": [-122.4595, 47.4810, -122.2244, 47.7341],
+            "center": [-122.3321, 47.6062],
+        },
+        "california": {
+            "name": "California",
+            "bbox": [-124.4820, 32.5288, -114.1315, 42.0095],
+            "center": [-119.4179, 36.7783],
+        },
+        "las vegas": {
+            "name": "Las Vegas",
+            "bbox": [-115.3711, 35.9630, -114.9372, 36.2610],
+            "center": [-115.1400, 36.1177],
+        },
+    }
+
+    def __init__(
+        self,
+        endpoint: str = "https://planetarycomputer.microsoft.com/api/stac/v1",
+    ) -> None:
+        """Initialize STAC tools.
+
+        Args:
+            endpoint: STAC API endpoint URL. Defaults to Microsoft Planetary Computer.
+        """
+        self.endpoint = endpoint
+        # Runtime cache for geocoding results
+        self._geocode_cache = {}
+
+    @tool(
+        description="List and search available STAC collections from Planetary Computer"
+    )
+    def list_collections(
+        self,
+        filter_keyword: Optional[str] = None,
+        detailed: bool = False,
+    ) -> str:
+        """List available STAC collections from Planetary Computer.
+
+        Args:
+            filter_keyword: Optional keyword to filter collections (searches in id, title, description).
+            detailed: If True, return detailed information including temporal extent, license, etc.
+
+        Returns:
+            JSON string containing list of collections with their metadata.
+        """
+        try:
+            # Get collections using existing function
[…]
