### 2025-10-18 408e938
Add STAC agent for interactive search (#327)
* Add STAC agent for interactive search

* Update geoai/agents/geo_agents.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update geoai/agents/geo_agents.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Fix bbox and time_range errors

* Update system instructions

* Fix agent issue

* Remove agent threading

* Add support for cloud cover filtering

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Add DEM examples

* Add notebook example

*
[…]
