### 2025-10-18 e065a97
Add catalog search agent
