+++ geoai/agents/catalog_tools.py
@@ -1,0 +1,907 @@
+"""Tools for searching data catalogs."""
+
+import io
+import json
+from typing import Any, Dict, List, Optional, Union
+
+import pandas as pd
+import requests
+from strands import tool
+
+from .catalog_models import CatalogDatasetInfo, CatalogSearchResult, LocationInfo
+
+
+class CatalogTools:
+    """Collection of tools for searching and interacting with data catalogs."""
+
+    # Common location cache to avoid repeated geocoding
+    _LOCATION_CACHE = {
+        "san francisco": {
+            "name": "San Francisco",
+            "bbox": [-122.5155, 37.7034, -122.3549, 37.8324],
+            "center": [-122.4194, 37.7749],
+        },
+        "new york": {
+            "name": "New York",
+            "bbox": [-74.0479, 40.6829, -73.9067, 40.8820],
+            "center": [-73.9352, 40.7306],
+        },
+        "new york city": {
+            "name": "New York City",
+            "bbox": [-74.0479, 40.6829, -73.9067, 40.8820],
+            "center": [-73.9352, 40.7306],
+        },
+        "paris": {
+            "name": "Paris",
+            "bbox": [2.2241, 48.8156, 2.4698, 48.9022],
+            "center": [2.3522, 48.8566],
+        },
+        "london": {
+            "name": "London",
+            "bbox": [-0.5103, 51.2868, 0.3340, 51.6919],
+            "center": [-0.1276, 51.5074],
+        },
+        "tokyo": {
+            "name": "Tokyo",
+            "bbox": [139.5694, 35.5232, 139.9182, 35.8173],
+            "center": [139.6917, 35.6895],
+        },
+        "los angeles": {
+            "name": "Los Angeles",
+            "bbox": [-118.6682, 33.7037, -118.1553, 34.3373],
+            "center": [-118.2437, 34.0522],
+        },
+        "chicago": {
+            "name": "Chicago",
+            "bbox": [-87.9401, 41.6445, -87.5241, 42.0230],
+            "center": [-87.6298, 41.8781],
+        },
+        "seattle": {
+            "name": "Seattle",
+            "bbox": [-122.4595, 47.4810, -122.2244, 47.7341],
+            "center": [-122.3321, 47.6062],
+        },
+        "california": {
+            "name": "California",
+            "bbox": [-124.4820, 32.5288, -114.1315, 42.0095],
+            "center": [-119.4179, 36.7783],
+        },
+        "las vegas": {
+            "name": "Las Vegas",
+            "bbox": [-115.3711, 35.9630, -114.9372, 36.2610],
+            "center": [-115.1400, 36.1177],
+        },
+    }
+
+    def __init__(
+        self,
+        catalog_url: Optional[str] = None,
+        catalog_df: Optional[pd.DataFrame] = None,
+    ) -> None:
+        """Initialize CatalogTools.
+
+        Args:
+            catalog_url: URL to a catalog file (TSV, CSV, or JSON). If None, must provide catalog_df.
+            catalog_df: Pre-loaded catalog as a pandas DataFrame. If None, must provide catalog_url.
+        """
+        self.catalog_url = catalog_url
+        self._catalog_df = catalog_df
+        self._cache = {}
+        # Runtime cache for geocoding results
+        self._geocode_cache = {}
+
+        # Load catalog if URL provided
+        if catalog_url and catalog_df is None:
+            self._catalog_df = self._load_catalog(catalog_url)
+
+    def _load_catalog(self, url: str) -> pd.DataFrame:
+        """Load catalog from a URL.
+
+        Args:
+            url: URL to catalog file (TSV, CSV, or JSON).
+
+        Returns:
+            DataFrame containing catalog data.
+        """
+        # Check cache first
+        if url in self._cache:
[…]
