### 2025-10-18 3e36d13
Update geoai/agents/catalog_tools.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
