### 2025-10-18 80e6a77
Add catalog search agent (#328)
* Add catalog search agent

* Update geoai/agents/catalog_tools.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

---------

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
