+++ geoai/agents/geo_agents.py
@@ -18,6 +18,7 @@ from strands.models.anthropic import AnthropicModel
 from strands.models.ollama import OllamaModel as _OllamaModel
 from strands.models.openai import OpenAIModel
 
+from .catalog_tools import CatalogTools
 from .map_tools import MapSession, MapTools
 from .stac_tools import STACTools
 
@@ -1182,3 +1183,238 @@ CRITICAL: Return ONLY JSON. NO explanatory text, NO made-up data."""
             [map_panel, right], layout=widgets.Layout(width="100%", gap="8px")
         )
         display(root)
+
+
+class CatalogAgent(Agent):
+    """AI agent for searching data catalogs with natural language queries."""
+
+    def __init__(
+        self,
+        *,
+        model: str = "llama3.1",
+        system_prompt: str = "default",
+        catalog_url: Optional[str] = None,
+        catalog_df: Optional[Any] = None,
+        model_args: dict = None,
+        **kwargs: Any,
+    ) -> None:
+        """Initialize the Catalog Agent.
+
+        Args:
+            model: Model identifier (default: "llama3.1").
+            system_prompt: System prompt for the agent (default: "default").
+            catalog_url: URL to a catalog file (TSV, CSV, or JSON). Use JSON format for spatial search support.
+                Example: "https://raw.githubusercontent.com/opengeos/Earth-Engine-Catalog/refs/heads/master/gee_catalog.json"
+            catalog_df: Pre-loaded catalog as a pandas DataFrame.
+            model_args: Additional keyword arguments for the model.
+            **kwargs: Additional keyword arguments for the Agent.
+        """
+        self.catalog_tools: CatalogTools = CatalogTools(
+            catalog_url=catalog_url, catalog_df=catalog_df
+        )
+
+        if model_args is None:
+            model_args = {}
+
+        # --- save a model factory we can call each turn ---
+        if model == "llama3.1":
+            self._model_factory: Callable[[], OllamaModel] = (
+                lambda: create_ollama_model(
+                    host="http://localhost:11434", model_id=model, **model_args
+                )
+            )
+        elif isinstance(model, str):
+            self._model_factory: Callable[[], BedrockModel] = (
+                lambda: create_bedrock_model(model_id=model, **model_args)
+            )
+        elif isinstance(model, OllamaModel):
+            # Extract configuration from existing OllamaModel and create new instances
+            model_id = model.config["model_id"]
+            host = model.host
+            client_args = model.client_args
+            self._model_factory: Callable[[], OllamaModel] = (
+                lambda: create_ollama_model(
+                    host=host, model_id=model_id, client_args=client_args, **model_args
+                )
+            )
+        elif isinstance(model, OpenAIModel):
+            # Extract configuration from existing OpenAIModel and create new instances
+            model_id = model.config["model_id"]
+            client_args = model.client_args.copy()
+            self._model_factory: Callable[[], OpenAIModel] = (
+                lambda mid=model_id, client_args=client_args: create_openai_model(
+                    model_id=mid, client_args=client_args, **model_args
+                )
+            )
+        elif isinstance(model, AnthropicModel):
+            # Extract configuration from existing AnthropicModel and create new instances
+            model_id = model.config["model_id"]
+            client_args = model.client_args.copy()
+            self._model_factory: Callable[[], AnthropicModel] = (
+                lambda mid=model_id, client_args=client_args: create_anthropic_model(
+                    model_id=mid, client_args=client_args, **model_args
+                )
+            )
+        else:
+            raise ValueError(f"Invalid model: {model}")
+
+        # build initial model (first turn)
[…]
