### 2025-10-20 2fe070f
Show agent status during execution
