### 2025-10-20 5b91f4e
Show spinning icon during agent execution
