+++ geoai/agents/geo_agents.py
@@ -55,7 +55,7 @@ class UICallbackHandler:
                 # Make tool names more user-friendly
                 friendly_name = tool_name.replace("_", " ").title()
                 self.status_widget.value = (
-                    f"<link rel='stylesheet' href='https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css'>"
+                    f"<link rel='stylesheet' href='https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css' integrity='sha512-papm6H1QJQvQ1BM8DT6vKrrO5gkP7FpC18JNpDutZCRa14Q6gttxyPjdvVSxGInxjeaUp43EIBosHuLlHotE3A==' crossorigin='anonymous'>"
                     f"<span style='color:#0a7'>"
                     f"<i class='fas fa-spinner fa-spin' style='font-size:1.2em'></i> "
                     f"{friendly_name}...</span>"
