### 2025-10-20 a4ae944
Show spinning icon during agent execution (#333)
* Show spinning icon during agent execution

* Update geoai/agents/geo_agents.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* improve import

---------

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
