### 2025-10-27 06b2e11
Fix NameError when torchange not installed (#341)
* Initial plan

* Fix NameError for AnyChange when torchange is not installed

Co-authored-by: giswqs <5016453+giswqs@users.noreply.github.com>

---------

Co-authored-by: copilot-swe-agent[bot] <198982749+Copilot@users.noreply.github.com>
Co-authored-by: giswqs <5016453+giswqs@users.noreply.github.com>
