+++ geoai/train.py
@@ -1917,6 +1917,92 @@ class SemanticRandomHorizontalFlip:
         return image, mask
 
 
+class SemanticRandomVerticalFlip:
+    """Random vertical flip transform for semantic segmentation."""
+
+    def __init__(self, prob: float = 0.5) -> None:
+        self.prob = prob
+
+    def __call__(
+        self, image: torch.Tensor, mask: torch.Tensor
+    ) -> Tuple[torch.Tensor, torch.Tensor]:
+        if random.random() < self.prob:
+            # Flip image and mask along height dimension
+            image = torch.flip(image, dims=[1])
+            mask = torch.flip(mask, dims=[0])
+        return image, mask
+
+
+class SemanticRandomRotation90:
+    """Random 90-degree rotation transform for semantic segmentation."""
+
+    def __init__(self, prob: float = 0.5) -> None:
+        self.prob = prob
+
+    def __call__(
+        self, image: torch.Tensor, mask: torch.Tensor
+    ) -> Tuple[torch.Tensor, torch.Tensor]:
+        if random.random() < self.prob:
+            # Randomly rotate by 90, 180, or 270 degrees
+            k = random.randint(1, 3)
+            image = torch.rot90(image, k, dims=[1, 2])
+            mask = torch.rot90(mask, k, dims=[0, 1])
+        return image, mask
+
+
+class SemanticBrightnessAdjustment:
+    """Random brightness adjustment transform for semantic segmentation."""
+
+    def __init__(self, brightness_range: Tuple[float, float] = (0.8, 1.2), prob: float = 0.5) -> None:
+        """
+        Initialize brightness adjustment transform.
+
+        Args:
+            brightness_range: Tuple of (min, max) brightness factors.
+            prob: Probability of applying the transform.
+        """
+        self.brightness_range = brightness_range
+        self.prob = prob
+
+    def __call__(
+        self, image: torch.Tensor, mask: torch.Tensor
+    ) -> Tuple[torch.Tensor, torch.Tensor]:
+        if random.random() < self.prob:
+            # Apply random brightness adjustment
+            factor = self.brightness_range[0] + random.random() * (
+                self.brightness_range[1] - self.brightness_range[0]
+            )
+            image = torch.clamp(image * factor, 0, 1)
+        return image, mask
+
+
+class SemanticContrastAdjustment:
+    """Random contrast adjustment transform for semantic segmentation."""
+
+    def __init__(self, contrast_range: Tuple[float, float] = (0.8, 1.2), prob: float = 0.5) -> None:
+        """
+        Initialize contrast adjustment transform.
+
+        Args:
+            contrast_range: Tuple of (min, max) contrast factors.
+            prob: Probability of applying the transform.
+        """
+        self.contrast_range = contrast_range
+        self.prob = prob
+
+    def __call__(
+        self, image: torch.Tensor, mask: torch.Tensor
+    ) -> Tuple[torch.Tensor, torch.Tensor]:
+        if random.random() < self.prob:
+            # Apply random contrast adjustment
+            factor = self.contrast_range[0] + random.random() * (
+                self.contrast_range[1] - self.contrast_range[0]
+            )
+            mean = image.mean(dim=(1, 2), keepdim=True)
+            image = torch.clamp((image - mean) * factor + mean, 0, 1)
+        return image, mask
+
+
 def get_semantic_transform(train: bool) -> Any:
     """
     Get transforms for semantic segmentation data augmentation.
@@ -2388,6 +2474,8 @@ def train_segmentation_model(
     resize_mode: str = "resize",
     num_workers: Optional[int] = None,
     early_stopping_patience: Optional[int] = None,
[…]
