### 2025-10-27 e4754ad
Update geoai/train.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
