### 2025-10-27 5d0511e
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
