+++ requirements.txt
@@ -22,10 +22,10 @@ scikit-image
 scikit-learn
 # segment-geospatial
 timm
+tokenizers>=0.22.1
 torch
 # torchange
 torchgeo
 torchinfo
 tqdm
 transformers
-tokenizers>=0.22.1
