### 2025-10-27 725de96
Add support for custom transform (#343)
* Add support for custom transform

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update geoai/train.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Pin tokenizers version

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Pin transformer version

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.c
[…]
