### 2025-10-27 d7d1c0a
Integrate multiclean for cleaning up segmentation results
