+++ geoai/tools/multiclean.py
@@ -1,0 +1,358 @@
+"""
+MultiClean integration utilities for cleaning segmentation results.
+
+This module provides functions to use MultiClean (https://github.com/DPIRD-DMA/MultiClean)
+for post-processing segmentation masks and classification rasters. MultiClean performs
+morphological operations to smooth edges, remove noise islands, and fill gaps.
+"""
+
+import os
+from typing import Optional, List, Union, Tuple
+import numpy as np
+
+try:
+    from multiclean import clean_array
+
+    MULTICLEAN_AVAILABLE = True
+except ImportError:
+    MULTICLEAN_AVAILABLE = False
+
+try:
+    import rasterio
+
+    RASTERIO_AVAILABLE = True
+except ImportError:
+    RASTERIO_AVAILABLE = False
+
+
+def check_multiclean_available():
+    """
+    Check if multiclean is installed.
+
+    Raises:
+        ImportError: If multiclean is not installed.
+    """
+    if not MULTICLEAN_AVAILABLE:
+        raise ImportError(
+            "multiclean is not installed. "
+            "Please install it with: pip install multiclean "
+            "or: pip install geoai-py[extra]"
+        )
+
+
+def clean_segmentation_mask(
+    mask: np.ndarray,
+    class_values: Optional[Union[int, List[int]]] = None,
+    smooth_edge_size: int = 2,
+    min_island_size: int = 100,
+    connectivity: int = 8,
+    max_workers: Optional[int] = None,
+    fill_nan: bool = False,
+) -> np.ndarray:
+    """
+    Clean a segmentation mask using MultiClean morphological operations.
+
+    This function applies three cleaning operations:
+    1. Edge smoothing - Uses morphological opening to reduce jagged boundaries
+    2. Island removal - Eliminates small connected components (noise)
+    3. Gap filling - Replaces invalid pixels with nearest valid class
+
+    Args:
+        mask (np.ndarray): 2D numpy array containing segmentation classes.
+            Can be int or float. NaN values are treated as nodata.
+        class_values (int, list of int, or None): Target class values to process.
+            If None, auto-detects unique values from the mask. Defaults to None.
+        smooth_edge_size (int): Kernel width in pixels for edge smoothing.
+            Set to 0 to disable smoothing. Defaults to 2.
+        min_island_size (int): Minimum area (in pixels) for connected components.
+            Components with area strictly less than this are removed. Defaults to 100.
+        connectivity (int): Connectivity for component detection. Use 4 or 8.
+            8-connectivity considers diagonal neighbors. Defaults to 8.
+        max_workers (int, optional): Thread pool size for parallel processing.
+            If None, uses default threading. Defaults to None.
+        fill_nan (bool): Whether to fill NaN pixels with nearest valid class.
+            Defaults to False.
+
+    Returns:
+        np.ndarray: Cleaned 2D segmentation mask with same shape as input.
+
+    Raises:
+        ImportError: If multiclean is not installed.
+        ValueError: If mask is not 2D or if connectivity is not 4 or 8.
+
+    Example:
+        >>> import numpy as np
+        >>> from geoai.tools.multiclean import clean_segmentation_mask
+        >>> mask = np.random.randint(0, 3, (512, 512))
+        >>> cleaned = clean_segmentation_mask(
+        ...     mask,
+        ...     class_values=[0, 1, 2],
+        ...     smooth_edge_size=2,
+        ...     min_island_size=50
+        ... )
+    """
+    check_multiclean_available()
+
+    if mask.ndim != 2:
+        raise ValueError(f"Mask must be 2D, got shape {mask.shape}")
+
+    if connectivity not in [4, 8]:
[…]
