### 2025-10-27 cf8de29
Update geoai/tools/multiclean.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
