### 2025-10-27 67ea61b
Integrate multiclean for cleaning up segmentation results (#344)
* Integrate multiclean for cleaning up segmentation results

* Update geoai/tools/multiclean.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update geoai/tools/multiclean.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

---------

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
