### 2025-10-27 55f7b18
Integrate Integrate OmniCloud for cloud detection
