+++ mkdocs.yml
@@ -156,6 +156,7 @@ nav:
           - examples/train_timm_segmentation.ipynb
           - examples/lightly_train_self_supervised.ipynb
           - examples/clean_segmentation_results.ipynb
+          - examples/cloud_detection.ipynb
     - Workshops:
           - workshops/GeoAI_Workshop_2025.ipynb
           - workshops/AWS_2025.ipynb
