### 2025-10-27 bedfcdf
Integrate Integrate OmniCloud for cloud detection
