+++ geoai/tools/cloudmask.py
@@ -1,0 +1,431 @@
+"""
+OmniCloudMask integration for cloud and cloud shadow detection in satellite imagery.
+
+This module provides functions to use OmniCloudMask (https://github.com/DPIRD-DMA/OmniCloudMask)
+for detecting clouds and cloud shadows in satellite imagery. OmniCloudMask performs semantic
+segmentation to classify pixels into: Clear (0), Thick Cloud (1), Thin Cloud (2), Cloud Shadow (3).
+
+Supports Sentinel-2, Landsat 8, PlanetScope, and Maxar imagery at 10-50m resolution.
+"""
+
+import os
+from typing import Optional, List, Tuple, Dict, Any
+import numpy as np
+
+try:
+    from omnicloudmask import predict_from_array
+
+    OMNICLOUDMASK_AVAILABLE = True
+except ImportError:
+    OMNICLOUDMASK_AVAILABLE = False
+
+try:
+    import rasterio
+
+    RASTERIO_AVAILABLE = True
+except ImportError:
+    RASTERIO_AVAILABLE = False
+
+
+# Cloud mask class values
+CLEAR = 0
+THICK_CLOUD = 1
+THIN_CLOUD = 2
+CLOUD_SHADOW = 3
+
+
+def check_omnicloudmask_available():
+    """
+    Check if omnicloudmask is installed.
+
+    Raises:
+        ImportError: If omnicloudmask is not installed.
+    """
+    if not OMNICLOUDMASK_AVAILABLE:
+        raise ImportError(
+            "omnicloudmask is not installed. "
+            "Please install it with: pip install omnicloudmask "
+            "or: pip install geoai-py[extra]"
+        )
+
+
+def predict_cloud_mask(
+    image: np.ndarray,
+    batch_size: int = 1,
+    inference_device: str = "cpu",
+    inference_dtype: str = "fp32",
+    patch_size: int = 1000,
+    export_confidence: bool = False,
+    model_version: int = 3,
+) -> np.ndarray:
+    """
+    Predict cloud mask from a numpy array using OmniCloudMask.
+
+    This function classifies each pixel into one of four categories:
+    - 0: Clear
+    - 1: Thick Cloud
+    - 2: Thin Cloud
+    - 3: Cloud Shadow
+
+    Args:
+        image (np.ndarray): Input image array with shape (3, height, width) or (height, width, 3).
+            Should contain Red, Green, and NIR bands. Values should be in reflectance (0-1)
+            or digital numbers (0-10000 typical for Sentinel-2/Landsat).
+        batch_size (int): Number of patches to process per inference batch. Defaults to 1.
+        inference_device (str): Device for inference ('cpu', 'cuda', or 'mps'). Defaults to 'cpu'.
+        inference_dtype (str): Data type for inference ('fp32', 'fp16', or 'bf16').
+            'bf16' recommended for speed on compatible hardware. Defaults to 'fp32'.
+        patch_size (int): Size of patches for processing large images. Defaults to 1000.
+        export_confidence (bool): If True, also returns confidence map. Defaults to False.
+        model_version (int): Model version to use (1, 2, or 3). Defaults to 3.
+
+    Returns:
+        np.ndarray: Cloud mask array with shape (height, width) containing class predictions.
+            If export_confidence=True, returns tuple of (mask, confidence).
+
+    Raises:
+        ImportError: If omnicloudmask is not installed.
+        ValueError: If image has wrong shape or number of channels.
+
+    Example:
+        >>> import numpy as np
+        >>> from geoai.tools.cloudmask import predict_cloud_mask
+        >>> # Create synthetic image (3 bands: R, G, NIR)
[…]
