+++ geoai/utils.py
@@ -771,6 +771,296 @@ def calc_segmentation_metrics(
     return results
 
 
+class FocalLoss(torch.nn.Module):
+    """
+    Focal Loss for addressing class imbalance in segmentation.
+    
+    Reference: Lin, T. Y., Goyal, P., Girshick, R., He, K., & Dollár, P. (2017).
+    Focal loss for dense object detection. ICCV.
+    """
+    
+    def __init__(self, alpha=1.0, gamma=2.0, ignore_index=-100, reduction='mean', weight=None):
+        """
+        Initialize Focal Loss.
+        
+        Args:
+            alpha (float): Weighting factor in range (0,1) to balance positive vs negative examples
+                or a list of weights for each class. Default: 1.0
+            gamma (float): Exponent of the modulating factor (1 - p_t)^gamma. Default: 2.0
+            ignore_index (int or bool): Specifies a target value that is ignored and does not 
+                contribute to the input gradient. If False, no pixels are ignored. Default: -100
+            reduction (str): Specifies the reduction to apply to the output: 
+                'none' | 'mean' | 'sum'. Default: 'mean'
+            weight (torch.Tensor, optional): Manual rescaling weight for each class. Default: None
+        """
+        super(FocalLoss, self).__init__()
+        self.alpha = alpha
+        self.gamma = gamma
+        self.ignore_index = ignore_index
+        self.reduction = reduction
+        self.weight = weight  # Class weights tensor
+        
+    def forward(self, inputs, targets):
+        """
+        Compute focal loss.
+        
+        Args:
+            inputs: Tensor of shape [N, C, H, W] where C is number of classes
+            targets: Tensor of shape [N, H, W] with class indices
+            
+        Returns:
+            torch.Tensor: Computed focal loss
+        """
+        import torch.nn.functional as F
+        
+        # Handle ignore_index parameter - if False, disable ignoring
+        if self.ignore_index is False:
+            ignore_idx = -100  # Use a value that won't match any target
+        else:
+            ignore_idx = self.ignore_index
+        
+        # Apply log_softmax to get log probabilities
+        log_pt = F.log_softmax(inputs, dim=1)
+        
+        # Use NLL loss with ignore_index and class weights to handle ignored pixels properly
+        ce_loss = F.nll_loss(log_pt, targets, ignore_index=ignore_idx, weight=self.weight, reduction='none')
+        
+        # Get probabilities by exponentiating the negative cross entropy
+        pt = torch.exp(-ce_loss)
+        
+        # Apply focal loss formula: FL = -alpha * (1-pt)^gamma * log(pt)
+        focal_loss = self.alpha * (1 - pt) ** self.gamma * ce_loss
+        
+        # Handle reduction
+        if self.reduction == 'mean':
+            # Only average over non-ignored pixels
+            if ignore_idx != -100:
+                valid_mask = (targets != ignore_idx)
+                if valid_mask.sum() > 0:
+                    return focal_loss[valid_mask].mean()
+                else:
+                    return torch.tensor(0.0, device=inputs.device, requires_grad=True)
+            else:
+                return focal_loss.mean()
+        elif self.reduction == 'sum':
+            if ignore_idx != -100:
+                valid_mask = (targets != ignore_idx)
+                return focal_loss[valid_mask].sum()
+            else:
+                return focal_loss.sum()
+        else:
+            return focal_loss
+
+
+def get_loss_function(loss_name: str, 
+                     ignore_index: Union[int, bool] = -100, 
+                     num_classes: int = 2,
+                     use_class_weights: bool = False,
+                     class_weights: Optional[torch.Tensor] = None,
+                     focal_alpha: float = 1.0,
+                     focal_gamma: float = 2.0,
[…]
