+++ examples/test_focal_loss.py
@@ -1,0 +1,225 @@
+#!/usr/bin/env python3
+"""
+Example script demonstrating the new FocalLoss and class weighting functionality for landcover classification.
+
+This script shows how to:
+1. Use FocalLoss for handling class imbalance
+2. Compute class weights from label files
+3. Use get_loss_function helper
+4. Apply min_feature_ratio filtering in tile generation
+"""
+
+import torch
+import numpy as np
+from geoai.utils import FocalLoss, get_loss_function, compute_class_weights
+
+
+def example_focal_loss():
+    """Demonstrate FocalLoss usage."""
+    print("=" * 60)
+    print("Example 1: Using FocalLoss for Class Imbalance")
+    print("=" * 60)
+    
+    # Create sample data
+    batch_size, num_classes, height, width = 2, 5, 64, 64
+    inputs = torch.randn(batch_size, num_classes, height, width)
+    
+    # Simulate imbalanced labels (class 0 is dominant)
+    targets = torch.zeros(batch_size, height, width, dtype=torch.long)
+    # Add some minority class pixels
+    targets[:, 10:20, 10:20] = 1  # Some class 1
+    targets[:, 30:35, 30:35] = 2  # Some class 2
+    
+    # Standard CrossEntropy Loss
+    ce_loss_fn = torch.nn.CrossEntropyLoss()
+    ce_loss = ce_loss_fn(inputs, targets)
+    print(f"\n📊 Standard CrossEntropy Loss: {ce_loss.item():.4f}")
+    
+    # Focal Loss (focuses on hard examples)
+    focal_loss_fn = FocalLoss(alpha=1.0, gamma=2.0)
+    focal_loss = focal_loss_fn(inputs, targets)
+    print(f"🎯 Focal Loss (alpha=1.0, gamma=2.0): {focal_loss.item():.4f}")
+    
+    # Focal Loss with higher gamma (more focus on hard examples)
+    focal_loss_fn_high = FocalLoss(alpha=1.0, gamma=3.0)
+    focal_loss_high = focal_loss_fn_high(inputs, targets)
+    print(f"🎯 Focal Loss (alpha=1.0, gamma=3.0): {focal_loss_high.item():.4f}")
+    
+    print("\n✅ Focal loss helps focus training on hard-to-classify pixels!")
+
+
+def example_get_loss_function():
+    """Demonstrate get_loss_function helper."""
+    print("\n" + "=" * 60)
+    print("Example 2: Using get_loss_function Helper")
+    print("=" * 60)
+    
+    num_classes = 5
+    
+    # Get CrossEntropy loss
+    print("\n🎯 Creating CrossEntropy Loss:")
+    ce_loss = get_loss_function("crossentropy", num_classes=num_classes)
+    
+    # Get Focal loss with custom parameters
+    print("\n🎯 Creating Focal Loss with custom parameters:")
+    focal_loss = get_loss_function(
+        "focal", 
+        num_classes=num_classes,
+        focal_alpha=0.25,
+        focal_gamma=2.0
+    )
+    
+    # Get loss with class weights
+    class_weights = torch.tensor([1.0, 2.0, 3.0, 4.0, 5.0])  # Boost rare classes
+    print("\n🎯 Creating Focal Loss with class weights:")
+    weighted_loss = get_loss_function(
+        "focal",
+        num_classes=num_classes,
+        use_class_weights=True,
+        class_weights=class_weights,
+        focal_alpha=0.25,
+        focal_gamma=2.0
+    )
+    
+    print("\n✅ get_loss_function makes it easy to select and configure loss functions!")
+
+
+def example_ignore_index():
+    """Demonstrate flexible ignore_index handling."""
[…]
