### 2025-10-30 01f2279
Add comprehensive documentation for landcover training enhancements
Co-authored-by: giswqs <5016453+giswqs@users.noreply.github.com>
