+++ LANDCOVER_ENHANCEMENTS.md
@@ -1,0 +1,284 @@
+# GeoAI Landcover Training Enhancements - Implementation Summary
+
+## Overview
+
+This PR implements significant enhancements to the GeoAI package to support training with discrete class landcover data, particularly for scenarios with severe class imbalance and sparse training datasets.
+
+## Problem Statement
+
+When training landcover classification models with discrete/sparse data, several challenges arise:
+
+1. **Severe Class Imbalance**: Common classes (forests >40%) dominate rare classes (roads <1%)
+2. **Sparse Training Data**: Many tiles contain mostly background with little labeled content
+3. **Inflexible Loss Functions**: Standard CrossEntropy doesn't focus on hard-to-classify pixels
+4. **Manual Weight Tuning**: No automated way to compute balanced class weights
+
+## Implemented Solutions
+
+### 1. FocalLoss Class (`geoai/utils.py`)
+
+**Location**: Lines 774-853
+
+A custom PyTorch implementation of Focal Loss for handling severe class imbalance.
+
+**Key Features**:
+- Focuses training on "hard examples" while down-weighting easy predictions
+- Configurable `focal_alpha` (class balancing) and `focal_gamma` (hard example focus)
+- Supports class weights via `weight` parameter
+- Flexible `ignore_index` (can be int or False)
+- Compatible with standard PyTorch training loops
+
+**Reference**: Lin, T. Y., et al. (2017). "Focal loss for dense object detection." ICCV.
+
+**Usage**:
+```python
+from geoai.utils import FocalLoss
+
+loss_fn = FocalLoss(alpha=1.0, gamma=2.0)
+loss = loss_fn(model_output, targets)
+```
+
+### 2. get_loss_function() Helper (`geoai/utils.py`)
+
+**Location**: Lines 855-908
+
+A convenient factory function for creating configured loss functions.
+
+**Supported Loss Functions**:
+- `'crossentropy'`: Standard CrossEntropyLoss
+- `'focal'`: FocalLoss with custom parameters
+
+**Key Features**:
+- Unified interface for loss function creation
+- Automatic device management for class weights
+- Support for ignore_index as int or False
+- Informative console output
+
+**Usage**:
+```python
+from geoai.utils import get_loss_function
+
+loss_fn = get_loss_function(
+    "focal",
+    num_classes=5,
+    use_class_weights=True,
+    class_weights=weights,
+    focal_alpha=0.25,
+    focal_gamma=2.0
+)
+```
+
+### 3. compute_class_weights() Function (`geoai/utils.py`)
+
+**Location**: Lines 910-1059
+
+Automatically computes class weights for imbalanced datasets with extensive configuration options.
+
+**Modes of Operation**:
+1. **Inverse Frequency Mode** (default): Auto-computes weights based on pixel abundance
+2. **Pure Custom Mode**: Manual weight control without frequency calculation
+
+**Key Features**:
+- Scans all label files (.tif, .tiff, .png, .jpg, .jpeg) in a directory
+- Computes inverse frequency weights automatically
+- Supports custom multipliers to fine-tune specific classes
+- Maximum weight capping to prevent training instability
+- Handles ignore_index properly
+- Informative progress reporting
+
+**Usage**:
+```python
+from geoai.utils import compute_class_weights
+
+# Automatic inverse frequency weights
+weights = compute_class_weights(
+    labels_dir="path/to/labels/",
+    num_classes=5,
+    ignore_index=-100
+)
+
+# With custom multipliers
+weights = compute_class_weights(
+    labels_dir="path/to/labels/",
+    num_classes=5,
[…]
