### 2025-10-30 4a25a55
Fix syntax error in scheduler initialization
