+++ geoai/timm_segment.py
@@ -241,7 +241,7 @@ class TimmSegmentationModel(pl.LightningModule):
         )
 
         scheduler = torch.optim.lr_scheduler.ReduceLROnPlateau(
-            optimizer, mode="min", factor=0.5, patience=5, verbose=True
+            optimizer, mode="min", factor=0.5, patience=5,
         )
 
         return {
