### 2025-10-30 cfdcb78
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
