+++ geoai/timm_segment.py
@@ -241,7 +241,10 @@ class TimmSegmentationModel(pl.LightningModule):
         )
 
         scheduler = torch.optim.lr_scheduler.ReduceLROnPlateau(
-            optimizer, mode="min", factor=0.5, patience=5,
+            optimizer,
+            mode="min",
+            factor=0.5,
+            patience=5,
         )
 
         return {
