### 2025-10-30 5f9fdb4
Fix syntax error in scheduler initialization (#348)
* Fix syntax error in scheduler initialization

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
