### 2025-11-04 ef87468
Add ASA workshop notebook (#349)
* Add ASA workshop notebook

* Update docs/workshops/ASA_2025.ipynb

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

---------

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
