### 2025-11-05 a5111a5
Update ASA workshop notebook (#351)
