### 2025-11-05 d6b08d6
Landcover Classification for Discrete Data
Included support for focal loss function including gamma and alpha parameters, per class assigned weights, inverse frequency weighting, ignore an index during training (background for instance), weighted IOU validation metric for best model selection, Boundary alpha to control this weighted IOU consideration of boundary in validation, minimum feature ratio for tile selection, tracking of weighted vs standard IOU or vice versa
