### 2025-11-05 20e0f76
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
