+++ geoai/train.py
@@ -12,9 +12,9 @@ MAJOR MODIFICATIONS SUMMARY:
    - Standard IoU: Unweighted average across all classes (original behavior preserved)
    - Per-class frequency-weighted IoU: Classes weighted by pixel frequency
    - Boundary-weighted IoU (wIoU): Distance-based weighting focusing on class boundaries
-   
+
    BOUNDARY-WEIGHTED IoU (wIoU) IMPLEMENTATION:
-   - Source: "Weighted Intersection over Union (wIoU): A New Evaluation Metric for 
+   - Source: "Weighted Intersection over Union (wIoU): A New Evaluation Metric for
      Image Segmentation" by Yeong-Jun Cho (arXiv:2107.09858)
    - Algorithm: Exponentially weights pixels based on distance from class boundaries
      * Pixels near boundaries receive higher weights (critical for accuracy)
@@ -26,13 +26,13 @@ MAJOR MODIFICATIONS SUMMARY:
      * Applies exponential decay with configurable alpha parameter
    - Use case: Boundary-critical applications (roads, building footprints, property lines)
    - Configuration: Set validation_iou_mode="boundary_weighted" and boundary_alpha (0.01-100)
-   
+
    IGNORE_INDEX SUPPORT:
    - Added proper handling of ignore_index in all IoU calculation modes
    - Pixels with ignore_index value are excluded from metric calculations
    - Ensures accurate evaluation when training with unlabeled/masked regions
    - Applies to both standard IoU and weighted variants
-   
+
    BACKWARD COMPATIBILITY:
    - New features activated only when optional parameters provided
 
@@ -44,7 +44,7 @@ MAJOR MODIFICATIONS SUMMARY:
    - Usage: loss_function='focal', focal_alpha=1.0, focal_gamma=2.0
 
 2. CLASS WEIGHT COMPUTATION (Lines 216-315):
-   - compute_class_weights function: Enhanced inverse frequency calculation  
+   - compute_class_weights function: Enhanced inverse frequency calculation
    - Custom multiplier support: Manual adjustment of specific class weights
    - Maximum weight capping: Prevent extreme weights (configurable, default 50.0)
     - Reasoning: Landcover data has severe imbalance (roads <1% vs forest >40%)
@@ -68,7 +68,7 @@ MAJOR MODIFICATIONS SUMMARY:
 KEY BENEFITS:
 =============
 - Handles severe class imbalance in landcover data (focal loss + class weights)
-- Prevents unstable training from extreme class weights (capping mechanism)  
+- Prevents unstable training from extreme class weights (capping mechanism)
 - Reduces manual logging errors (automatic callback system)
 - Supports multiple loss functions (crossentropy, focal, dice)
 - Provides fine-grained class weight control (custom multipliers)
@@ -115,19 +115,21 @@ except ImportError:
 class FocalLoss(torch.nn.Module):
     """
     Focal Loss for addressing class imbalance in segmentation.
-    
+
     Reference: Lin, T. Y., Goyal, P., Girshick, R., He, K., & Dollár, P. (2017).
     Focal loss for dense object detection. ICCV.
     """
-    
-    def __init__(self, alpha=1.0, gamma=2.0, ignore_index=-100, reduction='mean', weight=None):
+
+    def __init__(
+        self, alpha=1.0, gamma=2.0, ignore_index=-100, reduction="mean", weight=None
+    ):
         super(FocalLoss, self).__init__()
         self.alpha = alpha
         self.gamma = gamma
         self.ignore_index = ignore_index
         self.reduction = reduction
         self.weight = weight  # Class weights tensor
-        
+
     def forward(self, inputs, targets):
         """
         Args:
[…]
