### 2025-11-05 90defb2
Add ChangeDetection import and update installation docs
- Import ChangeDetection class in geoai/__init__.py with proper error handling
- Update README.md with instructions for installing optional dependencies
- Update docs/index.md with same installation instructions

Co-authored-by: giswqs <5016453+giswqs@users.noreply.github.com>
