### 2025-11-05 9246afb
Add tests for ChangeDetection import behavior
- Add comprehensive tests for change_detection module import
- Test graceful handling of missing torchange dependency
- Test error message quality when torchange is not installed

Co-authored-by: giswqs <5016453+giswqs@users.noreply.github.com>
