### 2025-11-05 45ca4b3
Update CANVAS workshop notebook (#356)
* Update CANVAS workshop notebook

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update website

* Remove duplicate link

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
