+++ mkdocs.yml
@@ -161,7 +161,7 @@ nav:
           - workshops/GeoAI_Workshop_2025.ipynb
           - workshops/AWS_2025.ipynb
           - workshops/TNView_2025.ipynb
-          - workshops/ASA_2025.ipynb
+          - workshops/CANVAS_2025.ipynb
     - API Reference:
           - change_detection module: change_detection.md
           - classify module: classify.md
