### 2025-11-05 153365d
Update view_raster function to use folium backend
