### 2025-11-05 4622821
Pin leafmap min version (#358)
