### 2025-11-05 726b73c
Improve view_vector function
