### 2025-11-05 d6e4132
Improve view_vector_interactive function (#359)
* Pin leafmap min version

* Improve view_vector function

* Pin leafmap version

* Fix bug
