### 2025-11-13 #9025 @mengqi-z
@RaczeQ - Thank you for your reply and for agreeing to review the submission!

The review aims to verify that the software functions as described in the accompanying paper and documentation. As a reviewer, your responsibilities include ensuring that the software runs correctly without errors, that the repository contains all essential components (e.g., license, contribution guide, automated tests, and sufficient documentation), and that the paper is well written and clearly presents the summary, statement of need, state of the field, and software functionality.

Once the review officially begins (we are currently in the pre-review issue), the process will be much more straightforward. You will be able to create a review checklist using our [EditorialBot](https://joss.readthedocs.io/en/latest/editorial_bot.html) (`@editorialbot`) which will outline all the items you should evaluate as part of the review process. You can check off any items that meet your expectations. For those that do not, the review will become an iterative process between you and the author to resolve the issues until they meet your satisfaction. The link is the list of items to be reviewed: https://joss.readthedocs.io/en/latest/review_checklist.html
