### 2025-11-13 c278566
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
