+++ geoai/utils.py
@@ -3575,7 +3575,9 @@ def export_geotiff_tiles(
                         if not np.issubdtype(img_for_aug.dtype, np.uint8):
                             # If image is float, scale to 0-255 and convert to uint8
                             if np.issubdtype(img_for_aug.dtype, np.floating):
-                                img_for_aug = (img_for_aug * 255).clip(0, 255).astype(np.uint8)
+                                img_for_aug = (
+                                    (img_for_aug * 255).clip(0, 255).astype(np.uint8)
+                                )
                             else:
                                 img_for_aug = img_for_aug.astype(np.uint8)
 
