+++ mkdocs.yml
@@ -159,7 +159,6 @@ nav:
           - examples/lightly_train_self_supervised.ipynb
           - examples/clean_segmentation_results.ipynb
           - examples/cloud_detection.ipynb
-          - examples/data_augmentation.ipynb
     - Workshops:
           - workshops/GeoAI_Workshop_2025.ipynb
           - workshops/AWS_2025.ipynb
