### 2025-11-13 18f5bae
Add data augmentation (#362)
* Add data augmentation

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update geoai/utils.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update geoai/utils.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update links

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Copilot <1757
[…]
