### 2025-11-16 cf2cce3
Fixed model default behaviour bug
