+++ geoai/agents/geo_agents.py
@@ -229,15 +229,12 @@ class GeoAgent(Agent):
             model_args = {}
 
         # --- save a model factory we can call each turn ---
-        if model == "llama3.1":
-            self._model_factory: Callable[[], OllamaModel] = (
-                lambda: create_ollama_model(
-                    host="http://localhost:11434", model_id=model, **model_args
+        if isinstance(model, str) and (":" in model or model.startswith("llama")):
+            # treat ANY "llama..." model id as Ollama
+            self._model_factory = (
+                lambda m=model: create_ollama_model(
+                    host="http://localhost:11434", model_id=m, **model_args
                 )
-            )
-        elif isinstance(model, str):
-            self._model_factory: Callable[[], BedrockModel] = (
-                lambda: create_bedrock_model(model_id=model, **model_args)
             )
         elif isinstance(model, OllamaModel):
             # Extract configuration from existing OllamaModel and create new instances
@@ -266,6 +263,11 @@ class GeoAgent(Agent):
                 lambda mid=model_id, client_args=client_args: create_anthropic_model(
                     model_id=mid, client_args=client_args, **model_args
                 )
+            )
+        elif isinstance(model, str):
+            # Only Bedrock IDs here, not LLaMA
+            self._model_factory = (
+                lambda m=model: create_bedrock_model(model_id=m, **model_args)
             )
         else:
             raise ValueError(f"Invalid model: {model}")
@@ -643,16 +645,14 @@ class STACAgent(Agent):
             model_args = {}
 
         # --- save a model factory we can call each turn ---
-        if model == "llama3.1":
-            self._model_factory: Callable[[], OllamaModel] = (
-                lambda: create_ollama_model(
-                    host="http://localhost:11434", model_id=model, **model_args
+        if isinstance(model, str) and (":" in model or model.startswith("llama")):
+            # treat ANY "llama..." model id as Ollama
+            self._model_factory = (
+                lambda m=model: create_ollama_model(
+                    host="http://localhost:11434", model_id=m, **model_args
                 )
             )
-        elif isinstance(model, str):
-            self._model_factory: Callable[[], BedrockModel] = (
-                lambda: create_bedrock_model(model_id=model, **model_args)
-            )
+
         elif isinstance(model, OllamaModel):
             # Extract configuration from existing OllamaModel and create new instances
             model_id = model.config["model_id"]
@@ -680,6 +680,11 @@ class STACAgent(Agent):
                 lambda mid=model_id, client_args=client_args: create_anthropic_model(
                     model_id=mid, client_args=client_args, **model_args
                 )
+            )
+        elif isinstance(model, str):
+            # Only Bedrock IDs here, not LLaMA
+            self._model_factory = (
+                lambda m=model: create_bedrock_model(model_id=m, **model_args)
             )
         else:
             raise ValueError(f"Invalid model: {model}")
@@ -1284,15 +1289,12 @@ class CatalogAgent(Agent):
             model_args = {}
 
         # --- save a model factory we can call each turn ---
-        if model == "llama3.1":
-            self._model_factory: Callable[[], OllamaModel] = (
-                lambda: create_ollama_model(
-                    host="http://localhost:11434", model_id=model, **model_args
+        if isinstance(model, str) and (":" in model or model.startswith("llama")):
+            # treat ANY "llama..." model id as Ollama
+            self._model_factory = (
+                lambda m=model: create_ollama_model(
+                    host="http://localhost:11434", model_id=m, **model_args
                 )
-            )
-        elif isinstance(model, str):
-            self._model_factory: Callable[[], BedrockModel] = (
-                lambda: create_bedrock_model(model_id=model, **model_args)
             )
[…]
