### 2025-11-16 ec68041
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
