### 2025-11-16 be47c8a
Fixed model default behaviour bug (#364)
* Fixed model default behaviour bug

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
