### 2025-11-16 1a3b507
Add AGU workshop notebook
