+++ mkdocs.yml
@@ -164,6 +164,7 @@ nav:
           - workshops/AWS_2025.ipynb
           - workshops/TNView_2025.ipynb
           - workshops/CANVAS_2025.ipynb
+          - workshops/AGU_2025.ipynb
     - API Reference:
           - change_detection module: change_detection.md
           - classify module: classify.md
