### 2025-11-17 484f83b
Fixed chat UI overflow issue
