### 2025-11-17 374b521
Fixing Chat UI Overflow Issue  (#366)
* Fixed model default behaviour bug

* Fixed chat UI overflow issue

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
