### 2025-11-18 2cacebd
Add super-resolution module
