+++ requirements.txt
@@ -29,3 +29,4 @@ torchgeo
 torchinfo
 tqdm
 transformers>=4.57.1
+opensr-model
