### 2025-11-18 98fa5d4
Add super-resolution module
