### 2025-11-18 422ab03
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
