+++ geoai/tools/__init__.py
@@ -69,6 +69,7 @@ except ImportError:
 
 try:
     from .sr import super_resolution
+
     __all__.extend(["super_resolution"])
 except ImportError:
     pass
