### 2025-11-20 d92ed9d
Make opensr-model optional
