+++ requirements.txt
@@ -29,4 +29,3 @@ torchgeo
 torchinfo
 tqdm
 transformers>=4.57.1
-opensr-model
