+++ requirements.txt
@@ -11,7 +11,6 @@ localtileserver
 mapclassify
 maplibre
 opencv-python-headless
-opensr-model
 overturemaps
 planetary-computer
 psutil
@@ -30,4 +29,3 @@ torchgeo
 torchinfo
 tqdm
 transformers>=4.57.1
-opensr-model
