### 2025-11-20 6ad96d3
Fix super-resolution georeference
