### 2025-11-20 9c2460e
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
