+++ geoai/tools/sr.py
@@ -46,7 +46,7 @@ def super_resolution(
         sampling_steps (int): Number of diffusion sampling steps. Default is 100.
         n_variations (int): Number of samples to compute uncertainty. Default is 25.
         scale (int, optional): Super-resolution scale factor. Default is 4.
-            This adjusts the affine transform to ensure georeference matches 
+            This adjusts the affine transform to ensure georeference matches
             the original image.
 
     Returns:
@@ -84,32 +84,34 @@ def super_resolution(
     # Generate super-resolution tensor
     sr_tensor = model.forward(lr_tensor, sampling_steps=sampling_steps)
     sr_image = sr_tensor.squeeze(0).cpu().numpy().astype(np.float32)
-    save_geotiff(sr_image, profile, output_sr_path,scale)
+    save_geotiff(sr_image, profile, output_sr_path, scale)
     print("Saved super-resolution image to:", output_sr_path)
 
     # Compute uncertainty map
     unc_tensor = model.uncertainty_map(lr_tensor, n_variations=n_variations)
     uncertainty = unc_tensor.squeeze(0).cpu().numpy().astype(np.float32)
-    save_geotiff(uncertainty, profile, output_uncertainty_path,scale)
+    save_geotiff(uncertainty, profile, output_uncertainty_path, scale)
     print("Saved uncertainty map to:", output_uncertainty_path)
 
     return sr_image, uncertainty
 
 
-def save_geotiff(data: np.ndarray, reference_profile: dict, output_path: str, scale: int = 4):
+def save_geotiff(
+    data: np.ndarray, reference_profile: dict, output_path: str, scale: int = 4
+):
     """
-    Save a 2D or 3D NumPy array as a GeoTIFF with super-resolution scaling 
+    Save a 2D or 3D NumPy array as a GeoTIFF with super-resolution scaling
     and corrected georeference.
 
     Args:
         data (np.ndarray): Image array to save. Can be:
             - 2D array (H, W) for a single-band image
             - 3D array (C, H, W) for multi-band images (e.g., RGB+NIR)
-        reference_profile (dict): Rasterio metadata from a reference GeoTIFF. 
+        reference_profile (dict): Rasterio metadata from a reference GeoTIFF.
             Used to preserve CRS, transform, and other metadata.
         output_path (str): Path to save the output GeoTIFF.
         scale (int, optional): Super-resolution scale factor. Default is 4.
-            This adjusts the affine transform to ensure georeference matches 
+            This adjusts the affine transform to ensure georeference matches
             the original image.
 
     Returns:
@@ -123,18 +125,22 @@ def save_geotiff(data: np.ndarray, reference_profile: dict, output_path: str, sc
 
     # Update profile and transform
     profile = reference_profile.copy()
-    old_transform = profile['transform']
+    old_transform = profile["transform"]
     new_transform = Affine(
-        old_transform.a / scale, old_transform.b, old_transform.c,
-        old_transform.d, old_transform.e / scale, old_transform.f
+        old_transform.a / scale,
+        old_transform.b,
+        old_transform.c,
+        old_transform.d,
+        old_transform.e / scale,
+        old_transform.f,
     )
     profile.update(
         dtype=rasterio.float32,
         count=data.shape[0],
         height=data.shape[1],
         width=data.shape[2],
-        compress='lzw',
-        transform=new_transform
+        compress="lzw",
+        transform=new_transform,
     )
 
     with rasterio.open(output_path, "w", **profile) as dst:
