### 2025-11-21 9cb438f
Update geoai/tools/sr.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
