### 2025-11-21 a00d8b9
Update geoai/tools/sr.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
