### 2025-11-21 989356f
Update geoai/tools/__init__.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
