### 2025-11-21 a52012e
Update geoai/__init__.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
