### 2025-11-21 4575062
Update geoai/tools/sr.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
