### 2025-11-21 f4bcc1c
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
