### 2025-11-23 e58b344
"Add new landcover classification module and notebook example"
