+++ geoai/landcover_utils.py
@@ -1,0 +1,372 @@
+"""
+Landcover Classification Utilities - Enhanced Tile Export Module
+
+This module extends the base geoai functionality with specialized utilities
+for discrete landcover classification. It provides enhanced tile generation
+with background filtering capabilities to improve training efficiency.
+
+Key Features:
+- Enhanced tile filtering with configurable feature ratio thresholds
+- Separate statistics tracking for different skip reasons
+- Maintains full compatibility with base geoai workflow
+- Optimized for discrete landcover classification tasks
+
+Date: November 2025
+"""
+
+import os
+import warnings
+from pathlib import Path
+from typing import Any, Dict, List, Optional, Tuple, Union
+
+import geopandas as gpd
+import numpy as np
+import rasterio
+from rasterio import features
+from rasterio.windows import Window
+from tqdm import tqdm
+
+
+def export_landcover_tiles(
+    in_raster: str,
+    out_folder: str,
+    in_class_data: Optional[Union[str, gpd.GeoDataFrame]] = None,
+    tile_size: int = 256,
+    stride: int = 128,
+    class_value_field: str = "class",
+    buffer_radius: float = 0,
+    max_tiles: Optional[int] = None,
+    quiet: bool = False,
+    all_touched: bool = True,
+    create_overview: bool = False,
+    skip_empty_tiles: bool = False,
+    min_feature_ratio: Union[bool, float] = False,
+    metadata_format: str = "PASCAL_VOC",
+) -> Dict[str, Any]:
+    """
+    Export GeoTIFF tiles optimized for landcover classification training.
+    
+    This function extends the base export_geotiff_tiles with enhanced filtering
+    capabilities specifically designed for discrete landcover classification.
+    It can filter out tiles dominated by background pixels to improve training
+    data quality and reduce dataset size.
+    
+    Args:
+        in_raster: Path to input raster (image to tile)
+        out_folder: Output directory for tiles
+        in_class_data: Path to vector mask or GeoDataFrame (optional for image-only export)
+        tile_size: Size of output tiles in pixels (default: 256)
+        stride: Stride for sliding window (default: 128)
+        class_value_field: Field name containing class values (default: "class")
+        buffer_radius: Buffer radius around features in pixels (default: 0)
+        max_tiles: Maximum number of tiles to export (default: None)
+        quiet: Suppress progress output (default: False)
+        all_touched: Include pixels touched by geometry (default: True)
+        create_overview: Create overview image showing tile locations (default: False)
+        skip_empty_tiles: Skip tiles with no features (default: False)
+        min_feature_ratio: Minimum ratio of non-background pixels required to keep tile
+            - False: Disable ratio filtering (default)
+            - 0.0-1.0: Minimum ratio threshold (e.g., 0.1 = 10% features required)
+        metadata_format: Annotation format ("PASCAL_VOC" or "YOLO")
+    
+    Returns:
+        Dictionary containing:
+            - tiles_exported: Number of tiles successfully exported
+            - tiles_skipped_empty: Number of completely empty tiles skipped
+            - tiles_skipped_ratio: Number of tiles filtered by min_feature_ratio
+            - output_dirs: Dictionary with paths to images and labels directories
+    
+    Examples:
+        # Original behavior (no filtering)
+        export_landcover_tiles(
+            "input.tif", 
+            "output", 
+            "mask.shp",
+            skip_empty_tiles=True
+        )
+        
+        # Light filtering (keep tiles with ≥5% features)
+        export_landcover_tiles(
+            "input.tif",
+            "output", 
+            "mask.shp",
+            skip_empty_tiles=True,
+            min_feature_ratio=0.05
+        )
+        
[…]
