+++ geoai/train.py
@@ -1,80 +1,3 @@
-"""
-GEOAI TRAIN.PY - MODIFICATIONS
-===========================================================
-
-This file has been extensively modified for landcover classification with the following enhancements:
-
-MAJOR MODIFICATIONS SUMMARY:
-============================
-
-1. ENHANCED IoU EVALUATION METRICS (Lines 2591-2752):
-   - Extended iou_coefficient function with multiple calculation modes
-   - Standard IoU: Unweighted average across all classes (original behavior preserved)
-   - Per-class frequency-weighted IoU: Classes weighted by pixel frequency
-   - Boundary-weighted IoU (wIoU): Distance-based weighting focusing on class boundaries
-   
-   BOUNDARY-WEIGHTED IoU (wIoU) IMPLEMENTATION:
-   - Source: "Weighted Intersection over Union (wIoU): A New Evaluation Metric for 
-     Image Segmentation" by Yeong-Jun Cho (arXiv:2107.09858)
-   - Algorithm: Exponentially weights pixels based on distance from class boundaries
-     * Pixels near boundaries receive higher weights (critical for accuracy)
-     * Pixels in region centers receive lower weights (less critical)
-     * Formula: weight_map = exp(-alpha * distance_from_boundary)
-   - Implementation: generate_boundary_weight_map() (Lines 2506-2589)
-     * Uses cv2.distanceTransform for accurate distance calculation
-     * Normalizes and accumulates distances across all classes
-     * Applies exponential decay with configurable alpha parameter
-   - Use case: Boundary-critical applications (roads, building footprints, property lines)
-   - Configuration: Set validation_iou_mode="boundary_weighted" and boundary_alpha (0.01-100)
-   
-   IGNORE_INDEX SUPPORT:
-   - Added proper handling of ignore_index in all IoU calculation modes
-   - Pixels with ignore_index value are excluded from metric calculations
-   - Ensures accurate evaluation when training with unlabeled/masked regions
-   - Applies to both standard IoU and weighted variants
-   
-   BACKWARD COMPATIBILITY:
-   - New features activated only when optional parameters provided
-
-2. ADVANCED LOSS FUNCTIONS (Lines 97-176):
-   - FocalLoss class: Custom PyTorch implementation for class imbalance
-   - Alpha/gamma parameters for fine-tuning hard example focus
-    - Reasoning: Standard CrossEntropy inadequate for highly imbalanced landcover data
-   - Functionality: Mathematically correct focal loss with class weights support
-   - Usage: loss_function='focal', focal_alpha=1.0, focal_gamma=2.0
-
-2. CLASS WEIGHT COMPUTATION (Lines 216-315):
-   - compute_class_weights function: Enhanced inverse frequency calculation  
-   - Custom multiplier support: Manual adjustment of specific class weights
-   - Maximum weight capping: Prevent extreme weights (configurable, default 50.0)
-    - Reasoning: Landcover data has severe imbalance (roads <1% vs forest >40%)
-   - Functionality: Auto-computation + manual fine-tuning + overflow protection
-   - Usage: custom_multipliers={1: 0.5}, max_class_weight=50.0
-
-3. ENHANCED TRAINING FUNCTION (Lines 2565-3216):
-   - train_segmentation_model: Extended parameter support for advanced loss functions
-   - NEW PARAMETERS:
-     * loss_function: 'crossentropy', 'focal', or 'dice'
-     * ignore_index: Pixel value to ignore (0, False, or custom int)
-     * use_class_weights: Enable automatic class balancing
-     * custom_multipliers: Manual class weight adjustments {class_id: multiplier}
-     * focal_alpha/focal_gamma: Focal loss hyperparameters
[…]
