+++ geoai/utils.py
@@ -1,79 +1,4 @@
-"""
-GEOAI UTILS.PY - ENHANCED VERSION WITH CUSTOM MODIFICATIONS
-===========================================================
-
-This file has been enhanced with improved tile generation capabilities:
-
-MAJOR MODIFICATIONS SUMMARY:
-============================
-
-1. ENHANCED TILE FILTERING (Lines 2675-3410):
-   - export_geotiff_tiles function: Added min_feature_ratio parameter (Line 2688)
-   - Background filtering: Skip tiles dominated by background pixels (class 0)
-   - Statistical tracking: Separate counters for different skip reasons
-   - NEW PARAMETER: min_feature_ratio=False (default, maintains compatibility)
-   - Reasoning: Training on 95% background tiles creates poor class balance
-   - Functionality: Configurable threshold (0.0-1.0) for minimum feature content
-   - Implementation: Lines 3044-3062 (validation and filtering logic)
-
-2. PARAMETER VALIDATION (Lines 3046-3050):
-   - Added robust validation for min_feature_ratio parameter
-   - Type checking: isinstance(min_feature_ratio, (int, float))
-   - Range validation: 0.0 <= min_feature_ratio <= 1.0
-   - Graceful fallback with warning messages for invalid inputs
-   - Reasoning: Prevent runtime errors from invalid parameter values
-   - Functionality: Input sanitization with descriptive error messages
-
-3. STATISTICAL ENHANCEMENT (Lines 2720-2730, 3057-3060):
-   - Enhanced statistics tracking with separate counters:
-     * tiles_skipped_empty: Completely empty tiles (original)
-     * tiles_skipped_ratio: Tiles filtered by min_feature_ratio (new)
-   - Improved summary reporting with detailed skip reasons
-   - Reasoning: Better understanding of filtering impact on dataset
-   - Functionality: Separate tracking for different filtering mechanisms
-
-KEY ENHANCEMENTS:
-=================
-
-export_geotiff_tiles() - Enhanced Tile Generation:
-- NEW PARAMETER: min_feature_ratio (default: False)
-  - False: Disable ratio filtering (original behavior)
-  - 0.0-1.0: Minimum ratio of non-background pixels required
-  - Example: 0.10 = keep only tiles with ≥10% feature pixels
-
-- IMPROVED STATISTICS: Enhanced tracking and reporting
-  - tiles_skipped_ratio: Count of tiles filtered by feature ratio
-  - tiles_skipped_empty: Count of completely empty tiles  
-  - Separate counters for different skip reasons
-
-- BETTER OUTPUT: Reduced verbosity while maintaining functionality
-  - Progress bar updates without individual tile messages
-  - Summary statistics at completion
-  - Error logging preserved for troubleshooting
-
-USAGE EXAMPLES:
-===============
-# Original behavior (no filtering)
-export_geotiff_tiles(raster, folder, mask, skip_empty_tiles=True)
-
-# Light filtering (keep tiles with ≥5% features)
-export_geotiff_tiles(raster, folder, mask, skip_empty_tiles=True, min_feature_ratio=0.05)
-
-# Moderate filtering (keep tiles with ≥15% features) 
-export_geotiff_tiles(raster, folder, mask, skip_empty_tiles=True, min_feature_ratio=0.15)
-
-# Aggressive filtering (keep tiles with ≥25% features)
-export_geotiff_tiles(raster, folder, mask, skip_empty_tiles=True, min_feature_ratio=0.25)
-
-BENEFITS FOR LANDCOVER CLASSIFICATION:
-====================================
-- Improved training efficiency: Focus on tiles with meaningful content
-- Better class balance: Reduce background dominance in training data
-- Faster convergence: Less noise from nearly-empty tiles
-- Reduced dataset size: Eliminate low-information tiles
[…]
