### 2025-11-23 10d581f
Add landcover classification modules and notebook example
