+++ .github/workflows/ubuntu.yml
@@ -17,7 +17,7 @@ jobs:
                 python-version: ["3.10", "3.11", "3.12"]
 
         steps:
-            - uses: actions/checkout@v5
+            - uses: actions/checkout@v6
 
             # - name: Install GDAL system dependencies
             #   run: sudo apt-get update && sudo apt-get install -y gdal-bin libgdal-dev
