+++ .github/workflows/docs.yml
@@ -12,7 +12,7 @@ jobs:
         env:
             TITILER_ENDPOINT: ${{ secrets.TITILER_ENDPOINT }}
         steps:
-            - uses: actions/checkout@v5
+            - uses: actions/checkout@v6
               with:
                   fetch-depth: 0
 
