+++ .github/workflows/docs-build.yml
@@ -14,7 +14,7 @@ jobs:
             TITILER_ENDPOINT: ${{ secrets.TITILER_ENDPOINT }}
 
         steps:
-            - uses: actions/checkout@v5
+            - uses: actions/checkout@v6
               with:
                   fetch-depth: 0
 
