### 2025-11-24 a3fb862
Add Super-Resolution Module for GeoAI (#368)
* Add super-resolution module

* Add super-resolution module

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Make opensr-model optional

* Update dependencies

* Fix super-resolution georeference

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update geoai/tools/sr.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update geoai/tools/sr.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com
[…]
