### 2025-11-25 c2c55f4
Remove landcover files from root directory - they belong in geoai/ subdirectory
