+++ geoai/landcover_train.py
@@ -1,0 +1,685 @@
+"""
+Landcover Classification Training Module
+
+This module extends the base geoai training functionality with specialized
+components for discrete landcover classification, including:
+- Enhanced loss functions with boundary weighting
+- Per-class frequency weighting for imbalanced datasets
+- Configurable ignore_index handling
+- Additional validation metrics
+
+Key Features:
+- Maintains full compatibility with base geoai workflow
+- Adds optional advanced loss computation modes
+- Provides flexible ignore_index configuration
+- Optimized for multi-class landcover segmentation
+
+Author: ValHab Project
+Date: November 2025
+"""
+
+from typing import Any, Callable, Dict, List, Optional, Tuple, Union
+
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+
+
+class FocalLoss(nn.Module):
+    """
+    Focal Loss for addressing class imbalance in segmentation.
+
+    Reference: Lin, T. Y., Goyal, P., Girshick, R., He, K., & Dollár, P. (2017).
+    Focal loss for dense object detection. ICCV.
+
+    Args:
+        alpha: Weighting factor in range (0,1) to balance positive/negative examples
+        gamma: Exponent of the modulating factor (1 - p_t)^gamma
+        ignore_index: Specifies a target value that is ignored
+        reduction: Specifies the reduction to apply to the output
+        weight: Manual rescaling weight given to each class
+    """
+
+    def __init__(
+        self, alpha=1.0, gamma=2.0, ignore_index=-100, reduction="mean", weight=None
+    ):
+        super(FocalLoss, self).__init__()
+        self.alpha = alpha
+        self.gamma = gamma
+        self.ignore_index = ignore_index
+        self.reduction = reduction
+        self.weight = weight
+
+    def forward(self, inputs, targets):
+        """
+        Forward pass of focal loss.
+
+        Args:
+            inputs: Predictions (N, C, H, W) where C = number of classes
+            targets: Ground truth (N, H, W) with class indices
+
+        Returns:
+            Loss value
+        """
+        # Get class probabilities
+        ce_loss = F.cross_entropy(
+            inputs,
+            targets,
+            weight=self.weight,
+            ignore_index=self.ignore_index,
+            reduction="none",
+        )
+
+        # Get probability of true class
+        p_t = torch.exp(-ce_loss)
+
+        # Calculate focal loss
+        focal_loss = self.alpha * (1 - p_t) ** self.gamma * ce_loss
+
+        # Apply reduction
+        if self.reduction == "mean":
+            return focal_loss.mean()
+        elif self.reduction == "sum":
+            return focal_loss.sum()
+        else:
+            return focal_loss
+
+
+class LandcoverCrossEntropyLoss(nn.Module):
+    """
+    Enhanced CrossEntropyLoss with optional ignore_index and class weights.
+
+    This extends the standard CrossEntropyLoss with more flexible ignore_index
+    handling, specifically designed for landcover classification tasks.
+
+    Args:
+        weight: Manual rescaling weight given to each class
+        ignore_index: Specifies a target value that is ignored (default: None)
+            - None: No values ignored (standard behavior)
+            - int: Specific class index to ignore (e.g., 0 for background)
+        reduction: Specifies the reduction to apply ('mean', 'sum', 'none')
+    """
+
+    def __init__(
+        self,
+        weight: Optional[torch.Tensor] = None,
+        ignore_index: Optional[int] = None,
+        reduction: str = "mean",
+    ):
+        super().__init__()
+        self.weight = weight
+        self.ignore_index = ignore_index if ignore_index is not None else -100
+        self.reduction = reduction
+
[…]
