### 2025-12-06 3a7bf29
added gemini model support and handle initialization error gracefully
