### 2025-12-06 1a3f8f9
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
