### 2025-12-06 5a99bd6
feat: #377 support for Gemini model integration into geoagents (#378)
* added gemini model support and handle initialization error gracefully

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fixed GEMINI_API_KEY reference in example to GOOGLE_API_KEY

* Fixed docs codespell errors

* Update geoai/agents/geo_agents.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update docs/examples/AI_agents.ipynb

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update docs/examples/AI_agents.ipynb

C
[…]
