+++ geoai/agents/__init__.py
@@ -6,6 +6,7 @@ from .geo_agents import (
     create_ollama_model,
     create_anthropic_model,
     create_openai_model,
+    create_gemini_model,
     create_bedrock_model,
 )
 from .map_tools import MapTools
