### 2025-12-10 627915c
Add support for moondream
