### 2025-12-10 b5945db
Add support for moondream (#382)
