+++ geoai/moondream.py
@@ -1,0 +1,943 @@
+"""Moondream Vision Language Model module for GeoAI.
+
+This module provides an interface for using Moondream vision language models
+(moondream2 and moondream3-preview) with geospatial imagery, supporting
+GeoTIFF input and georeferenced output.
+
+Supported models:
+- moondream2: https://huggingface.co/vikhyatk/moondream2
+- moondream3-preview: https://huggingface.co/moondream/moondream3-preview
+"""
+
+import os
+from typing import Any, Dict, List, Optional, Tuple, Union
+
+import geopandas as gpd
+import numpy as np
+import rasterio
+import torch
+from PIL import Image
+from shapely.geometry import Point, box
+from transformers.utils import logging as hf_logging
+
+from .utils import get_device
+
+
+hf_logging.set_verbosity_error()  # silence HF load reports
+
+
+class MoondreamGeo:
+    """Moondream Vision Language Model processor with GeoTIFF support.
+
+    This class provides an interface for using Moondream models for
+    geospatial image analysis, including captioning, visual querying,
+    object detection, and pointing.
+
+    Attributes:
+        model: The loaded Moondream model.
+        model_name: Name of the model being used.
+        device: Torch device for inference.
+        model_version: Either "moondream2" or "moondream3".
+    """
+
+    def __init__(
+        self,
+        model_name: str = "vikhyatk/moondream2",
+        revision: Optional[str] = None,
+        device: Optional[str] = None,
+        compile_model: bool = False,
+        **kwargs: Any,
+    ) -> None:
+        """Initialize the Moondream processor.
+
+        Args:
+            model_name: HuggingFace model name. Options:
+                - "vikhyatk/moondream2" (default)
+                - "moondream/moondream3-preview"
+            revision: Model revision/checkpoint date. For moondream2, recommended
+                to use a specific date like "2025-06-21" for reproducibility.
+            device: Device for inference ("cuda", "mps", or "cpu").
+                If None, automatically selects the best available device.
+            compile_model: Whether to compile the model (recommended for
+                moondream3-preview for faster inference).
+            **kwargs: Additional arguments passed to from_pretrained.
+
+        Raises:
+            ImportError: If transformers is not installed.
+            RuntimeError: If model loading fails.
+        """
+        self.model_name = model_name
+        self.device = device or get_device()
+        self._source_path: Optional[str] = None
+        self._metadata: Optional[Dict] = None
+
+        # Determine model version
+        if "moondream3" in model_name.lower():
+            self.model_version = "moondream3"
+        else:
+            self.model_version = "moondream2"
+
+        # Load the model
+        self.model = self._load_model(revision, compile_model, **kwargs)
+
+    def _load_model(
+        self,
+        revision: Optional[str],
+        compile_model: bool,
+        **kwargs: Any,
+    ) -> Any:
+        """Load the Moondream model.
+
+        Args:
+            revision: Model revision/checkpoint.
+            compile_model: Whether to compile the model.
+            **kwargs: Additional arguments for from_pretrained.
+
+        Returns:
+            Loaded model instance.
+
+        Raises:
+            RuntimeError: If model loading fails.
+        """
+        try:
+            from transformers import AutoModelForCausalLM
+
+            # Default kwargs
+            load_kwargs = {
+                "trust_remote_code": True,
+            }
+
+            # Try to use device_map with accelerate, fall back to manual device placement
+            use_device_map = False
+            try:
+                import accelerate  # noqa: F401
+
+                # Build device map
+                if isinstance(self.device, str):
+                    device_map = {"": self.device}
+                else:
+                    device_map = {"": str(self.device)}
+                load_kwargs["device_map"] = device_map
+                use_device_map = True
[…]
