+++ mkdocs.yml
@@ -159,6 +159,7 @@ nav:
           - examples/lightly_train_self_supervised.ipynb
           - examples/clean_segmentation_results.ipynb
           - examples/cloud_detection.ipynb
+          - examples/moondream.ipynb
     - Workshops:
           - workshops/GeoAI_Workshop_2025.ipynb
           - workshops/AWS_2025.ipynb
@@ -177,6 +178,7 @@ nav:
           - hf module: hf.md
           - map_tools module: map_tools.md
           - map_widgets module: map_widgets.md
+          - moondream module: moondream.md
           - sam module: sam.md
           - segmentation module: segmentation.md
           - timm_train module: timm_train.md
