### 2025-12-10 af60720
Fix view_image bug
