### 2025-12-10 ebf4296
Add moondream interactive GUI
