### 2025-12-10 bce27ad
Update notebook
