### 2025-12-10 9758ba3
Update geoai/map_widgets.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
