+++ geoai/map_widgets.py
@@ -4,7 +4,7 @@ import os
 import string
 import random
 import tempfile
-from typing import Any, Dict, Optional
+from typing import Any, Optional
 
 import ipywidgets as widgets
 
