### 2025-12-10 d23d274
Update geoai/map_widgets.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
