### 2025-12-10 142d3f1
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
