### 2025-12-10 5166c61
Update geoai/moondream.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
