### 2025-12-10 9587c93
Add moondream interactive GUI (#384)
* Add moondream interactive GUI

* Update notebook

* Update geoai/map_widgets.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update geoai/map_widgets.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update geoai/map_widgets.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update geoai/map_widgets.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* 
[…]
