+++ geoai/map_widgets.py
@@ -1,8 +1,27 @@
 """Interactive widget for GeoAI."""
 
+import os
+import string
+import random
+import tempfile
+from typing import Any, Optional
+
 import ipywidgets as widgets
 
 from .utils import dict_to_image, dict_to_rioxarray
+
+
+def random_string(string_length: int = 6) -> str:
+    """Generate a random string of fixed length.
+
+    Args:
+        string_length: The length of the random string. Defaults to 6.
+
+    Returns:
+        A random string of the specified length.
+    """
+    letters = string.ascii_lowercase
+    return "".join(random.choice(letters) for _ in range(string_length))
 
 
 class DINOv3GUI(widgets.VBox):
@@ -172,3 +191,540 @@ class DINOv3GUI(widgets.VBox):
 
             host_map.on_interaction(handle_map_interaction)
             host_map.default_style = {"cursor": "crosshair"}
+
+
+def moondream_gui(
+    moondream,
+    basemap: str = "SATELLITE",
+    out_dir: Optional[str] = None,
+    opacity: float = 0.5,
+    **kwargs: Any,
+):
+    """Display an interactive GUI for using Moondream with leafmap.
+
+    This function creates an interactive map interface for using Moondream
+    vision language model capabilities including:
+    - Image captioning (short, normal, long) with streaming output
+    - Visual question answering (query) with streaming output
+    - Object detection with bounding boxes displayed on map
+    - Point detection for locating objects with markers on map
+
+    Args:
+        moondream (MoondreamGeo): The MoondreamGeo object with a loaded image.
+            Must have called load_image() or load_geotiff() first.
+        basemap (str, optional): The basemap to use. Defaults to "SATELLITE".
+        out_dir (str, optional): The output directory for saving results.
+            Defaults to None (uses temp directory).
+        opacity (float, optional): The opacity of overlay layers. Defaults to 0.5.
+        **kwargs: Additional keyword arguments passed to leafmap.Map().
+
+    Returns:
+        leafmap.Map: The interactive map with the Moondream GUI.
+
+    Example:
+        >>> from geoai import MoondreamGeo, moondream_gui
+        >>> moondream = MoondreamGeo()
+        >>> moondream.load_image("image.tif")
+        >>> m = moondream_gui(moondream)
+        >>> m
+    """
+    try:
+        import ipyevents
+        import ipyleaflet
+        import leafmap
+        from ipyfilechooser import FileChooser
+    except ImportError:
+        raise ImportError(
+            "The moondream_gui function requires additional packages. "
+            "Please install them with: pip install leafmap ipyevents ipyfilechooser"
+        )
+
+    if out_dir is None:
+        out_dir = tempfile.gettempdir()
+
+    # Create the map
+    m = leafmap.Map(**kwargs)
+    m.default_style = {"cursor": "crosshair"}
+    if basemap is not None:
+        m.add_basemap(basemap, show=False)
+
+    # Try to add the image layer if source is available
+    if moondream._source_path is not None:
+        try:
+            m.add_raster(moondream._source_path, layer_name="Image")
+        except Exception:
+            pass
+
+    # Initialize marker storage for detection results
+    m.detection_markers = []
+    m.point_markers = []
+
+    # Removed unused LayerGroups for detections and points.
+    m.last_result_as_gdf = None
+
+    # Widget styling
+    widget_width = "300px"
+    button_width = "90px"
+    padding = "0px 4px 0px 4px"
+    style = {"description_width": "initial"}
+
+    # Create toolbar buttons
+    toolbar_button = widgets.ToggleButton(
+        value=True,
+        tooltip="Toolbar",
+        icon="gear",
[…]
