### 2025-12-10 bdf51f6
Bump version: 0.18.2 → 0.19.0
