### 2025-12-10 0095b8b
Add QGIS plugin
