+++ qgis_plugin/geoai_plugin/dialogs/moondream.py
@@ -1,0 +1,622 @@
+"""
+Moondream Dock Widget for GeoAI Plugin
+
+This dock widget provides an interactive interface for using the Moondream
+vision-language model for geospatial image analysis in QGIS.
+"""
+
+import os
+import tempfile
+from typing import Optional
+
+from qgis.PyQt.QtCore import Qt, QThread, pyqtSignal
+from qgis.PyQt.QtWidgets import (
+    QDockWidget,
+    QWidget,
+    QVBoxLayout,
+    QHBoxLayout,
+    QLabel,
+    QLineEdit,
+    QComboBox,
+    QPushButton,
+    QTextEdit,
+    QFileDialog,
+    QGroupBox,
+    QFormLayout,
+    QProgressBar,
+    QMessageBox,
+    QCheckBox,
+    QColorDialog,
+    QScrollArea,
+    QSizePolicy,
+)
+from qgis.PyQt.QtGui import QColor
+
+from qgis.core import (
+    QgsProject,
+    QgsVectorLayer,
+    QgsRasterLayer,
+)
+from qgis.gui import QgsMapLayerComboBox
+from qgis.core import QgsMapLayerProxyModel
+
+
+class MoondreamWorker(QThread):
+    """Worker thread for running Moondream operations."""
+
+    finished = pyqtSignal(dict)
+    error = pyqtSignal(str)
+    progress = pyqtSignal(str)
+
+    def __init__(
+        self,
+        moondream,
+        mode: str,
+        source_path: str,
+        prompt: str = "",
+        caption_length: str = "normal",
+    ):
+        super().__init__()
+        self.moondream = moondream
+        self.mode = mode
+        self.source_path = source_path
+        self.prompt = prompt
+        self.caption_length = caption_length
+
+    def run(self):
+        """Execute the Moondream operation."""
+        try:
+            if self.mode == "Caption":
+                self.progress.emit(f"Generating {self.caption_length} caption...")
+                result = self.moondream.caption(
+                    self.source_path,
+                    length=self.caption_length,
+                    stream=False,
+                )
+                self.finished.emit({"type": "caption", "result": result})
+
+            elif self.mode == "Query":
+                self.progress.emit(f"Processing query: {self.prompt}")
+                result = self.moondream.query(
+                    self.prompt,
+                    source=self.source_path,
+                    stream=False,
+                )
+                self.finished.emit({"type": "query", "result": result, "question": self.prompt})
+
+            elif self.mode == "Detect":
+                self.progress.emit(f"Detecting: {self.prompt}")
+                result = self.moondream.detect(
+                    self.source_path,
+                    self.prompt,
+                )
+                self.finished.emit({"type": "detect", "result": result, "object_type": self.prompt})
+
+            elif self.mode == "Point":
+                self.progress.emit(f"Locating: {self.prompt}")
+                result = self.moondream.point(
+                    self.source_path,
+                    self.prompt,
+                )
+                self.finished.emit({"type": "point", "result": result, "description": self.prompt})
+
+        except Exception as e:
+            self.error.emit(str(e))
+
+
+class MoondreamDockWidget(QDockWidget):
+    """Dockable widget for Moondream vision-language model interaction."""
+
+    def __init__(self, iface, parent=None):
+        """Initialize the Moondream dock widget.
+
+        Args:
+            iface: QGIS interface instance.
+            parent: Parent widget.
+        """
+        super().__init__("GeoAI - Moondream", parent)
+        self.iface = iface
+        self.moondream = None
+        self.current_image_path = None
+        self.last_result = None
+        self.worker = None
+
[…]
