+++ qgis_plugin/geoai_plugin/geoai_plugin.py
@@ -1,0 +1,263 @@
+"""
+GeoAI Plugin for QGIS - Main Plugin Class
+
+This module contains the main plugin class that manages the QGIS interface
+integration, menu items, and toolbar buttons.
+"""
+
+import os
+from pathlib import Path
+
+from qgis.PyQt.QtCore import Qt
+from qgis.PyQt.QtGui import QIcon
+from qgis.PyQt.QtWidgets import QAction, QMenu, QToolBar, QMessageBox
+
+from qgis.core import Qgis
+
+
+class GeoAIPlugin:
+    """GeoAI Plugin implementation class for QGIS."""
+
+    def __init__(self, iface):
+        """Constructor.
+
+        Args:
+            iface: An interface instance that provides the hook to QGIS.
+        """
+        self.iface = iface
+        self.plugin_dir = os.path.dirname(__file__)
+        self.actions = []
+        self.menu = None
+        self.toolbar = None
+
+        # Dock widgets (lazy loaded)
+        self._moondream_dock = None
+        self._segmentation_dock = None
+
+    def add_action(
+        self,
+        icon_path,
+        text,
+        callback,
+        enabled_flag=True,
+        add_to_menu=True,
+        add_to_toolbar=True,
+        status_tip=None,
+        checkable=False,
+        parent=None,
+    ):
+        """Add a toolbar icon to the toolbar.
+
+        Args:
+            icon_path: Path to the icon for this action.
+            text: Text that appears in the menu for this action.
+            callback: Function to be called when the action is triggered.
+            enabled_flag: A flag indicating if the action should be enabled.
+            add_to_menu: Flag indicating whether action should be added to menu.
+            add_to_toolbar: Flag indicating whether action should be added to toolbar.
+            status_tip: Optional text to show in status bar when mouse hovers over action.
+            checkable: Whether the action is checkable (toggle).
+            parent: Parent widget for the new action.
+
+        Returns:
+            The action that was created.
+        """
+        icon = QIcon(icon_path)
+        action = QAction(icon, text, parent)
+        action.triggered.connect(callback)
+        action.setEnabled(enabled_flag)
+        action.setCheckable(checkable)
+
+        if status_tip is not None:
+            action.setStatusTip(status_tip)
+
+        if add_to_toolbar:
+            self.toolbar.addAction(action)
+
+        if add_to_menu:
+            self.menu.addAction(action)
+
+        self.actions.append(action)
+
+        return action
+
+    def initGui(self):
+        """Create the menu entries and toolbar icons inside the QGIS GUI."""
+        # Create menu
+        self.menu = QMenu("&GeoAI")
+        self.iface.mainWindow().menuBar().addMenu(self.menu)
+
+        # Create toolbar
+        self.toolbar = QToolBar("GeoAI Toolbar")
+        self.toolbar.setObjectName("GeoAIToolbar")
+        self.iface.addToolBar(self.toolbar)
+
+        # Get icon paths
+        icon_base = os.path.join(self.plugin_dir, "icons")
+
+        # Icon paths with SVG support
+        moondream_icon = os.path.join(icon_base, "moondream.svg")
+        if not os.path.exists(moondream_icon):
+            moondream_icon = ":/images/themes/default/mIconRaster.svg"
+
+        segment_icon = os.path.join(icon_base, "segment.svg")
+        if not os.path.exists(segment_icon):
+            segment_icon = ":/images/themes/default/mActionAddRasterLayer.svg"
+
+        about_icon = os.path.join(icon_base, "about.svg")
+        if not os.path.exists(about_icon):
+            about_icon = ":/images/themes/default/mActionHelpContents.svg"
+
+        # Add Moondream action (checkable for dock toggle)
+        self.moondream_action = self.add_action(
+            moondream_icon,
[…]
