+++ qgis_plugin/geoai_plugin/dialogs/segmentation.py
@@ -88,7 +88,9 @@ class TrainingWorker(QThread):
                 input_format=self.input_format,
                 architecture=self.architecture,
                 encoder_name=self.encoder_name,
-                encoder_weights=self.encoder_weights if self.encoder_weights != "None" else None,
+                encoder_weights=(
+                    self.encoder_weights if self.encoder_weights != "None" else None
+                ),
                 num_channels=self.num_channels,
                 num_classes=self.num_classes,
                 batch_size=self.batch_size,
@@ -103,6 +105,7 @@ class TrainingWorker(QThread):
 
         except Exception as e:
             import traceback
+
             self.error.emit(f"{str(e)}\n\n{traceback.format_exc()}")
 
 
@@ -150,6 +153,7 @@ class TileExportWorker(QThread):
 
         except Exception as e:
             import traceback
+
             self.error.emit(f"{str(e)}\n\n{traceback.format_exc()}")
 
 
@@ -220,6 +224,7 @@ class InferenceWorker(QThread):
 
         except Exception as e:
             import traceback
+
             self.error.emit(f"{str(e)}\n\n{traceback.format_exc()}")
 
 
@@ -267,6 +272,7 @@ class VectorizeWorker(QThread):
 
         except Exception as e:
             import traceback
+
             self.error.emit(f"{str(e)}\n\n{traceback.format_exc()}")
 
 
@@ -499,19 +505,36 @@ class SegmentationDockWidget(QDockWidget):
 
         self.architecture_combo = QComboBox()
         self.architecture_combo.setStyleSheet(self.combo_style)
-        self.architecture_combo.addItems([
-            "unet", "unetplusplus", "deeplabv3", "deeplabv3plus",
-            "fpn", "pspnet", "linknet", "manet", "segformer",
-        ])
+        self.architecture_combo.addItems(
+            [
+                "unet",
+                "unetplusplus",
+                "deeplabv3",
+                "deeplabv3plus",
+                "fpn",
+                "pspnet",
+                "linknet",
+                "manet",
+                "segformer",
+            ]
+        )
         arch_layout.addRow("Architecture:", self.architecture_combo)
 
         self.encoder_combo = QComboBox()
         self.encoder_combo.setStyleSheet(self.combo_style)
-        self.encoder_combo.addItems([
-            "resnet34", "resnet50", "resnet101", "resnet152",
-            "efficientnet-b0", "efficientnet-b1", "efficientnet-b2",
-            "mobilenet_v2", "vgg16",
-        ])
+        self.encoder_combo.addItems(
+            [
+                "resnet34",
+                "resnet50",
+                "resnet101",
+                "resnet152",
+                "efficientnet-b0",
+                "efficientnet-b1",
+                "efficientnet-b2",
+                "mobilenet_v2",
+                "vgg16",
+            ]
+        )
         arch_layout.addRow("Encoder:", self.encoder_combo)
 
         self.encoder_weights_combo = QComboBox()
@@ -645,19 +668,36 @@ class SegmentationDockWidget(QDockWidget):
 
         self.inf_architecture_combo = QComboBox()
         self.inf_architecture_combo.setStyleSheet(self.combo_style)
-        self.inf_architecture_combo.addItems([
-            "unet", "unetplusplus", "deeplabv3", "deeplabv3plus",
-            "fpn", "pspnet", "linknet", "manet", "segformer",
-        ])
+        self.inf_architecture_combo.addItems(
+            [
+                "unet",
+                "unetplusplus",
+                "deeplabv3",
+                "deeplabv3plus",
+                "fpn",
+                "pspnet",
+                "linknet",
+                "manet",
[…]
