### 2025-12-10 d18dbf8
Update qgis_plugin/README.md
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
