### 2025-12-10 4474c93
Update qgis_plugin/README.md
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
