### 2025-12-10 1b38b31
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
