+++ qgis_plugin/geoai_plugin/geoai_plugin.py
@@ -12,6 +12,7 @@ from qgis.PyQt.QtCore import Qt
 from qgis.PyQt.QtGui import QIcon
 from qgis.PyQt.QtWidgets import QAction, QMenu, QToolBar, QMessageBox
 
+
 class GeoAIPlugin:
     """GeoAI Plugin implementation class for QGIS."""
 
