### 2025-12-10 663b9ac
Update qgis_plugin/install.sh
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
