+++ qgis_plugin/install.sh
@@ -36,7 +36,7 @@ echo ""
 echo "To use the plugin:"
 echo "1. Restart QGIS"
 echo "2. Go to Plugins -> Manage and Install Plugins..."
-echo "3. Enable 'GeoAI Plugin'"
+echo "3. Enable 'GeoAI'"
 echo ""
 echo "Make sure you have the required Python packages installed:"
 echo "  pip install geoai-py torch torchvision"
