### 2025-12-10 a676c0b
Update qgis_plugin/geoai_plugin/geoai_plugin.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
