### 2025-12-10 90d66ed
Update qgis_plugin/geoai_plugin/dialogs/moondream.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
