### 2025-12-11 8a2e97e
Improve GUI and add clean up button
