+++ qgis_plugin/geoai_plugin/dialogs/segmentation.py
@@ -36,12 +36,40 @@ from qgis.gui import QgsMapLayerComboBox
 from qgis.core import QgsMapLayerProxyModel
 
 
+class OutputCapture:
+    """Capture stdout and emit lines to a callback in real-time."""
+
+    def __init__(self, callback, original_stdout):
+        self.callback = callback
+        self.original_stdout = original_stdout
+        self.buffer = ""
+
+    def write(self, text):
+        # Also write to original stdout
+        if self.original_stdout:
+            self.original_stdout.write(text)
+
+        self.buffer += text
+        while "\n" in self.buffer:
+            line, self.buffer = self.buffer.split("\n", 1)
+            if line.strip():
+                self.callback(line)
+
+    def flush(self):
+        if self.original_stdout:
+            self.original_stdout.flush()
+        if self.buffer.strip():
+            self.callback(self.buffer)
+            self.buffer = ""
+
+
 class TrainingWorker(QThread):
     """Worker thread for training segmentation models."""
 
     finished = pyqtSignal(str)
     error = pyqtSignal(str)
     progress = pyqtSignal(str)
+    epoch_progress = pyqtSignal(int, int, str)  # current_epoch, total_epochs, metrics
 
     def __init__(
         self,
@@ -58,6 +86,7 @@ class TrainingWorker(QThread):
         learning_rate: float,
         val_split: float,
         input_format: str = "directory",
+        plot_curves: bool = False,
     ):
         super().__init__()
         self.images_dir = images_dir
@@ -73,32 +102,63 @@ class TrainingWorker(QThread):
         self.learning_rate = learning_rate
         self.val_split = val_split
         self.input_format = input_format
+        self.plot_curves = plot_curves
+        self._epoch_pattern = None
+
+    def _parse_output_line(self, line: str):
+        """Parse a line of output for epoch progress."""
+        import re
+
+        if self._epoch_pattern is None:
+            self._epoch_pattern = re.compile(
+                r"Epoch (\d+)/(\d+): Train Loss: ([\d.]+), Val Loss: ([\d.]+), "
+                r"Val IoU: ([\d.]+), Val F1: ([\d.]+)"
+            )
+
+        match = self._epoch_pattern.search(line)
+        if match:
+            current_epoch = int(match.group(1))
+            total_epochs = int(match.group(2))
+            metrics = (
+                f"Loss: {match.group(3)}, IoU: {match.group(5)}, F1: {match.group(6)}"
+            )
+            self.epoch_progress.emit(current_epoch, total_epochs, metrics)
 
     def run(self):
         """Execute the training."""
+        import sys
+
         try:
             import geoai
 
             self.progress.emit("Starting training...")
 
-            geoai.train_segmentation_model(
-                images_dir=self.images_dir,
-                labels_dir=self.labels_dir,
-                output_dir=self.output_dir,
-                input_format=self.input_format,
-                architecture=self.architecture,
-                encoder_name=self.encoder_name,
-                encoder_weights=(
-                    self.encoder_weights if self.encoder_weights != "None" else None
-                ),
-                num_channels=self.num_channels,
-                num_classes=self.num_classes,
-                batch_size=self.batch_size,
-                num_epochs=self.num_epochs,
-                learning_rate=self.learning_rate,
-                val_split=self.val_split,
-                verbose=True,
-            )
+            # Capture stdout to parse epoch progress in real-time
+            old_stdout = sys.stdout
+            sys.stdout = OutputCapture(self._parse_output_line, old_stdout)
+
+            try:
+                geoai.train_segmentation_model(
+                    images_dir=self.images_dir,
+                    labels_dir=self.labels_dir,
[…]
