### 2025-12-11 90821fb
Update qgis_plugin/geoai_plugin/dialogs/moondream.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
