### 2025-12-11 32c4f89
Fix input height error
