### 2025-12-11 07ce22c
Add QGIS plugin (#386)
* Add QGIS plugin

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update qgis_plugin/geoai_plugin/dialogs/moondream.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update qgis_plugin/README.md

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update qgis_plugin/README.md

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update qgis_plugin/geoai_plugin/geoai_plugin.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* 
[…]
