+++ qgis_plugin/geoai_plugin/dialogs/segmentation.py
@@ -1,0 +1,1450 @@
+"""
+Segmentation Dock Widget for GeoAI Plugin
+
+This dock widget provides an interface for training semantic segmentation models
+and running inference, combined in a single dockable panel.
+"""
+
+import os
+from typing import Optional
+
+from qgis.PyQt.QtCore import Qt, QThread, pyqtSignal
+from qgis.PyQt.QtWidgets import (
+    QDockWidget,
+    QWidget,
+    QVBoxLayout,
+    QHBoxLayout,
+    QLabel,
+    QLineEdit,
+    QComboBox,
+    QPushButton,
+    QTextEdit,
+    QFileDialog,
+    QGroupBox,
+    QFormLayout,
+    QProgressBar,
+    QMessageBox,
+    QSpinBox,
+    QDoubleSpinBox,
+    QCheckBox,
+    QTabWidget,
+    QScrollArea,
+)
+
+from qgis.core import QgsProject, QgsVectorLayer, QgsRasterLayer
+from qgis.gui import QgsMapLayerComboBox
+from qgis.core import QgsMapLayerProxyModel
+
+
+class OutputCapture:
+    """Capture stdout and emit lines to a callback in real-time."""
+
+    def __init__(self, callback, original_stdout):
+        self.callback = callback
+        self.original_stdout = original_stdout
+        self.buffer = ""
+
+    def write(self, text):
+        # Also write to original stdout
+        if self.original_stdout:
+            self.original_stdout.write(text)
+
+        self.buffer += text
+        while "\n" in self.buffer:
+            line, self.buffer = self.buffer.split("\n", 1)
+            if line.strip():
+                self.callback(line)
+
+    def flush(self):
+        if self.original_stdout:
+            self.original_stdout.flush()
+        if self.buffer.strip():
+            self.callback(self.buffer)
+            self.buffer = ""
+
+
+class TrainingWorker(QThread):
+    """Worker thread for training segmentation models."""
+
+    finished = pyqtSignal(str)
+    error = pyqtSignal(str)
+    progress = pyqtSignal(str)
+    epoch_progress = pyqtSignal(int, int, str)  # current_epoch, total_epochs, metrics
+
+    def __init__(
+        self,
+        images_dir: str,
+        labels_dir: str,
+        output_dir: str,
+        architecture: str,
+        encoder_name: str,
+        encoder_weights: str,
+        num_channels: int,
+        num_classes: int,
+        batch_size: int,
+        num_epochs: int,
+        learning_rate: float,
+        val_split: float,
+        input_format: str = "directory",
+        plot_curves: bool = False,
+    ):
+        super().__init__()
+        self.images_dir = images_dir
+        self.labels_dir = labels_dir
+        self.output_dir = output_dir
+        self.architecture = architecture
+        self.encoder_name = encoder_name
+        self.encoder_weights = encoder_weights
+        self.num_channels = num_channels
+        self.num_classes = num_classes
+        self.batch_size = batch_size
+        self.num_epochs = num_epochs
+        self.learning_rate = learning_rate
+        self.val_split = val_split
+        self.input_format = input_format
+        self.plot_curves = plot_curves
+        self._epoch_pattern = None
+
+    def _parse_output_line(self, line: str):
+        """Parse a line of output for epoch progress."""
+        import re
+
+        if self._epoch_pattern is None:
+            self._epoch_pattern = re.compile(
+                r"Epoch (\d+)/(\d+): Train Loss: ([\d.]+), Val Loss: ([\d.]+), "
+                r"Val IoU: ([\d.]+), Val F1: ([\d.]+)"
+            )
+
+        match = self._epoch_pattern.search(line)
+        if match:
+            current_epoch = int(match.group(1))
+            total_epochs = int(match.group(2))
+            metrics = (
[…]
