### 2025-12-11 c5186d9
Add SamGeo panel to the QGIS plugin
