+++ qgis_plugin/geoai_plugin/geoai_plugin.py
@@ -30,6 +30,7 @@ class GeoAIPlugin:
         # Dock widgets (lazy loaded)
         self._moondream_dock = None
         self._segmentation_dock = None
+        self._samgeo_dock = None
 
     def add_action(
         self,
@@ -101,6 +102,10 @@ class GeoAIPlugin:
         if not os.path.exists(segment_icon):
             segment_icon = ":/images/themes/default/mActionAddRasterLayer.svg"
 
+        samgeo_icon = os.path.join(icon_base, "samgeo.png")
+        if not os.path.exists(samgeo_icon):
+            samgeo_icon = ":/images/themes/default/mIconPolygonLayer.svg"
+
         about_icon = os.path.join(icon_base, "about.svg")
         if not os.path.exists(about_icon):
             about_icon = ":/images/themes/default/mActionHelpContents.svg"
@@ -121,6 +126,16 @@ class GeoAIPlugin:
             "Segmentation",
             self.toggle_segmentation_dock,
             status_tip="Toggle Segmentation Training & Inference panel",
+            checkable=True,
+            parent=self.iface.mainWindow(),
+        )
+
+        # Add SamGeo action (checkable for dock toggle)
+        self.samgeo_action = self.add_action(
+            samgeo_icon,
+            "SamGeo",
+            self.toggle_samgeo_dock,
+            status_tip="Toggle SamGeo Segmentation panel (SAM, SAM2, SAM3)",
             checkable=True,
             parent=self.iface.mainWindow(),
         )
@@ -167,6 +182,11 @@ class GeoAIPlugin:
             self.iface.removeDockWidget(self._segmentation_dock)
             self._segmentation_dock.deleteLater()
             self._segmentation_dock = None
+
+        if self._samgeo_dock:
+            self.iface.removeDockWidget(self._samgeo_dock)
+            self._samgeo_dock.deleteLater()
+            self._samgeo_dock = None
 
         # Remove actions from menu
         for action in self.actions:
@@ -258,6 +278,44 @@ class GeoAIPlugin:
         """Handle Segmentation dock visibility change."""
         self.segmentation_action.setChecked(visible)
 
+    def toggle_samgeo_dock(self):
+        """Toggle the SamGeo dock widget visibility."""
+        if self._samgeo_dock is None:
+            try:
+                from .dialogs.samgeo import SamGeoDockWidget
+
+                self._samgeo_dock = SamGeoDockWidget(
+                    self.iface, self.iface.mainWindow()
+                )
+                self._samgeo_dock.setObjectName("GeoAISamGeoDock")
+                self._samgeo_dock.visibilityChanged.connect(
+                    self._on_samgeo_visibility_changed
+                )
+                self.iface.addDockWidget(Qt.RightDockWidgetArea, self._samgeo_dock)
+                self._samgeo_dock.show()
+                self._samgeo_dock.raise_()
+                return
+
+            except Exception as e:
+                QMessageBox.critical(
+                    self.iface.mainWindow(),
+                    "Error",
+                    f"Failed to create SamGeo panel:\n{str(e)}",
+                )
+                self.samgeo_action.setChecked(False)
+                return
+
+        # Toggle visibility
+        if self._samgeo_dock.isVisible():
+            self._samgeo_dock.hide()
+        else:
+            self._samgeo_dock.show()
+            self._samgeo_dock.raise_()
+
+    def _on_samgeo_visibility_changed(self, visible):
+        """Handle SamGeo dock visibility change."""
+        self.samgeo_action.setChecked(visible)
+
     def clear_gpu_memory(self):
         """Clear GPU memory and release CUDA resources."""
         import gc
@@ -325,6 +383,50 @@ class GeoAIPlugin:
[…]
