### 2025-12-11 65de523
Update qgis_plugin/geoai_plugin/dialogs/samgeo.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
