### 2025-12-11 45a4db0
Update qgis_plugin/geoai_plugin/geoai_plugin.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
