### 2025-12-11 e3ffe4c
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
