### 2025-12-11 5379a82
Update qgis_plugin/geoai_plugin/dialogs/map_tools.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
