### 2025-12-11 2ab4fea
Update qgis_plugin/geoai_plugin/dialogs/samgeo.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
