### 2025-12-11 a8327bb
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
