### 2025-12-11 444d9e0
Add SamGeo panel to the QGIS plugin (#387)
* Add SamGeo panel to the QGIS plugin

* Update qgis_plugin/geoai_plugin/dialogs/samgeo.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update qgis_plugin/geoai_plugin/geoai_plugin.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update qgis_plugin/geoai_plugin/dialogs/samgeo.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update qgis_plugin/geoai_plugin/geoai_plugin.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* [pre-commit.ci] auto fixes f
[…]
