+++ qgis_plugin/geoai_plugin/dialogs/samgeo.py
@@ -1,0 +1,1438 @@
+"""
+SamGeo Dock Widget for GeoAI Plugin
+
+This dock widget provides an interface for remote sensing image segmentation
+using the SamGeo library (SAM, SAM2, and SAM3 models).
+"""
+
+import os
+
+from qgis.PyQt.QtCore import Qt, QCoreApplication
+from qgis.PyQt.QtGui import QColor
+from qgis.PyQt.QtWidgets import (
+    QDockWidget,
+    QWidget,
+    QVBoxLayout,
+    QHBoxLayout,
+    QLabel,
+    QPushButton,
+    QLineEdit,
+    QComboBox,
+    QSpinBox,
+    QDoubleSpinBox,
+    QCheckBox,
+    QGroupBox,
+    QFileDialog,
+    QMessageBox,
+    QProgressBar,
+    QTabWidget,
+    QTextEdit,
+    QListWidget,
+    QListWidgetItem,
+    QScrollArea,
+)
+from qgis.core import (
+    QgsProject,
+    QgsRasterLayer,
+    QgsVectorLayer,
+    QgsWkbTypes,
+    Qgis,
+    QgsMessageLog,
+)
+
+from .map_tools import PointPromptTool, BoxPromptTool
+
+
+class SamGeoDockWidget(QDockWidget):
+    """Dock widget for SamGeo segmentation operations."""
+
+    def __init__(self, iface, parent=None):
+        """Initialize the SamGeo dock widget.
+
+        Args:
+            iface: The QGIS interface instance.
+            parent: Parent widget.
+        """
+        super().__init__("SamGeo Segmentation", parent)
+        self.iface = iface
+        self.canvas = iface.mapCanvas()
+        self.setAllowedAreas(Qt.LeftDockWidgetArea | Qt.RightDockWidgetArea)
+
+        # SamGeo model instance
+        self.sam = None
+        self.current_layer = None
+        self.current_image_path = None
+
+        # Point and box prompts
+        self.point_coords = []
+        self.point_labels = []
+        self.box_coords = None
+
+        # Batch point prompts
+        self.batch_point_coords = []
+        self.batch_point_coords_map = []  # Map coordinates for display
+
+        # Map tools
+        self.point_tool = None
+        self.batch_point_tool = None
+        self.box_tool = None
+        self.previous_tool = None
+
+        self._setup_ui()
+
+    def _setup_ui(self):
+        """Set up the user interface."""
+        # Main widget with scroll area
+        scroll_area = QScrollArea()
+        scroll_area.setWidgetResizable(True)
+        scroll_area.setHorizontalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
+
+        main_widget = QWidget()
+        main_layout = QVBoxLayout()
+        main_widget.setLayout(main_layout)
+
+        # Tab widget for different modes
+        tab_widget = QTabWidget()
+
+        # === Model Settings Tab ===
+        model_tab = self._create_model_tab()
+        tab_widget.addTab(model_tab, "Model")
+
+        # === Text Prompts Tab ===
+        text_tab = self._create_text_tab()
+        tab_widget.addTab(text_tab, "Text")
+
+        # === Interactive Tab ===
+        interactive_tab = self._create_interactive_tab()
+        tab_widget.addTab(interactive_tab, "Interactive")
+
+        # === Batch Tab ===
+        batch_tab = self._create_batch_tab()
+        tab_widget.addTab(batch_tab, "Batch")
+
+        # === Output Tab ===
+        output_tab = self._create_output_tab()
+        tab_widget.addTab(output_tab, "Output")
+
+        # Add tab widget to main layout
+        main_layout.addWidget(tab_widget)
+
[…]

[…]
