### 2025-12-11 ed121fa
Bump version: 0.19.0 → 0.20.0
