+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "geoai-py"
-version = "0.19.0"
+version = "0.20.0"
 dynamic = [
     "dependencies",
 ]
@@ -45,7 +45,7 @@ universal = true
 
 
 [tool.bumpversion]
-current_version = "0.19.0"
+current_version = "0.20.0"
 commit = true
 tag = true
 
