### 2025-12-12 187a03a
Add auto module for supporting HF models
