### 2025-12-12 a9cbd69
Update geoai/auto.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
