### 2025-12-12 6479ce7
Update geoai/auto.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
