### 2025-12-12 5db963e
Update geoai/auto.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
