### 2025-12-12 4db9129
Add auto module for supporting HF models (#390)
* Add auto module for supporting HF models

* Update geoai/auto.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update docs/examples/AutoML.ipynb

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update docs/examples/AutoML.ipynb

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update geoai/auto.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update geoai/auto.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update geo
[…]
