+++ geoai/auto.py
@@ -1,0 +1,1982 @@
+"""Auto classes for geospatial model inference with GeoTIFF support.
+
+This module provides AutoGeoModel and AutoGeoImageProcessor that extend
+Hugging Face transformers' AutoModel and AutoImageProcessor to support
+processing geospatial data (GeoTIFF) and saving outputs as GeoTIFF or vector data.
+
+Supported tasks:
+    - Semantic segmentation (e.g., SegFormer, Mask2Former)
+    - Image classification (e.g., ViT, ResNet)
+    - Object detection (e.g., DETR, YOLOS)
+    - Zero-shot object detection (e.g., Grounding DINO, OWL-ViT)
+    - Depth estimation (e.g., Depth Anything, DPT)
+    - Mask generation (e.g., SAM)
+
+Example:
+    >>> from geoai import AutoGeoModel
+    >>> model = AutoGeoModel.from_pretrained(
+    ...     "nvidia/segformer-b0-finetuned-ade-512-512",
+    ...     task="semantic-segmentation"
+    ... )
+    >>> result = model.predict("input.tif", output_path="output.tif")
+"""
+
+import os
+from typing import Any, Dict, List, Optional, Tuple, Union
+
+import cv2
+import geopandas as gpd
+import numpy as np
+import rasterio
+import requests
+import torch
+from PIL import Image
+from rasterio.features import shapes
+from rasterio.windows import Window
+from shapely.geometry import box, shape
+from tqdm import tqdm
+from transformers import (
+    AutoConfig,
+    AutoImageProcessor,
+    AutoModel,
+    AutoModelForImageClassification,
+    AutoModelForImageSegmentation,
+    AutoModelForSemanticSegmentation,
+    AutoModelForUniversalSegmentation,
+    AutoModelForDepthEstimation,
+    AutoModelForMaskGeneration,
+    AutoModelForObjectDetection,
+    AutoModelForZeroShotObjectDetection,
+    AutoProcessor,
+)
+
+from transformers.utils import logging as hf_logging
+
+from .utils import get_device
+
+
+hf_logging.set_verbosity_error()  # silence HF load reports
+
+
+class AutoGeoImageProcessor:
+    """
+    Image processor for geospatial data that wraps AutoImageProcessor.
+
+    This class provides functionality to load and preprocess GeoTIFF images
+    while preserving geospatial metadata (CRS, transform, bounds). It wraps
+    Hugging Face's AutoImageProcessor and adds geospatial capabilities.
+
+    Use `from_pretrained` to instantiate this class, following the transformers pattern.
+
+    Attributes:
+        processor: The underlying AutoImageProcessor instance.
+        device (str): The device being used ('cuda' or 'cpu').
+
+    Example:
+        >>> processor = AutoGeoImageProcessor.from_pretrained("facebook/sam-vit-base")
+        >>> data, metadata = processor.load_geotiff("input.tif")
+        >>> inputs = processor(data)
+    """
+
+    def __init__(
+        self,
+        processor: "AutoImageProcessor",
+        processor_name: Optional[str] = None,
+        device: Optional[str] = None,
+    ) -> None:
+        """Initialize the AutoGeoImageProcessor with an existing processor.
+
+        Note: Use `from_pretrained` class method to load from Hugging Face Hub.
+
+        Args:
+            processor: An AutoImageProcessor instance.
+            processor_name: Name or path of the processor (for reference).
+            device: Device to use ('cuda', 'cpu'). If None, auto-detect.
+        """
+        self.processor = processor
+        self.processor_name = processor_name
+
+        if device is None:
+            self.device = get_device()
+        else:
+            self.device = device
+
+    @classmethod
+    def from_pretrained(
[…]
