### 2025-12-12 ac2640d
Improve QGIS plugin documentation
