+++ mkdocs.yml
@@ -96,6 +96,7 @@ markdown_extensions:
 nav:
     - Home: index.md
     - Installation: installation.md
+    - QGIS Plugin: qgis_plugin.md
     - Contributing: contributing.md
     - Changelog: https://github.com/opengeos/geoai/releases
     - Report Issues: https://github.com/opengeos/geoai/issues
