+++ docs/qgis_plugin.md
@@ -2,38 +2,44 @@
 
 A QGIS plugin providing AI-powered geospatial analysis tools from the [geoai](https://github.com/opengeos/geoai) package.
 
+![demo](https://github.com/user-attachments/assets/557feb58-ca58-4e27-800f-f3e8a8d3d362)
+
 ## Features
 
 The plugin provides **dockable panels** that can be attached to the left or right side of QGIS.
 
 ### Moondream Vision-Language Model Panel
-- **Caption**: Generate descriptions of geospatial imagery (short, normal, or long)
-- **Query**: Ask questions about images using natural language
-- **Detect**: Detect and locate objects with bounding boxes
-- **Point**: Locate specific objects with point markers
+
+-   **Caption**: Generate descriptions of geospatial imagery (short, normal, or long)
+-   **Query**: Ask questions about images using natural language
+-   **Detect**: Detect and locate objects with bounding boxes
+-   **Point**: Locate specific objects with point markers
 
 ### Segmentation Panel (Combined Training & Inference)
-- **Tab 1 - Create Training Data**: Export GeoTIFF tiles from raster and vector data
-- **Tab 2 - Train Models**: Train custom segmentation models (U-Net, DeepLabV3+, FPN, etc.)
-- **Tab 3 - Run Inference**: Apply trained models to new imagery and vectorize results
+
+-   **Tab 1 - Create Training Data**: Export GeoTIFF tiles from raster and vector data
+-   **Tab 2 - Train Models**: Train custom segmentation models (U-Net, DeepLabV3+, FPN, etc.)
+-   **Tab 3 - Run Inference**: Apply trained models to new imagery and vectorize results
 
 ### SamGeo Panel (Segment Anything Model)
-- **Model Tab**: Load SAM models (SAM1, SAM2, or SAM3) with configurable backend and device settings
-- **Text Tab**: Segment objects using text prompts (e.g., "tree", "building", "road")
-- **Interactive Tab**: Segment using point prompts (foreground/background) or box prompts drawn on the map
-- **Batch Tab**: Process multiple points interactively or from vector files/layers
-- **Output Tab**: Save results as raster (GeoTIFF) or vector (GeoPackage, Shapefile) with optional regularization (orthogonalize polygons, filter by minimum area)
+
+-   **Model Tab**: Load SAM models (SAM1, SAM2, or SAM3) with configurable backend and device settings
+-   **Text Tab**: Segment objects using text prompts (e.g., "tree", "building", "road")
+-   **Interactive Tab**: Segment using point prompts (foreground/background) or box prompts drawn on the map
+-   **Batch Tab**: Process multiple points interactively or from vector files/layers
+-   **Output Tab**: Save results as raster (GeoTIFF) or vector (GeoPackage, Shapefile) with optional regularization (orthogonalize polygons, filter by minimum area)
 
 ### GPU Memory Management
-- **Clear GPU Memory**: Release GPU memory and clear CUDA cache for all loaded models
+
+-   **Clear GPU Memory**: Release GPU memory and clear CUDA cache for all loaded models
 
 ## Requirements
 
-- QGIS 3.28 or later
-- Python 3.10+
-- PyTorch (with CUDA support for GPU acceleration)
-- geoai-py package
-- samgeo package (for SamGeo panel)
+-   QGIS 3.28 or later
+-   Python 3.10+
+-   PyTorch (with CUDA support for GPU acceleration)
+-   geoai-py package
+-   samgeo package (for SamGeo panel)
 
 ## Installation
 
[…]
