### 2025-12-12 6707546
Update plugin docs
