+++ docs/qgis_plugin.md
@@ -1,12 +1,18 @@
 # QGIS Plugin for GeoAI
 
-A QGIS plugin providing AI-powered geospatial analysis tools from the [geoai](https://github.com/opengeos/geoai) package.
+A QGIS plugin providing AI-powered geospatial analysis tools from the [geoai](https://github.com/opengeos/geoai) package. The plugin bundles several model-driven workflows (Moondream VLM, segmentation training/inference, SamGeo) into dockable panels so you can keep QGIS as the main workspace while experimenting with GeoAI.
 
-![demo](https://github.com/user-attachments/assets/557feb58-ca58-4e27-800f-f3e8a8d3d362)
+## Quick Start
+
+1. Create a fresh conda environment (`conda create -n geo python=3.12`) and install QGIS plus GeoAI dependencies as described below.
+2. Install the GeoAI QGIS plugin either from the script (`python install.py`) or by copying the `qgis_plugin` folder into your QGIS profile.
+3. Restart QGIS, enable the plugin under `Plugins` → `Manage and Install Plugins...`, and open one of the panels from the GeoAI toolbar to begin experimenting with sample datasets.
+
+[![demo](https://github.com/user-attachments/assets/557feb58-ca58-4e27-800f-f3e8a8d3d362)](https://github.com/user-attachments/assets/557feb58-ca58-4e27-800f-f3e8a8d3d362)
 
 ## Features
 
-The plugin provides **dockable panels** that can be attached to the left or right side of QGIS.
+Each tool lives inside a dockable panel that can be attached to either side of the QGIS interface, so you can keep layers, maps, and models visible simultaneously.
 
 ### Moondream Vision-Language Model Panel
 
@@ -19,7 +25,7 @@ The plugin provides **dockable panels** that can be attached to the left or righ
 
 -   **Tab 1 - Create Training Data**: Export GeoTIFF tiles from raster and vector data
 -   **Tab 2 - Train Models**: Train custom segmentation models (U-Net, DeepLabV3+, FPN, etc.)
--   **Tab 3 - Run Inference**: Apply trained models to new imagery and vectorize results
+-   **Tab 3 - Run Inference**: Apply trained models to new imagery and vectorize results. Vector outputs can optionally be smoothed or simplified for immediate use in GIS workflows.
 
 ### SamGeo Panel (Segment Anything Model)
 
@@ -43,51 +49,11 @@ The plugin provides **dockable panels** that can be attached to the left or righ
 
 ## Installation
 
-### Option 1: Install from Plugin Manager (Recommended)
+### Install QGIS and Python Packages
 
-1. Open QGIS
-2. Go to `Plugins` → `Manage and Install Plugins...`
-3. Search for "GeoAI"
-4. Click "Install Plugin"
+Before using the plugin, create a new conda environment dedicated to QGIS and the Python dependencies. Please do not try to reuse an existing QGIS install—mismatched Python packages almost always break the plugin.
 
-### Option 2: Install Using Script (Recommended for Developers)
-
-1. Clone or download this repository
-2. Run the installation script:
-
-    **Linux/macOS:**
-
-    ```bash
-    python install.py
-    ```
-
-    **Windows:**
-
-    ```cmd
-    python install.py
-    ```
-
-3. Restart QGIS
-4. Enable the plugin in `Plugins` → `Manage and Install Plugins...`
-
-To remove the plugin:
-
-```bash
-python install.py --remove
-```
-
-### Option 3: Manual Installation
-
-1. Copy the `geoai_plugin` folder to your QGIS plugins directory:
[…]
