+++ qgis_plugin/README.md
@@ -1,84 +1,59 @@
 # QGIS Plugin for GeoAI
 
-A QGIS plugin providing AI-powered geospatial analysis tools from the [geoai](https://github.com/opengeos/geoai) package.
+A QGIS plugin providing AI-powered geospatial analysis tools from the [geoai](https://github.com/opengeos/geoai) package. The plugin bundles several model-driven workflows (Moondream VLM, segmentation training/inference, SamGeo) into dockable panels so you can keep QGIS as the main workspace while experimenting with GeoAI.
+
+## Quick Start
+
+1. Create a fresh conda environment (`conda create -n geo python=3.12`) and install QGIS plus GeoAI dependencies as described below.
+2. Install the GeoAI QGIS plugin either from the script (`python install.py`) or by copying the `qgis_plugin` folder into your QGIS profile.
+3. Restart QGIS, enable the plugin under `Plugins` → `Manage and Install Plugins...`, and open one of the panels from the GeoAI toolbar to begin experimenting with sample datasets.
+
+[![demo](https://github.com/user-attachments/assets/557feb58-ca58-4e27-800f-f3e8a8d3d362)](https://github.com/user-attachments/assets/557feb58-ca58-4e27-800f-f3e8a8d3d362)
 
 ## Features
 
-The plugin provides **dockable panels** that can be attached to the left or right side of QGIS.
+Each tool lives inside a dockable panel that can be attached to either side of the QGIS interface, so you can keep layers, maps, and models visible simultaneously.
 
 ### Moondream Vision-Language Model Panel
-- **Caption**: Generate descriptions of geospatial imagery (short, normal, or long)
-- **Query**: Ask questions about images using natural language
-- **Detect**: Detect and locate objects with bounding boxes
-- **Point**: Locate specific objects with point markers
+
+-   **Caption**: Generate descriptions of geospatial imagery (short, normal, or long)
+-   **Query**: Ask questions about images using natural language
+-   **Detect**: Detect and locate objects with bounding boxes
+-   **Point**: Locate specific objects with point markers
 
 ### Segmentation Panel (Combined Training & Inference)
-- **Tab 1 - Create Training Data**: Export GeoTIFF tiles from raster and vector data
-- **Tab 2 - Train Models**: Train custom segmentation models (U-Net, DeepLabV3+, FPN, etc.)
-- **Tab 3 - Run Inference**: Apply trained models to new imagery and vectorize results
+
+-   **Tab 1 - Create Training Data**: Export GeoTIFF tiles from raster and vector data
+-   **Tab 2 - Train Models**: Train custom segmentation models (U-Net, DeepLabV3+, FPN, etc.)
+-   **Tab 3 - Run Inference**: Apply trained models to new imagery and vectorize results. Vector outputs can optionally be smoothed or simplified for immediate use in GIS workflows.
 
 ### SamGeo Panel (Segment Anything Model)
-- **Model Tab**: Load SAM models (SAM1, SAM2, or SAM3) with configurable backend and device settings
-- **Text Tab**: Segment objects using text prompts (e.g., "tree", "building", "road")
-- **Interactive Tab**: Segment using point prompts (foreground/background) or box prompts drawn on the map
-- **Batch Tab**: Process multiple points interactively or from vector files/layers
-- **Output Tab**: Save results as raster (GeoTIFF) or vector (GeoPackage, Shapefile) with optional regularization (orthogonalize polygons, filter by minimum area)
+
[…]
