### 2025-12-12 9fe8307
Add QGIS plugin to home page
