### 2025-12-12 944fc40
Improve QGIS plugin documentation (#391)
* Improve QGIS plugin documentation

* Update docs

* Update plugin docs

* Add QGIS plugin to home page

* Update docs
