### 2025-12-12 6944374
Fix QGIS plugin import issue
