### 2025-12-12 c78ac47
Update qgis_plugin/geoai_plugin/_geoai_lib.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
