### 2025-12-12 227677c
Update qgis_plugin/geoai_plugin/_geoai_lib.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
