### 2025-12-12 4e6e597
Update qgis_plugin/geoai_plugin/metadata.txt
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
