### 2025-12-12 501b5a5
Update qgis_plugin/geoai_plugin/_geoai_lib.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
