### 2025-12-12 2c613b8
Update qgis_plugin/geoai_plugin/_geoai_lib.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
