### 2025-12-12 d1ac4cf
Fix QGIS plugin import issue (#392)
* Fix QGIS plugin import issue

* Update qgis_plugin/geoai_plugin/_geoai_lib.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update qgis_plugin/geoai_plugin/_geoai_lib.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update qgis_plugin/geoai_plugin/_geoai_lib.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update qgis_plugin/geoai_plugin/metadata.txt

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update qgis_plugin/geoai_plugin/_geoai_lib.py

Co-autho
[…]
