### 2025-12-12 a6dcd67
Update QGIS installation instructions
