+++ qgis_plugin/README.md
@@ -1,14 +1,22 @@
 # QGIS Plugin for GeoAI
 
-A QGIS plugin providing AI-powered geospatial analysis tools from the [geoai](https://github.com/opengeos/geoai) package. The plugin bundles several model-driven workflows (Moondream VLM, segmentation training/inference, SamGeo) into dockable panels so you can keep QGIS as the main workspace while experimenting with GeoAI.
+A QGIS plugin that brings the [geoai](https://github.com/opengeos/geoai) models into dockable panels (Moondream VLM, segmentation training/inference, SamGeo) so you can keep QGIS as your main workspace while experimenting with GeoAI.
 
 ## Quick Start
 
-1. Create a fresh conda environment (`conda create -n geo python=3.12`) and install QGIS plus GeoAI dependencies as described below.
-2. Install the GeoAI QGIS plugin either from the script (`python install.py`) or by copying the `qgis_plugin` folder into your QGIS profile.
-3. Restart QGIS, enable the plugin under `Plugins` → `Manage and Install Plugins...`, and open one of the panels from the GeoAI toolbar to begin experimenting with sample datasets.
+-   Create a fresh conda env (`conda create -n geo python=3.12`) and install QGIS + deps (see below).
+-   Install the plugin (`python install.py`) from this repo.
+-   Restart QGIS → `Plugins` → `Manage and Install Plugins...` → enable `GeoAI`.
+-   Open a GeoAI toolbar panel and try the sample datasets below.
 
 [![demo](https://github.com/user-attachments/assets/557feb58-ca58-4e27-800f-f3e8a8d3d362)](https://github.com/user-attachments/assets/557feb58-ca58-4e27-800f-f3e8a8d3d362)
+
+## Requirements
+
+-   QGIS 3.28 or later
+-   Python 3.10+ (conda recommended)
+-   PyTorch (CUDA if you want GPU acceleration)
+-   `geoai` and `samgeo` packages
 
 ## Features
 
@@ -39,35 +47,45 @@ Each tool lives inside a dockable panel that can be attached to either side of t
 
 -   **Clear GPU Memory**: Release GPU memory and clear CUDA cache for all loaded models
 
-## Requirements
-
--   QGIS 3.28 or later
--   Python 3.10+
--   PyTorch (with CUDA support for GPU acceleration)
--   geoai-py package
--   samgeo package (for SamGeo panel)
-
 ## Installation
 
-### Install QGIS and Python Packages
+### 1) Set up the environment
 
-Before using the plugin, create a new conda environment dedicated to QGIS and the Python dependencies. Please do not try to reuse an existing QGIS install—mismatched Python packages almost always break the plugin.
-
-It is recommended to use [Miniconda](https://www.anaconda.com/docs/getting-started/miniconda/install) to manage conda environments. Once Miniconda is installed, open the Terminal or Anaconda Prompt and run the following commands to create a new conda environment and install QGIS and Python packages:
+Use a clean conda env dedicated to QGIS—mixing with an existing QGIS install often breaks dependencies.
 
 ```bash
 conda create -n geo python=3.12
 conda activate geo
-conda install -c conda-forge qgis segment-geospatial geoai
 ```
 
-Some SamGeo dependencies are only available on PyPI. Run the following command to install the remaining dependencies:
+Install core geospatial deps first:
 
 ```bash
-pip install -U "segment-geospatial[samgeo3]"
+conda install -c conda-forge --strict-channel-priority gdal rasterio libnetcdf netcdf4
[…]
