### 2025-12-12 085734e
Update QGIS plugin sample datasets
