### 2025-12-13 636024f
Add support for auto showing results for SamGeo
