+++ qgis_plugin/geoai_plugin/dialogs/samgeo.py
@@ -92,30 +92,30 @@ class SamGeoDockWidget(QDockWidget):
         main_widget.setLayout(main_layout)
 
         # Tab widget for different modes
-        tab_widget = QTabWidget()
+        self.tab_widget = QTabWidget()
 
         # === Model Settings Tab ===
         model_tab = self._create_model_tab()
-        tab_widget.addTab(model_tab, "Model")
+        self.tab_widget.addTab(model_tab, "Model")
 
         # === Text Prompts Tab ===
         text_tab = self._create_text_tab()
-        tab_widget.addTab(text_tab, "Text")
+        self.tab_widget.addTab(text_tab, "Text")
 
         # === Interactive Tab ===
         interactive_tab = self._create_interactive_tab()
-        tab_widget.addTab(interactive_tab, "Interactive")
+        self.tab_widget.addTab(interactive_tab, "Interactive")
 
         # === Batch Tab ===
         batch_tab = self._create_batch_tab()
-        tab_widget.addTab(batch_tab, "Batch")
+        self.tab_widget.addTab(batch_tab, "Batch")
 
         # === Output Tab ===
         output_tab = self._create_output_tab()
-        tab_widget.addTab(output_tab, "Output")
+        self.tab_widget.addTab(output_tab, "Output")
 
         # Add tab widget to main layout
-        main_layout.addWidget(tab_widget)
+        main_layout.addWidget(self.tab_widget)
 
         # Progress bar
         self.progress_bar = QProgressBar()
@@ -520,6 +520,15 @@ class SamGeoDockWidget(QDockWidget):
         self.add_to_map_check.setChecked(True)
         output_group_layout.addWidget(self.add_to_map_check)
 
+        # Auto-show results after segmentation
+        self.auto_show_check = QCheckBox("Auto-show results after segmentation")
+        self.auto_show_check.setChecked(True)
+        self.auto_show_check.setToolTip(
+            "Automatically save and display results after running segmentation\n"
+            "in the Text, Interactive, or Batch tabs"
+        )
+        output_group_layout.addWidget(self.auto_show_check)
+
         # Output path
         output_path_row = QHBoxLayout()
         self.output_path_edit = QLineEdit()
@@ -858,9 +867,12 @@ class SamGeoDockWidget(QDockWidget):
 
             # Update status label
             if num_masks > 0:
-                self.text_status_label.setText(
-                    f"Found {num_masks} object(s). Go to Output tab to save."
-                )
+                if self.auto_show_check.isChecked():
+                    self.text_status_label.setText(f"Found {num_masks} object(s).")
+                else:
+                    self.text_status_label.setText(
+                        f"Found {num_masks} object(s). Go to Output tab to save."
+                    )
                 self.text_status_label.setStyleSheet("color: green;")
             else:
                 self.text_status_label.setText(
@@ -869,6 +881,10 @@ class SamGeoDockWidget(QDockWidget):
                 self.text_status_label.setStyleSheet("color: orange;")
 
             self.log_message(f"Text segmentation complete. Found {num_masks} objects.")
+
+            # Auto-show results if enabled
+            if num_masks > 0:
+                self._auto_show_results()
 
         except Exception as e:
             self.text_status_label.setText("Segmentation failed!")
@@ -1029,9 +1045,12 @@ class SamGeoDockWidget(QDockWidget):
 
             # Update status label
             if num_masks > 0:
-                self.point_status_label.setText(
-                    f"Found {num_masks} object(s). Go to Output tab to save."
-                )
+                if self.auto_show_check.isChecked():
+                    self.point_status_label.setText(f"Found {num_masks} object(s).")
+                else:
+                    self.point_status_label.setText(
[…]
