### 2025-12-13 f57f9ca
Update qgis_plugin/geoai_plugin/dialogs/samgeo.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
