### 2025-12-13 09fe6df
Add support for auto showing results for SamGeo (#397)
* Add support for auto showing results for SamGeo

* Update qgis_plugin/geoai_plugin/dialogs/samgeo.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update qgis_plugin/geoai_plugin/dialogs/samgeo.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

---------

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
