### 2025-12-13 8eed2ff
Add band selection support for SamGeo
