### 2025-12-14 #9605 @gmarupilla
@giswqs I ran into a macOS-specific issue and wanted to check if this is expected behavior.

On an Intel Mac (x86_64), importing the package fails with:

`_ARRAY_API not found (Triggered internally at .../torch/csrc/utils/tensor_numpy.cpp:84.)`

On an Apple Silicon machine (M2), install and import work fine and I can see the version without issues.

I tried pinning `numpy<2` as a workaround on Intel, but that runs into dependency conflicts since several geo deps now require NumPy ≥ 2, for example:
- rioxarray 0.20.0 (numpy>=2)
- buildingregulariser 0.2.4 (numpy>=2.0.0)
- opencv-python-headless 4.12.0.88 (numpy>=2)

So at least on Intel mac, it looks hard to satisfy both:
- torch importing correctly, and
- the geo dependencies installing cleanly.

Is Intel macOS currently expected to be unsupported?
If yes, would it make sense to document this (or note any NumPy / torch constraints) so users don’t get stuck here?

Happy to share more details or test things if helpful.
