### 2025-12-14 #9605 @giswqs
@gmarupilla Thank you for reporting. I don't have a Intel Mac for testing. The geoai package does not pin the numpy version. See https://github.com/opengeos/geoai/blob/main/requirements.txt. It is because the upstream dependencies like rioxarray and buildingregulariser that require numpy >=2.0. If you manage to install pytorch, rioxarry and opencv properly before installing geoai, then geoai should work fine. Sometimes using a mix of conda and pip can make it work.

As I said, I don't have an Intel Mac for testing. If it is concluded that it does not work on Intel Mac, I update the docs to reflect that.
