### 2025-12-23 524f597
Add QGIS plugin error handling (#451)
* Add QGIS plugin error handling

* Update qgis_plugin/README.md

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update docs/qgis_plugin.md

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

---------

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
