### 2026-01-07 opengeos/geoai#462 @jbousquin
Broken link in contributing docs
### Environment Information

Browser info: Chrome v143.0.7499.170

### Description

In [Pull Request Guidelines](https://opengeoai.org/contributing/#pull-request-guidelines) section of Contributing docs, bullet 3, the link provided to check pull requests is https://github.com/opengeos/geoai/pull_requests and gives a 404, should it be https://github.com/opengeos/geoai/pulls?

Url appears to be hard coded in [docs/contributing.md](https://github.com/opengeos/geoai/blob/524f597035b04a543a5b03767faa6ae78ee9a0b6/docs/contributing.md?plain=1#L101)

(Submitting this issue as part of JOSS review [9605](https://github.com/openjournals/joss-reviews/issues/9605))
