### 2026-01-29 #9605 @jbousquin
Recommending for acceptance, no further edits required.

The GEOAI package mainly wraps available tools, but unifying them into a single API makes for a more consistent user experience and is very valuable. It's been on my radar and after digging into it I plan to use it in the future.

The package is mature with extensive documentation in a variety of formats, including the API docs, example notebooks, tutorials, videos, etc. CI builds appear to keep most of this synchronized. A couple very minor things:
- Check the link for the tutorials badge (tinyurl coming through as not secure for me).
- Navigation capitalization, e.g., under examples 'Download naip', 'Building footprints usa', etc where it is capitalized everywhere else (may be how the pages are being built).

For contributions/maintenance github actions are used for pytest tests, I see coverage in docs requirements but didn't see coverage reporting for contributions (may be picked up by co-pilot review?). Definitely not required, just thinking that can be useful to ensure feature adds are part of tests during your PR reviews.
