### 2026-01-29 #9605 @giswqs
@jbousquin Thank you for your feedback.

1. The tinyurl issue has been resolved in https://github.com/opengeos/geoai/pull/495
2. The documentation site (https://opengeoai.org) is built with MkDocs, which automatically generates the navigation menu based on file names. All notebook examples follow a consistent lowercase naming convention, which is why entries appear as “Download naip” and “Building footprints usa.”. While we could use each notebook’s level-1 heading for menu labels, many titles are quite long. Given that we currently have 69 notebook examples, we prefer to keep the navigation concise and easy to scan.
3. Pytest coverage for PRs is recommended but not required. We can rely on Copilot review to add or enhance test coverage when needed.
