### 2026-01-29 opengeos/geoai#496 @mengqi-z
JOSS Paper Proofread
Hi @giswqs,

I've proofread the paper for the JOSS review openjournals/joss-reviews#9605.

The only update I'd suggest is to use backticks (e.g., \`software name\`) for software and package names in `paper.md`. For example, `GeoAI`, `QGIS`, `PyTorch`, etc.

Please update the paper and let me know once it's done. Thank you.
