### 2026-01-30 #9605 @giswqs
@mengqi-z I confirm that all the list items have been reviewed and completed. Here is the DOI:
https://doi.org/10.5281/zenodo.18424047
