### 2026-01-30 #9605 @mengqi-z
@giswqs - Thank you! I noticed two things that need to be updated.

1. The version number on Zenodo (currently v1) should match the version number of your latest release (v0.27.0).
2. The license on Zenodo (currently Creative Commons Attribution 4.0 International) should match the license listed in your GitHub repository (MIT license)
