### 2026-01-30 #9605 @mengqi-z
@giswqs - You can modify the metadata on Zenodo directly to update those information.
