### 2026-01-30 #9605 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
