### 2026-02-03 #9605 @kthyng
@giswqs Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example "jupyter".
