### 2026-02-03 #9605 @giswqs
@kthyng Thank you for the feedback. The issue has been fixed in https://github.com/opengeos/geoai/pull/512.
