### 2026-02-03 #9605 @kthyng
Congratulations on your new publication @giswqs! Many thanks to editor @mengqi-z and to reviewers @RaczeQ, @jbousquin, and @GMarupilla for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
